no more crashes on conds and while loops, messy wiring though
