+++ skema/skema-rs/skema/src/database.rs
@@ -55,15 +55,16 @@ pub struct MetadataNode {
 
 #[derive(Debug, Clone)]
 pub struct Node {
-    pub n_type: String,
+    pub n_type: String, // name of type of node, function or expression
     pub value: Option<String>,
     pub name: Option<String>,
     pub node_id: String,
     pub out_idx: Option<Vec<u32>>, // opo or pof index, directly matching the wire src/tgt notation
     pub in_indx: Option<Vec<u32>>, // opi or pif index
-    pub contents: u32, // This indexes which index this node has inside the attribute list, but for functions this is overloaded
-    pub nbox: u8,
-    pub att_bf_idx: u32,
+    pub contents: usize, // This indexes which index this node has inside the fn_array list, but for functions this is overloaded
+    pub nbox: usize, // this indexs the top level box call for the node, if any, default for none is 0
+    pub att_bf_idx: usize, // this indexes the attribute index of one higher scope if any, default for none is 0
+    pub box_counter: usize, // this indexes the box call for the node one scope up, matches nbox if higher scope is top level
 }
 
 #[derive(Debug, Clone)]
@@ -71,7 +72,18 @@ pub struct Edge {
     pub src: String,
     pub tgt: String,
     pub e_type: String,
-    pub prop: Option<u32>, // option because of opo's and opi's
+    pub prop: Option<usize>, // option because of opo's and opi's
+}
+
+#[derive(Debug, Clone)]
+pub struct ConstructorArgs {
+    pub att_box: FunctionNet, // current attribute object is in
+    pub cur_box: GrometBox,   // current box object is in
+    pub parent_node: Node,    // parent node of the current constructor
+    pub att_idx: usize,       // This will index the attribute the function is in, 0 if not
+    pub bf_counter: usize, // This indexes which top level box the function is under, inherited from parent if not explicit, 0 if not
+    pub att_bf_idx: usize, // This indexes if the function is a subscope of a larger function, 0 if not
+    pub box_counter: usize, // this is the index of the box if called inside another function, 0 if not
 }
 
 pub fn execute_query(query: &str, host: &str) -> Result<(), MgError> {
@@ -390,9 +402,9 @@ fn create_function_net_lib(gromet: &ModuleCollection, mut start: u32) -> Vec<Str
     for node in nodes.clone() {
         if (1 == node.contents) && (node.n_type == "Function") {
             function_call_repeat = true;
-            if node.nbox < original_bf {
-                original_bf = node.nbox; // This grabs the first instance of bf that the function was called
-                                         // and thus is the nbox value of the nodes of the original contents
+            if node.nbox < original_bf as usize {
+                original_bf = node.nbox as u8; // This grabs the first instance of bf that the function was called
+                                               // and thus is the nbox value of the nodes of the original contents
             }
         }
     }
@@ -414,8 +426,9 @@ fn create_function_net_lib(gromet: &ModuleCollection, mut start: u32) -> Vec<Str
             out_idx: None,
             in_indx: None,
             contents: 1,
-            nbox: bf_counter,
+            nbox: bf_counter as usize,
             att_bf_idx: 0,
+            box_counter: 0,
         };
         let e1 = Edge {
             src: String::from("mod"),
@@ -449,28 +462,37 @@ fn create_function_net_lib(gromet: &ModuleCollection, mut start: u32) -> Vec<Str
         start += 1;
         let idx = 0;
         let eboxf = gromet.modules[0].attributes[idx as usize].clone();
+
+        let mut c_args = ConstructorArgs {
+            att_box: eboxf.clone(),
+            cur_box: eboxf.bf.unwrap()[0].clone(),
+            parent_node: n1.clone(),
+            att_idx: 1,
+            bf_counter: 1,
+            att_bf_idx: 0,
+            box_counter: 0,
+        };
+
         // create the ports
-        (nodes, edges, meta_nodes, start) = create_opo(
-            nodes.clone(),
-            edges.clone(),
-            meta_nodes.clone(),
-            &gromet.clone(),
-            start,
-            n1.clone(),
-            bf_counter,
+        create_opo(
+            &gromet,    // gromet for metadata
+            &mut nodes, // nodes
+            &mut edges,
+            &mut meta_nodes,
+            &mut start,
+            c_args.clone(),
         );
-        (nodes, edges, meta_nodes, start) = create_opi(
-            nodes.clone(),
-            edges.clone(),
-            meta_nodes.clone(),
-            &gromet.clone(),
-            start,
-            n1.clone(),
-            bf_counter,
+        create_opi(
+            &gromet,    // gromet for metadata
+            &mut nodes, // nodes
+            &mut edges,
+            &mut meta_nodes,
+            &mut start,
+            c_args.clone(),
         );
         // now to add the contains wires for the additional function call onto the original contents nodes:
         for node in nodes.clone() {
-            if (node.nbox == original_bf)
+            if (node.nbox == original_bf as usize)
                 && (node.contents == (idx + 1))
                 && ((node.n_type == "Literal")
                     || (node.n_type == "Primitive")
@@ -496,7 +518,7 @@ fn create_function_net_lib(gromet: &ModuleCollection, mut start: u32) -> Vec<Str
             // find the src node
             for node in nodes.iter() {
                 // make sure in correct box
-                if original_bf == node.nbox {
+                if original_bf == node.nbox as u8 {
                     // make sure only looking in current attribute nodes for srcs and tgts
                     if (idx + 1) == node.contents {
                         // only include nodes with pifs
@@ -518,7 +540,7 @@ fn create_function_net_lib(gromet: &ModuleCollection, mut start: u32) -> Vec<Str
             // find the tgt node
             for node in nodes.iter() {
                 // make sure in correct box
-                if bf_counter == node.nbox {
+                if bf_counter == node.nbox as u8 {
                     // make sure only looking in current attribute nodes for srcs and tgts
                     if (idx + 1) == node.contents {
                         // only opi's
@@ -549,7 +571,7 @@ fn create_function_net_lib(gromet: &ModuleCollection, mut start: u32) -> Vec<Str
             // find the src node
             for node in nodes.iter() {
                 // make sure in correct box
-                if bf_counter == node.nbox {
+                if bf_counter == node.nbox as u8 {
                     // make sure only looking in current attribute nodes for srcs and tgts
                     if (idx + 1) == node.contents {
                         // only opo's
@@ -568,7 +590,7 @@ fn create_function_net_lib(gromet: &ModuleCollection, mut start: u32) -> Vec<Str
             // finding the tgt node
             for node in nodes.iter() {
                 // make sure in correct box
-                if original_bf == node.nbox {
+                if original_bf == node.nbox as u8 {
                     // make sure only looking in current attribute nodes for srcs and tgts
                     if (idx + 1) == node.contents {
                         // only include nodes with pofs
@@ -601,178 +623,46 @@ fn create_function_net_lib(gromet: &ModuleCollection, mut start: u32) -> Vec<Str
         /////////////////////////
         // This is for no function repeat in a library representation
         /////////////////////////
-        let eboxf = gromet.modules[0].attributes[0_usize].clone();
-        let n1 = Node {
-            n_type: String::from("Function"),
+        let att_box = gromet.modules[0].attributes[0_usize].clone();
+
+        let mut temp_node = Node {
+            n_type: String::from("Temp"),
             value: None,
-            name: Some(
-                eboxf.b.as_ref().unwrap()[0]
-                    .name
-                    .clone()
-                    .map_or_else(|| format!("Function{}", start), |x| x),
-            ),
-            node_id: format!("n{}", start),
+            name: None,
+            node_id: "mod".to_string(),
             out_idx: None,
             in_indx: None,
-            contents: 1,
-            nbox: bf_counter,
+            contents: 0,
+            nbox: 0,
             att_bf_idx: 0,
+            box_counter: 0,
         };
-        let e1 = Edge {
-            src: String::from("mod"),
-            tgt: format!("n{}", start),
-            e_type: String::from("Contains"),
-            prop: Some(1),
+
+        // this represents the default state for this struct
+        let mut c_args = ConstructorArgs {
+            att_box: att_box.clone(),
+            cur_box: att_box.bf.unwrap()[0].clone(),
+            parent_node: temp_node,
+            att_idx: 1,
+            bf_counter: 0,
+            att_bf_idx: 0,
+            box_counter: 0,
         };
-        nodes.push(n1.clone());
-        edges.push(e1);
 
-        if eboxf.b.as_ref().unwrap()[0].metadata.as_ref().is_some() {
-            metadata_idx = eboxf.b.unwrap()[0].metadata.unwrap();
-            let mut repeat_meta = false;
-            for node in meta_nodes.iter() {
-                if node.metadata_idx == metadata_idx {
-                    repeat_meta = true;
-                }
-            }
-            if !repeat_meta {
-                meta_nodes.append(&mut create_metadata_node(&gromet.clone(), metadata_idx));
-                // adding the metadata edge
-                let me1 = Edge {
-                    src: n1.node_id.clone(),
-                    tgt: format!("m{}", metadata_idx),
-                    e_type: String::from("Metadata"),
-                    prop: None,
-                };
-                edges.push(me1);
-            }
-        }
-        // now travel to contents index of the attribute list (note it is 1 index,
-        // so contents=1 => attribute[0])
-        // create nodes and edges for this entry, include opo's and opi's
-        start += 1;
-        let idx = 1;
-        let eboxf = gromet.modules[0].attributes[(idx - 1) as usize].clone();
-        // create the ports
-        (nodes, edges, meta_nodes, start) = create_opo(
-            nodes.clone(),
-            edges.clone(),
-            meta_nodes.clone(),
-            &gromet.clone(),
-            start,
-            n1.clone(),
-            bf_counter,
+        create_function(
+            &gromet,    // gromet for metadata
+            &mut nodes, // nodes
+            &mut edges,
+            &mut meta_nodes,
+            &mut start,
+            c_args.clone(),
         );
-        (nodes, edges, meta_nodes, start) = create_opi(
-            nodes.clone(),
-            edges.clone(),
-            meta_nodes.clone(),
-            &gromet.clone(),
-            start,
-            n1.clone(),
-            bf_counter,
-        );
-        // now to construct the nodes inside the function, currently supported Literals and Primitives
-        // first include an Expression for increased depth
-        let mut box_counter: u8 = 1;
-        if eboxf.bf.is_some() {
-            for sboxf in eboxf.bf.as_ref().unwrap().iter() {
-                match sboxf.function_type {
-                    FunctionType::Predicate => {
-                        (nodes, edges, start, meta_nodes) = create_att_predicate(
-                            &gromet.clone(),
-                            eboxf.clone(),
-                            sboxf.clone(),
-                            nodes.clone(),
-                            edges.clone(),
-                            n1.clone(),
-                            idx,
-                            box_counter,
-                            bf_counter,
-                            start,
-                            meta_nodes.clone(),
-                        );
-                    }
-                    FunctionType::Expression => {
-                        (nodes, edges, start, meta_nodes) = create_att_expression(
-                            &gromet.clone(),
-                            eboxf.clone(),
-                            sboxf.clone(),
-                            nodes.clone(),
-                            edges.clone(),
-                            n1.clone(),
-                            idx,
-                            box_counter,
-                            bf_counter,
-                            start,
-                            meta_nodes.clone(),
-                        );
-                    }
-                    FunctionType::Literal => {
-                        (nodes, edges, meta_nodes) = create_att_literal(
-                            &gromet.clone(),
-                            eboxf.clone(),
-                            sboxf.clone(),
-                            nodes.clone(),
-                            edges.clone(),
-                            n1.clone(),
-                            idx,
-                            box_counter,
-                            bf_counter,
-                            start,
-                            meta_nodes.clone(),
-                        );
-                    }
-                    FunctionType::Primitive => {
-                        (nodes, edges, meta_nodes) = create_att_primitive(
-                            &gromet.clone(),
-                            eboxf.clone(),
-                            sboxf.clone(),
-                            nodes.clone(),
-                            edges.clone(),
-                            n1.clone(),
-                            idx,
-                            box_counter,
-                            bf_counter,
-                            start,
-                            meta_nodes.clone(),
-                        );
-                    }
-                    FunctionType::Abstract => {
-                        (nodes, edges, meta_nodes) = create_att_primitive(
-                            &gromet.clone(),
-                            eboxf.clone(),
-                            sboxf.clone(),
-                            nodes.clone(),
-                            edges.clone(),
-                            n1.clone(),
-                            idx,
-                            box_counter,
-                            bf_counter,
-                            start,
-                            meta_nodes.clone(),
-                        );
-                    }
-                    _ => {}
-                }
-                box_counter += 1;
-                start += 1;
-            }
-        }
-
-        // Now we perform the internal wiring of this branch
-        edges = internal_wiring(eboxf.clone(), nodes.clone(), edges, idx, bf_counter);
-        // perform cross attributal wiring of function
-        (edges, nodes) = cross_att_wiring(eboxf, nodes.clone(), edges, idx, bf_counter);
     }
-    start += 1;
-    bf_counter += 1;
-
     // add wires for inbetween attribute level boxes, so opo's, opi's and module level literals
     // between attributes
     // get wired through module level wff field, will require reading through node list to
     // match contents field to box field on wff entries
-    edges = external_wiring(gromet, nodes.clone(), edges);
+    external_wiring(gromet, &mut nodes.clone(), &mut edges);
 
     // make conditionals if they exist
     if gromet.modules[0].r#fn.bc.as_ref().is_some() {
@@ -787,20 +677,27 @@ fn create_function_net_lib(gromet: &ModuleCollection, mut start: u32) -> Vec<Str
             contents: 0,
             nbox: 0,
             att_bf_idx: 0,
+            box_counter: 0,
+        };
+        let mut c_args = ConstructorArgs {
+            att_box: gromet.modules[0].r#fn.clone(),
+            cur_box: gromet.modules[0].attributes[0].clone().bf.unwrap()[0].clone(), // this is a temp fill
+            parent_node: temp_mod_node,
+            att_idx: 1,
+            bf_counter: 0,
+            att_bf_idx: 0,
+            box_counter: 0,
         };
         for _cond in gromet.modules[0].r#fn.bc.as_ref().unwrap().iter() {
             // now lets check for and setup any conditionals at this level
-            (nodes, edges, start, meta_nodes) = create_conditional(
-                &gromet.clone(),
-                gromet.modules[0].r#fn.clone(), // This is gromet but is more generalizable based on scope
-                nodes.clone(),
-                edges.clone(),
-                temp_mod_node.clone(),
-                0,            // because top level
-                cond_counter, // This indexes the conditional in the list of conditionals (bc)
-                0,            // because top level
-                start,
-                meta_nodes.clone(),
+            create_conditional(
+                &gromet,    // gromet for metadata
+                &mut nodes, // nodes
+                &mut edges,
+                &mut meta_nodes,
+                &mut start,
+                c_args.clone(),
+                cond_counter,
             );
             cond_counter += 1;
         }
@@ -818,20 +715,27 @@ fn create_function_net_lib(gromet: &ModuleCollection, mut start: u32) -> Vec<Str
             contents: 0,
             nbox: 0,
             att_bf_idx: 0,
+            box_counter: 0,
+        };
+        let mut c_args = ConstructorArgs {
+            att_box: gromet.modules[0].r#fn.clone(),
+            cur_box: gromet.modules[0].attributes[0].clone().bf.unwrap()[0].clone(),
+            parent_node: temp_mod_node,
+            att_idx: 1,
+            bf_counter: 0,
+            att_bf_idx: 0,
+            box_counter: 0,
         };
         for _while_l in gromet.modules[0].r#fn.bl.as_ref().unwrap().iter() {
             // now lets check for and setup any conditionals at this level
-            (nodes, edges, start, meta_nodes) = create_while_loop(
-                &gromet.clone(),
-                gromet.modules[0].r#fn.clone(), // This is gromet but is more generalizable based on scope
-                nodes.clone(),
-                edges.clone(),
-                temp_mod_node.clone(),
-                0,             // because top level
-                while_counter, // This indexes the conditional in the list of conditionals (bc)
-                0,             // because top level
-                start,
-                meta_nodes.clone(),
+            create_while_loop(
+                &gromet,    // gromet for metadata
+                &mut nodes, // nodes
+                &mut edges,
+                &mut meta_nodes,
+                &mut start,
+                c_args.clone(),
+                while_counter,
             );
             while_counter += 1;
         }
@@ -886,283 +790,82 @@ fn create_function_net(gromet: &ModuleCollection, mut start: u32) -> Vec<String>
     let mut metadata_idx = 0;
     let mut edges: Vec<Edge> = vec![];
 
-    let mut bf_counter: u8 = 1;
+    // temp node for calling the module as parent for top level
+    let mut temp_node = Node {
+        n_type: String::from("Temp"),
+        value: None,
+        name: None,
+        node_id: "mod".to_string(),
+        out_idx: None,
+        in_indx: None,
+        contents: 0,
+        nbox: 0,
+        att_bf_idx: 0,
+        box_counter: 0,
+    };
+
+    // this represents the default state for this struct
+    let mut c_args = ConstructorArgs {
+        att_box: gromet.modules[0].r#fn.clone(),
+        cur_box: gromet.modules[0].r#fn.bf.as_ref().unwrap()[0].clone(),
+        parent_node: temp_node.clone(),
+        att_idx: 0,
+        bf_counter: 0,
+        att_bf_idx: 0,
+        box_counter: 0,
+    };
+
+    let mut bf_counter: usize = 1;
     for boxf in gromet.modules[0].r#fn.bf.as_ref().unwrap().iter() {
-        // construct the sub module level boxes along with their metadata and connection to module
+        c_args.bf_counter = bf_counter;
+        c_args.cur_box = boxf.clone();
+        c_args.att_idx = 0; // reset incase we hit a function and then literal for example
+                            // construct the sub module level boxes along with their metadata and connection to module
+        c_args.parent_node = temp_node.clone(); // for if overwritten in function constructor
         match boxf.function_type {
             FunctionType::Literal => {
-                // first we find the pof value for the literal box
-                let mut pof: Vec<u32> = vec![];
-                if gromet.modules[0].r#fn.pof.clone().is_some() {
-                    let mut po_idx: u32 = 1;
-                    for port in gromet.modules[0].r#fn.pof.clone().unwrap().iter() {
-                        if port.r#box == bf_counter {
-                            pof.push(po_idx);
-                        }
-                        po_idx += 1;
-                    }
-                }
-                let n1 = Node {
-                    n_type: String::from("Literal"),
-                    value: Some(format!("{:?}", boxf.value.clone().as_ref().unwrap())),
-                    name: Some("Literal".to_string()),
-                    node_id: format!("n{}", start),
-                    out_idx: Some(pof),
-                    in_indx: None,
-                    contents: 0,
-                    nbox: bf_counter,
-                    att_bf_idx: 0,
-                };
-                let e1 = Edge {
-                    src: String::from("mod"),
-                    tgt: format!("n{}", start),
-                    e_type: String::from("Contains"),
-                    prop: boxf.contents,
-                };
-                nodes.push(n1.clone());
-                edges.push(e1);
-                if boxf.metadata.as_ref().is_some() {
-                    metadata_idx = boxf.metadata.unwrap();
-                    let mut repeat_meta = false;
-                    for node in meta_nodes.iter() {
-                        if node.metadata_idx == metadata_idx {
-                            repeat_meta = true;
-                        }
-                    }
-                    if !repeat_meta {
-                        meta_nodes.append(&mut create_metadata_node(&gromet.clone(), metadata_idx));
-                        let me1 = Edge {
-                            src: n1.node_id.clone(),
-                            tgt: format!("m{}", metadata_idx),
-                            e_type: String::from("Metadata"),
-                            prop: None,
-                        };
-                        edges.push(me1);
-                    }
-                }
+                create_att_literal(
+                    &gromet,    // gromet for metadata
+                    &mut nodes, // nodes
+                    &mut edges,
+                    &mut meta_nodes,
+                    &mut start,
+                    c_args.clone(),
+                );
             }
             FunctionType::Predicate => {
-                let n1 = Node {
-                    n_type: String::from("Predicate"),
-                    value: None,
-                    name: Some(format!("Predicate{}", start)),
-                    node_id: format!("n{}", start),
-                    out_idx: None,
-                    in_indx: None,
-                    contents: boxf.contents.unwrap(),
-                    nbox: bf_counter,
-                    att_bf_idx: 0,
-                };
-                let e1 = Edge {
-                    src: String::from("mod"),
-                    tgt: format!("n{}", start),
-                    e_type: String::from("Contains"),
-                    prop: boxf.contents,
-                };
-                nodes.push(n1.clone());
-                edges.push(e1);
-
-                if boxf.metadata.as_ref().is_some() {
-                    metadata_idx = boxf.metadata.unwrap();
-                    let mut repeat_meta = false;
-                    for node in meta_nodes.iter() {
-                        if node.metadata_idx == metadata_idx {
-                            repeat_meta = true;
-                        }
-                    }
-                    if !repeat_meta {
-                        meta_nodes.append(&mut create_metadata_node(&gromet.clone(), metadata_idx));
-                        let me1 = Edge {
-                            src: n1.node_id.clone(),
-                            tgt: format!("m{}", metadata_idx),
-                            e_type: String::from("Metadata"),
-                            prop: None,
-                        };
-                        edges.push(me1);
-                    }
-                }
-
-                // now travel to contents index of the attribute list (note it is 1 index,
-                // so contents=1 => attribute[0])
-                // create nodes and edges for this entry, include opo's and opi's
-                start += 1;
-                let idx = boxf.contents.unwrap() - 1;
-                let eboxf = gromet.modules[0].attributes[idx as usize].clone();
-                // create the ports
-                (nodes, edges, meta_nodes, start) = create_opo(
-                    nodes.clone(),
-                    edges.clone(),
-                    meta_nodes.clone(),
-                    &gromet.clone(),
-                    start,
-                    n1.clone(),
-                    bf_counter,
+                c_args.att_idx = boxf.contents.unwrap() as usize;
+                create_att_predicate(
+                    &gromet,    // gromet for metadata
+                    &mut nodes, // nodes
+                    &mut edges,
+                    &mut meta_nodes,
+                    &mut start,
+                    c_args.clone(),
                 );
-                (nodes, edges, meta_nodes, start) = create_opi(
-                    nodes.clone(),
-                    edges.clone(),
-                    meta_nodes.clone(),
-                    &gromet.clone(),
-                    start,
-                    n1.clone(),
-                    bf_counter,
-                );
-                // now to construct the nodes inside the expression, Literal and Primitives
-                let mut box_counter: u8 = 1;
-                for sboxf in eboxf.bf.clone().as_ref().unwrap().iter() {
-                    match sboxf.function_type {
-                        FunctionType::Literal => {
-                            (nodes, edges, meta_nodes) = create_att_literal(
-                                &gromet.clone(),
-                                eboxf.clone(),
-                                sboxf.clone(),
-                                nodes.clone(),
-                                edges.clone(),
-                                n1.clone(),
-                                idx,
-                                box_counter,
-                                bf_counter,
-                                start,
-                                meta_nodes.clone(),
-                            );
-                        }
-                        FunctionType::Primitive => {
-                            (nodes, edges, meta_nodes) = create_att_primitive(
-                                &gromet.clone(),
-                                eboxf.clone(),
-                                sboxf.clone(),
-                                nodes.clone(),
-                                edges.clone(),
-                                n1.clone(),
-                                idx,
-                                box_counter,
-                                bf_counter,
-                                start,
-                                meta_nodes.clone(),
-                            );
-                        }
-                        _ => {}
-                    }
-                    box_counter += 1;
-                    start += 1;
-                }
-                // Now we perform the internal wiring of this branch
-                edges = internal_wiring(eboxf.clone(), nodes.clone(), edges, idx, bf_counter);
             }
             FunctionType::Expression => {
-                let n1 = Node {
-                    n_type: String::from("Expression"),
-                    value: None,
-                    name: Some(format!("Expression{}", start)),
-                    node_id: format!("n{}", start),
-                    out_idx: None,
-                    in_indx: None,
-                    contents: boxf.contents.unwrap(),
-                    nbox: bf_counter,
-                    att_bf_idx: 0,
-                };
-                let e1 = Edge {
-                    src: String::from("mod"),
-                    tgt: format!("n{}", start),
-                    e_type: String::from("Contains"),
-                    prop: boxf.contents,
-                };
-                nodes.push(n1.clone());
-                edges.push(e1);
-                if boxf.metadata.as_ref().is_some() {
-                    metadata_idx = boxf.metadata.unwrap();
-                    let mut repeat_meta = false;
-                    for node in meta_nodes.iter() {
-                        if node.metadata_idx == metadata_idx {
-                            repeat_meta = true;
-                        }
-                    }
-                    if !repeat_meta {
-                        meta_nodes.append(&mut create_metadata_node(&gromet.clone(), metadata_idx));
-                        let me1 = Edge {
-                            src: n1.node_id.clone(),
-                            tgt: format!("m{}", metadata_idx),
-                            e_type: String::from("Metadata"),
-                            prop: None,
-                        };
-                        edges.push(me1);
-                    }
-                }
-
-                // now travel to contents index of the attribute list (note it is 1 index,
-                // so contents=1 => attribute[0])
-                // create nodes and edges for this entry, include opo's and opi's
-                start += 1;
-                let idx = boxf.contents.unwrap() - 1;
-
-                let eboxf = gromet.modules[0].attributes[idx as usize].clone();
-                (nodes, edges, meta_nodes, start) = create_opo(
-                    nodes.clone(),
-                    edges.clone(),
-                    meta_nodes.clone(),
-                    &gromet.clone(),
-                    start,
-                    n1.clone(),
-                    bf_counter,
+                c_args.att_idx = boxf.contents.unwrap() as usize;
+                create_att_expression(
+                    &gromet,    // gromet for metadata
+                    &mut nodes, // nodes
+                    &mut edges,
+                    &mut meta_nodes,
+                    &mut start,
+                    c_args.clone(),
                 );
-                (nodes, edges, meta_nodes, start) = create_opi(
-                    nodes.clone(),
-                    edges.clone(),
-                    meta_nodes.clone(),
-                    &gromet.clone(),
-                    start,
-                    n1.clone(),
-                    bf_counter,
-                );
-                // now to construct the nodes inside the expression, Literal and Primitives
-                let mut box_counter: u8 = 1;
-                for sboxf in eboxf.bf.clone().as_ref().unwrap().iter() {
-                    match sboxf.function_type {
-                        FunctionType::Literal => {
-                            (nodes, edges, meta_nodes) = create_att_literal(
-                                &gromet.clone(),
-                                eboxf.clone(),
-                                sboxf.clone(),
-                                nodes.clone(),
-                                edges.clone(),
-                                n1.clone(),
-                                idx,
-                                box_counter,
-                                bf_counter,
-                                start,
-                                meta_nodes.clone(),
-                            );
-                        }
-                        FunctionType::Primitive => {
-                            (nodes, edges, meta_nodes) = create_att_primitive(
-                                &gromet.clone(),
-                                eboxf.clone(),
-                                sboxf.clone(),
-                                nodes.clone(),
-                                edges.clone(),
-                                n1.clone(),
-                                idx,
-                                box_counter,
-                                bf_counter,
-                                start,
-                                meta_nodes.clone(),
-                            );
-                        }
-                        _ => {}
-                    }
-                    box_counter += 1;
-                    start += 1;
-                }
-                // Now we perform the internal wiring of this branch
-                edges = internal_wiring(eboxf.clone(), nodes.clone(), edges, idx, bf_counter);
             }
             FunctionType::Function => {
                 // in order to have less repetition for multiple function calls and to setup support for recursive functions
                 // We check if the function node and thus contents were already made, and not duplicate the contents if already made
+                c_args.att_idx = boxf.contents.unwrap() as usize;
+
                 let mut function_call_repeat = false;
                 let mut original_bf = bf_counter;
                 for node in nodes.clone() {
-                    if (boxf.contents.unwrap() == node.contents) && (node.n_type == "Function") {
+                    if (boxf.contents.unwrap() == node.contents as u32)
+                        && (node.n_type == "Function")
+                    {
                         function_call_repeat = true;
                         if node.nbox < original_bf {
                             original_bf = node.nbox; // This grabs the first instance of bf that the function was called
@@ -1190,15 +893,16 @@ fn create_function_net(gromet: &ModuleCollection, mut start: u32) -> Vec<String>
                         node_id: format!("n{}", start),
                         out_idx: None,
                         in_indx: None,
-                        contents: idx + 1,
+                        contents: idx as usize + 1,
                         nbox: bf_counter,
                         att_bf_idx: 0,
+                        box_counter: 0,
                     };
                     let e1 = Edge {
                         src: String::from("mod"),
                         tgt: format!("n{}", start),
                         e_type: String::from("Contains"),
-                        prop: boxf.contents,
+                        prop: Some(boxf.contents.unwrap() as usize),
                     };
                     nodes.push(n1.clone());
                     edges.push(e1);
@@ -1246,29 +950,30 @@ fn create_function_net(gromet: &ModuleCollection, mut start: u32) -> Vec<String>
                     }
                     // we still construct unique ports for this function, however the contents will not be repeated
                     start += 1;
+
+                    c_args.parent_node = n1.clone();
+
                     // construct opo nodes, if not none
-                    (nodes, edges, meta_nodes, start) = create_opo(
-                        nodes.clone(),
-                        edges.clone(),
-                        meta_nodes.clone(),
-                        &gromet.clone(),
-                        start,
-                        n1.clone(),
-                        bf_counter,
+                    create_opo(
+                        &gromet,    // gromet for metadata
+                        &mut nodes, // nodes
+                        &mut edges,
+                        &mut meta_nodes,
+                        &mut start,
+                        c_args.clone(),
                     );
-                    (nodes, edges, meta_nodes, start) = create_opi(
-                        nodes.clone(),
-                        edges.clone(),
-                        meta_nodes.clone(),
-                        &gromet.clone(),
-                        start,
-                        n1.clone(),
-                        bf_counter,
+                    create_opi(
+                        &gromet,    // gromet for metadata
+                        &mut nodes, // nodes
+                        &mut edges,
+                        &mut meta_nodes,
+                        &mut start,
+                        c_args.clone(),
                     );
                     // now to add the contains wires for the additional function call onto the original contents nodes:
                     for node in nodes.clone() {
                         if (node.nbox == original_bf)
-                            && (node.contents == (idx + 1))
+                            && (node.contents == (idx as usize + 1))
                             && ((node.n_type == "Literal")
                                 || (node.n_type == "Primitive")
                                 || (node.n_type == "Predicate")
@@ -1295,7 +1000,7 @@ fn create_function_net(gromet: &ModuleCollection, mut start: u32) -> Vec<String>
                             // make sure in correct box
                             if original_bf == node.nbox {
                                 // make sure only looking in current attribute nodes for srcs and tgts
-                                if (idx + 1) == node.contents {
+                                if (idx + 1) == node.contents as u32 {
                                     // only include nodes with pifs
                                     if node.in_indx.is_some() {
                                         // exclude opi's
@@ -1317,7 +1022,7 @@ fn create_function_net(gromet: &ModuleCollection, mut start: u32) -> Vec<String>
                             // make sure in correct box
                             if bf_counter == node.nbox {
                                 // make sure only looking in current attribute nodes for srcs and tgts
-                                if (idx + 1) == node.contents {
+                                if (idx + 1) == node.contents as u32 {
                                     // only opi's
                                     if node.n_type == "Opi" {
                                         // iterate through port to check for tgt
@@ -1348,7 +1053,7 @@ fn create_function_net(gromet: &ModuleCollection, mut start: u32) -> Vec<String>
                             // make sure in correct box
                             if bf_counter == node.nbox {
                                 // make sure only looking in current attribute nodes for srcs and tgts
-                                if (idx + 1) == node.contents {
+                                if (idx + 1) == node.contents as u32 {
                                     // only opo's
                                     if node.n_type == "Opo" {
                                         // iterate through port to check for tgt
@@ -1367,7 +1072,7 @@ fn create_function_net(gromet: &ModuleCollection, mut start: u32) -> Vec<String>
                             // make sure in correct box
                             if original_bf == node.nbox {
                                 // make sure only looking in current attribute nodes for srcs and tgts
-                                if (idx + 1) == node.contents {
+                                if (idx + 1) == node.contents as u32 {
                                     // only include nodes with pofs
                                     if node.out_idx.is_some() {
                                         // exclude opo's
@@ -1395,35 +1100,15 @@ fn create_function_net(gromet: &ModuleCollection, mut start: u32) -> Vec<String>
                         }
                     }
                 } else {
-                    // mark
-                    // get attribute function net to pass into function
-                    let pfn =
-                        gromet.modules[0].attributes[(boxf.contents.unwrap() - 1) as usize].clone();
-
-                    // need a parent node for the module at this level
-                    let temp_mod_node = Node {
-                        n_type: String::from("module"),
-                        value: None,
-                        name: None,
-                        node_id: "mod".to_string(),
-                        out_idx: None,
-                        in_indx: None,
-                        contents: 0,
-                        nbox: 0,
-                        att_bf_idx: 0,
-                    };
-
-                    (nodes, edges, start, meta_nodes) = create_function(
-                        &gromet.clone(),        // gromet for metadata
-                        pfn.clone(),            // function net for attribute
-                        nodes.clone(),          // nodes
-                        edges.clone(),          // edges
-                        temp_mod_node.clone(),  // parent node
-                        boxf.contents.unwrap(), // attribute idx
-                        bf_counter,             // top level bf idx
-                        0,                      // sub box idx, 0 for top level
-                        start,                  // start
-                        meta_nodes.clone(),     // metadata nodes
+                    // new function call
+                    c_args.att_idx = boxf.contents.unwrap() as usize;
+                    create_function(
+                        &gromet,    // gromet for metadata
+                        &mut nodes, // nodes
+                        &mut edges,
+                        &mut meta_nodes,
+                        &mut start,
+                        c_args.clone(),
                     );
                 }
             }
@@ -1432,12 +1117,13 @@ fn create_function_net(gromet: &ModuleCollection, mut start: u32) -> Vec<String>
         start += 1;
         bf_counter += 1;
     }
+    start += 1;
 
     // add wires for inbetween attribute level boxes, so opo's, opi's and module level literals
     // between attributes
     // get wired through module level wff field, will require reading through node list to
     // match contents field to box field on wff entries
-    edges = external_wiring(gromet, nodes.clone(), edges);
+    external_wiring(gromet, &mut nodes.clone(), &mut edges);
 
     // make conditionals if they exist
     if gromet.modules[0].r#fn.bc.as_ref().is_some() {
@@ -1452,20 +1138,18 @@ fn create_function_net(gromet: &ModuleCollection, mut start: u32) -> Vec<String>
             contents: 0,
             nbox: 0,
             att_bf_idx: 0,
+            box_counter: 0,
         };
         for _cond in gromet.modules[0].r#fn.bc.as_ref().unwrap().iter() {
             // now lets check for and setup any conditionals at this level
-            (nodes, edges, start, meta_nodes) = create_conditional(
-                &gromet.clone(),
-                gromet.modules[0].r#fn.clone(), // This is gromet but is more generalizable based on scope
-                nodes.clone(),
-                edges.clone(),
-                temp_mod_node.clone(),
-                0,            // because top level
-                cond_counter, // This indexes the conditional in the list of conditionals (bc)
-                0,            // because top level
-                start,
-                meta_nodes.clone(),
+            create_conditional(
+                &gromet,    // gromet for metadata
+                &mut nodes, // nodes
+                &mut edges,
+                &mut meta_nodes,
+                &mut start,
+                c_args.clone(),
+                cond_counter,
             );
             cond_counter += 1;
         }
@@ -1483,20 +1167,18 @@ fn create_function_net(gromet: &ModuleCollection, mut start: u32) -> Vec<String>
             contents: 0,
             nbox: 0,
             att_bf_idx: 0,
+            box_counter: 0,
         };
         for _while_l in gromet.modules[0].r#fn.bl.as_ref().unwrap().iter() {
             // now lets check for and setup any conditionals at this level
-            (nodes, edges, start, meta_nodes) = create_while_loop(
-                &gromet.clone(),
-                gromet.modules[0].r#fn.clone(), // This is gromet but is more generalizable based on scope
-                nodes.clone(),
-                edges.clone(),
-                temp_mod_node.clone(),
-                0,             // because top level
-                while_counter, // This indexes the conditional in the list of conditionals (bc)
-                0,             // because top level
-                start,
-                meta_nodes.clone(),
+            create_while_loop(
+                &gromet,    // gromet for metadata
+                &mut nodes, // nodes
+                &mut edges,
+                &mut meta_nodes,
+                &mut start,
+                c_args.clone(),
+                while_counter,
             );
             while_counter += 1;
         }
@@ -1542,27 +1224,25 @@ fn create_function_net(gromet: &ModuleCollection, mut start: u32) -> Vec<String>
     queries
 }
 // this method creates an import type function
+// currently assumes top level call
 #[allow(unused_assignments)]
 pub fn create_import(
     gromet: &ModuleCollection, // needed still for metadata unfortunately
-    mut nodes: Vec<Node>,
-    mut edges: Vec<Edge>,
-    parent_node: Node, // used for contains construction
-    att_idx: u32,      // This will index the attribute the function is in
-    bf_counter: u8, // This indexes which box the function is under, inherited from parent if not explicit
-    att_bf_idx: u32, // This indexes if the function is a subscope of a larger function, 0 if not
-    start: u32,     // for node and edge indexing
-    mut meta_nodes: Vec<MetadataNode>,
-) -> (Vec<Node>, Vec<Edge>, u32, Vec<MetadataNode>) {
+    nodes: &mut Vec<Node>,
+    edges: &mut Vec<Edge>,
+    meta_nodes: &mut Vec<MetadataNode>,
+    start: &mut u32, // for node and edge indexing
+    c_args: ConstructorArgs,
+) -> () {
     let eboxf = gromet.modules[0].clone();
-    let sboxf = gromet.modules[0].attributes[(att_idx - 1) as usize].clone();
-    let mboxf = eboxf.r#fn.bf.unwrap()[(bf_counter - 1) as usize].clone();
+    let sboxf = gromet.modules[0].attributes[(c_args.att_idx - 1) as usize].clone();
+    let mboxf = eboxf.r#fn.bf.unwrap()[(c_args.bf_counter - 1) as usize].clone();
 
     let mut pof: Vec<u32> = vec![];
     if eboxf.r#fn.pof.is_some() {
         let mut po_idx: u32 = 1;
         for port in eboxf.r#fn.pof.clone().unwrap().iter() {
-            if port.r#box == bf_counter {
+            if port.r#box == c_args.bf_counter as u8 {
                 pof.push(po_idx);
             }
             po_idx += 1;
@@ -1573,7 +1253,7 @@ pub fn create_import(
     if eboxf.r#fn.pif.is_some() {
         let mut pi_idx: u32 = 1;
         for port in eboxf.r#fn.pif.unwrap().iter() {
-            if port.r#box == bf_counter {
+            if port.r#box == c_args.bf_counter as u8 {
                 pif.push(pi_idx);
             }
             pi_idx += 1;
@@ -1588,14 +1268,15 @@ pub fn create_import(
         node_id: format!("n{}", start),
         out_idx: Some(pof),
         in_indx: Some(pif),
-        contents: att_idx,
-        nbox: bf_counter,
-        att_bf_idx,
+        contents: c_args.att_idx,
+        nbox: c_args.bf_counter,
+        att_bf_idx: c_args.att_bf_idx,
+        box_counter: c_args.box_counter,
     };
     nodes.push(n3.clone());
     // make edge connecting to expression
     let e4 = Edge {
-        src: parent_node.node_id,
+        src: c_args.parent_node.node_id,
         tgt: n3.node_id.clone(),
         e_type: String::from("Contains"),
         prop: None,
@@ -1620,61 +1301,44 @@ pub fn create_import(
             edges.push(me1);
         }
     }
-
-    (nodes, edges, start, meta_nodes)
+    *start += 1;
 }
 
 // this creates a function node including all the contents included in it, including additional functions
+// CHANGEs: removed function_net arg and added box_counter arg
 #[allow(unused_assignments)]
 pub fn create_function(
     gromet: &ModuleCollection, // needed still for metadata unfortunately
-    function_net: FunctionNet, // This is gromet but is more generalizable based on scope
-    mut nodes: Vec<Node>,
-    mut edges: Vec<Edge>,
-    parent_node: Node, // used for contains construction
-    att_idx: u32,      // This will index the attribute the function is in
-    bf_counter: u8, // This indexes which box the function is under, inherited from parent if not explicit
-    att_bf_idx: u32, // This indexes if the function is a subscope of a larger function, 0 if not
-    mut start: u32, // for node and edge indexing
-    mut meta_nodes: Vec<MetadataNode>,
-) -> (Vec<Node>, Vec<Edge>, u32, Vec<MetadataNode>) {
+    nodes: &mut Vec<Node>,
+    edges: &mut Vec<Edge>,
+    meta_nodes: &mut Vec<MetadataNode>,
+    start: &mut u32, // for node and edge indexing
+    c_args: ConstructorArgs,
+) -> () {
     // function is not repeated
-    let eboxf = gromet.modules[0].attributes[(att_idx - 1) as usize].clone();
+    let att_box = gromet.modules[0].attributes[(c_args.att_idx - 1) as usize].clone(); // current expression attribute
 
     // now we add a check for if this is an imported function
-    match eboxf.b.clone().unwrap()[0].function_type.clone() {
+    match att_box.b.clone().unwrap()[0].function_type.clone() {
         FunctionType::Imported => {
-            (nodes, edges, start, meta_nodes) = create_import(
-                &gromet.clone(),
-                nodes.clone(),
-                edges.clone(),
-                parent_node.clone(),
-                att_idx,
-                bf_counter,
-                att_bf_idx,
-                start,
-                meta_nodes.clone(),
-            );
-            start += 1;
+            create_import(&gromet, nodes, edges, meta_nodes, start, c_args.clone());
+            *start += 1;
             // now to implement wiring
-            edges = import_wiring(
+            import_wiring(
                 &gromet.clone(),
-                eboxf,
-                nodes.clone(),
+                nodes,
                 edges,
-                att_idx,
-                bf_counter,
-                parent_node,
+                c_args.att_idx,
+                c_args.bf_counter,
+                c_args.parent_node,
             );
-
-            (nodes, edges, start, meta_nodes)
         }
         _ => {
             let n1 = Node {
                 n_type: String::from("Function"),
                 value: None,
                 name: Some(
-                    eboxf.b.as_ref().unwrap()[0]
+                    att_box.b.as_ref().unwrap()[0]
                         .name
                         .clone()
                         .map_or_else(|| format!("Function{}", start), |x| x),
@@ -1682,22 +1346,23 @@ pub fn create_function(
                 node_id: format!("n{}", start),
                 out_idx: None,
                 in_indx: None,
-                contents: att_idx,
-                nbox: bf_counter,
-                att_bf_idx,
+                contents: c_args.att_idx,
+                nbox: c_args.bf_counter,
+                att_bf_idx: c_args.att_bf_idx,
+                box_counter: c_args.box_counter,
             };
             let e1 = Edge {
-                src: parent_node.node_id,
+                src: c_args.parent_node.node_id.clone(),
                 tgt: n1.node_id.clone(),
                 e_type: String::from("Contains"),
-                prop: Some(att_idx),
+                prop: Some(c_args.att_idx),
             };
             nodes.push(n1.clone());
             edges.push(e1);
             let mut metadata_idx = 0;
             // attribute b level metadata reference
-            if eboxf.b.as_ref().unwrap()[0].metadata.as_ref().is_some() {
-                metadata_idx = eboxf.b.as_ref().unwrap()[0].metadata.unwrap();
+            if att_box.b.as_ref().unwrap()[0].metadata.as_ref().is_some() {
+                metadata_idx = att_box.b.as_ref().unwrap()[0].metadata.unwrap();
                 let mut repeat_meta = false;
                 for node in meta_nodes.iter() {
                     if node.metadata_idx == metadata_idx {
@@ -1720,163 +1385,113 @@ pub fn create_function(
             // now travel to contents index of the attribute list (note it is 1 index,
             // so contents=1 => attribute[0])
             // create nodes and edges for this entry, include opo's and opi's
-            start += 1;
+            *start += 1;
+
+            let mut new_c_args = c_args.clone();
+            new_c_args.parent_node = n1.clone();
 
             // construct opo nodes, if not none, might need to
-            (nodes, edges, meta_nodes, start) = create_opo(
-                nodes.clone(),
-                edges.clone(),
-                meta_nodes.clone(),
-                &gromet.clone(),
+            create_opo(
+                &gromet, // gromet for metadata
+                nodes,   // nodes
+                edges,
+                meta_nodes,
                 start,
-                n1.clone(),
-                bf_counter,
+                new_c_args.clone(),
             );
-            (nodes, edges, meta_nodes, start) = create_opi(
-                nodes.clone(),
-                edges.clone(),
-                meta_nodes.clone(),
-                &gromet.clone(),
+            create_opi(
+                &gromet, // gromet for metadata
+                nodes,   // nodes
+                edges,
+                meta_nodes,
                 start,
-                n1.clone(),
-                bf_counter,
+                new_c_args.clone(),
             );
             // now to construct the nodes inside the function, currently supported Literals and Primitives
             // first include an Expression for increased depth
-            let mut box_counter: u8 = 1;
-            for sboxf in eboxf.bf.as_ref().unwrap().iter() {
-                match sboxf.function_type {
+            let mut box_counter: usize = 1;
+            for att_sub_box in att_box.bf.as_ref().unwrap().iter() {
+                new_c_args.box_counter = box_counter;
+                new_c_args.cur_box = att_sub_box.clone();
+                new_c_args.att_idx = c_args.att_idx;
+                match att_sub_box.function_type {
                     FunctionType::Function => {
-                        (nodes, edges, start, meta_nodes) = create_function(
-                            &gromet.clone(),         // gromet for metadata
-                            eboxf.clone(),           // function net for attribute
-                            nodes.clone(),           // nodes
-                            edges.clone(),           // edges
-                            n1.clone(),              // parent node
-                            sboxf.contents.unwrap(), // attribute idx
-                            bf_counter,              // top level bf idx
-                            box_counter.into(),      // sub box idx, 0 for top level
-                            start,                   // start
-                            meta_nodes.clone(),      // metadata nodes
+                        new_c_args.att_idx = att_sub_box.contents.unwrap() as usize;
+                        create_function(
+                            &gromet, // gromet for metadata
+                            nodes,   // nodes
+                            edges,
+                            meta_nodes,
+                            start,
+                            new_c_args.clone(),
                         );
                     }
 
                     FunctionType::Predicate => {
-                        (nodes, edges, start, meta_nodes) = create_att_predicate(
-                            &gromet.clone(),
-                            eboxf.clone(),
-                            sboxf.clone(),
-                            nodes.clone(),
-                            edges.clone(),
-                            n1.clone(),
-                            att_idx,
-                            box_counter,
-                            bf_counter,
+                        new_c_args.att_idx = att_sub_box.contents.unwrap() as usize;
+                        create_att_predicate(
+                            &gromet, // gromet for metadata
+                            nodes,   // nodes
+                            edges,
+                            meta_nodes,
                             start,
-                            meta_nodes.clone(),
+                            new_c_args.clone(),
                         );
                     }
                     FunctionType::Expression => {
-                        (nodes, edges, start, meta_nodes) = create_att_expression(
-                            &gromet.clone(),
-                            eboxf.clone(),
-                            sboxf.clone(),
-                            nodes.clone(),
-                            edges.clone(),
-                            n1.clone(),
-                            att_idx,
-                            box_counter,
-                            bf_counter,
+                        new_c_args.att_idx = att_sub_box.contents.unwrap() as usize;
+                        create_att_expression(
+                            &gromet, // gromet for metadata
+                            nodes,   // nodes
+                            edges,
+                            meta_nodes,
                             start,
-                            meta_nodes.clone(),
+                            new_c_args.clone(),
                         );
                     }
                     FunctionType::Literal => {
-                        (nodes, edges, meta_nodes) = create_att_literal(
-                            &gromet.clone(),
-                            eboxf.clone(),
-                            sboxf.clone(),
-                            nodes.clone(),
-                            edges.clone(),
-                            n1.clone(),
-                            att_idx,
-                            box_counter,
-                            bf_counter,
+                        create_att_literal(
+                            &gromet, // gromet for metadata
+                            nodes,   // nodes
+                            edges,
+                            meta_nodes,
                             start,
-                            meta_nodes.clone(),
+                            new_c_args.clone(),
                         );
                     }
                     FunctionType::Primitive => {
-                        (nodes, edges, meta_nodes) = create_att_primitive(
-                            &gromet.clone(),
-                            eboxf.clone(),
-                            sboxf.clone(),
-                            nodes.clone(),
-                            edges.clone(),
-                            n1.clone(),
-                            att_idx,
-                            box_counter,
-                            bf_counter,
+                        create_att_primitive(
+                            &gromet, // gromet for metadata
+                            nodes,   // nodes
+                            edges,
+                            meta_nodes,
                             start,
-                            meta_nodes.clone(),
+                            new_c_args.clone(),
                         );
                     }
                     // stub for where to put in create function for nested functions
                     _ => {}
                 }
                 box_counter += 1;
-                start += 1;
+                *start += 1;
             }
 
             // Now we perform the internal wiring of this branch
-            edges = internal_wiring(eboxf.clone(), nodes.clone(), edges, att_idx, bf_counter);
+            internal_wiring(
+                att_box.clone(),
+                nodes,
+                edges,
+                c_args.att_idx,
+                c_args.bf_counter,
+            );
             // perform cross attributal wiring of function
-            (edges, nodes) =
-                cross_att_wiring(eboxf.clone(), nodes.clone(), edges, att_idx, bf_counter);
-
-            // make conditionals if they exist
-            // might need to make wiring more robust
-            if gromet.modules[0].r#fn.bc.as_ref().is_some() {
-                let mut cond_counter = 0;
-                for _cond in gromet.modules[0].r#fn.bc.as_ref().unwrap().iter() {
-                    // now lets check for and setup any conditionals at this level
-                    (nodes, edges, start, meta_nodes) = create_conditional(
-                        &gromet.clone(),
-                        eboxf.clone(), // This is gromet but is more generalizable based on scope
-                        nodes.clone(),
-                        edges.clone(),
-                        n1.clone(),
-                        att_idx - 1,  // because top level
-                        cond_counter, // This indexes the conditional in the list of conditionals (bc)
-                        bf_counter,   // because top level
-                        start,
-                        meta_nodes.clone(),
-                    );
-                    cond_counter += 1;
-                }
-            }
-            // make loops if they exist
-            // might need to make wiring more robust
-            if gromet.modules[0].r#fn.bl.as_ref().is_some() {
-                let mut while_counter = 0;
-                for _while_l in gromet.modules[0].r#fn.bl.as_ref().unwrap().iter() {
-                    // now lets check for and setup any conditionals at this level
-                    (nodes, edges, start, meta_nodes) = create_while_loop(
-                        &gromet.clone(),
-                        eboxf.clone(), // This is gromet but is more generalizable based on scope
-                        nodes.clone(),
-                        edges.clone(),
-                        n1.clone(),
-                        att_idx - 1,   // because top level
-                        while_counter, // This indexes the conditional in the list of conditionals (bc)
-                        bf_counter,    // because top level
-                        start,
-                        meta_nodes.clone(),
-                    );
-                    while_counter += 1;
-                }
-            }
-            (nodes, edges, start, meta_nodes)
+            cross_att_wiring(
+                att_box.clone(),
+                nodes,
+                edges,
+                c_args.att_idx,
+                c_args.bf_counter,
+            );
         }
     }
 }
@@ -1886,17 +1501,20 @@ pub fn create_function(
 // The iterator through the conditionals will need to be outside this function
 #[allow(unused_assignments)]
 pub fn create_conditional(
-    gromet: &ModuleCollection,
-    function_net: FunctionNet, // This is gromet but is more generalizable based on scope
-    mut nodes: Vec<Node>,
-    mut edges: Vec<Edge>,
-    parent_node: Node,
-    idx_in: u32,       // This will index the attribute the conditional is in, if any
-    cond_counter: u32, // This indexes the conditional in the list of conditionals (bc)
-    bf_counter: u8,    // This indexes which box the conditional is under, if any
-    mut start: u32,
-    mut meta_nodes: Vec<MetadataNode>,
-) -> (Vec<Node>, Vec<Edge>, u32, Vec<MetadataNode>) {
+    gromet: &ModuleCollection, // needed still for metadata unfortunately
+    nodes: &mut Vec<Node>,
+    edges: &mut Vec<Edge>,
+    meta_nodes: &mut Vec<MetadataNode>,
+    start: &mut u32, // for node and edge indexing
+    c_args: ConstructorArgs,
+    cond_counter: u32,
+) -> () {
+    let mut bf_counter = c_args.bf_counter;
+    let mut att_idx = c_args.att_idx;
+    let mut att_bf_idx = c_args.att_bf_idx;
+    let mut box_counter = c_args.box_counter;
+    let function_net = c_args.att_box.clone();
+
     let mut metadata_idx = 0;
     let n1 = Node {
         n_type: String::from("Conditional"),
@@ -1905,45 +1523,21 @@ pub fn create_conditional(
         node_id: format!("n{}", start),
         out_idx: None,
         in_indx: None,
-        contents: idx_in + 1,
+        contents: att_idx,
         nbox: bf_counter,
-        att_bf_idx: 0,
+        att_bf_idx: att_bf_idx,
+        box_counter: box_counter,
     };
     let e1 = Edge {
-        src: parent_node.node_id,
+        src: c_args.parent_node.node_id.clone(),
         tgt: format!("n{}", start),
         e_type: String::from("Contains"),
-        prop: Some(cond_counter),
+        prop: Some(cond_counter as usize),
     };
     nodes.push(n1.clone());
     edges.push(e1);
-    if function_net.bc.as_ref().unwrap()[cond_counter as usize]
-        .metadata
-        .as_ref()
-        .is_some()
-    {
-        metadata_idx = function_net.bc.as_ref().unwrap()[cond_counter as usize]
-            .metadata
-            .unwrap();
-        let mut repeat_meta = false;
-        for node in meta_nodes.iter() {
-            if node.metadata_idx == metadata_idx {
-                repeat_meta = true;
-            }
-        }
-        if !repeat_meta {
-            meta_nodes.append(&mut create_metadata_node(&gromet.clone(), metadata_idx));
-            let me1 = Edge {
-                src: n1.node_id.clone(),
-                tgt: format!("m{}", metadata_idx),
-                e_type: String::from("Metadata"),
-                prop: None,
-            };
-            edges.push(me1);
-        }
-    }
 
-    start += 1;
+    *start += 1;
 
     // now we make the pic and poc ports and connect them to the conditional node
     if function_net.pic.is_some() {
@@ -1963,9 +1557,10 @@ pub fn create_conditional(
                 node_id: format!("n{}", start),
                 out_idx: None,
                 in_indx: Some([port_count].to_vec()),
-                contents: idx_in + 1,
+                contents: att_idx,
                 nbox: bf_counter,
                 att_bf_idx: 0,
+                box_counter: 0,
             };
             let e3 = Edge {
                 src: n1.node_id.clone(),
@@ -1996,7 +1591,7 @@ pub fn create_conditional(
             }
 
             port_count += 1;
-            start += 1;
+            *start += 1;
         }
     }
     if function_net.poc.is_some() {
@@ -2015,9 +1610,10 @@ pub fn create_conditional(
                 node_id: format!("n{}", start),
                 out_idx: Some([port_count].to_vec()),
                 in_indx: None,
-                contents: idx_in + 1,
+                contents: att_idx,
                 nbox: bf_counter,
                 att_bf_idx: 0,
+                box_counter: 0,
             };
             let e5 = Edge {
                 src: n1.node_id.clone(),
@@ -2047,7 +1643,7 @@ pub fn create_conditional(
                 }
             }
             port_count += 1;
-            start += 1;
+            *start += 1;
         }
     }
 
@@ -2066,45 +1662,53 @@ pub fn create_conditional(
     let body_else_box = function_net.bc.as_ref().unwrap()[cond_counter as usize]
         .body_else
         .unwrap();
+
+    // now we make the conditional, if, and else nodes
+
+    let cond_att_box = gromet.modules[0].attributes[(cond_box - 1) as usize].clone();
+    let if_att_box = gromet.modules[0].attributes[(body_if_box - 1) as usize].clone();
+    let else_att_box = gromet.modules[0].attributes[(body_else_box - 1) as usize].clone();
+
+    let mut new_c_args = c_args.clone();
+    // creating the conditional
+    new_c_args.att_box = cond_att_box.clone();
+    new_c_args.att_idx = cond_box as usize;
+    create_att_predicate(&gromet, nodes, edges, meta_nodes, start, new_c_args.clone());
+    *start += 1;
+
+    // creating the if body
+    new_c_args.att_box = if_att_box.clone();
+    new_c_args.att_idx = body_if_box as usize;
+    create_function(&gromet, nodes, edges, meta_nodes, start, new_c_args.clone());
+    *start += 1;
+
+    // creating the else body
+    new_c_args.att_box = else_att_box.clone();
+    new_c_args.att_idx = body_else_box as usize;
+    create_function(&gromet, nodes, edges, meta_nodes, start, new_c_args.clone());
+    *start += 1;
+
     for node in nodes.clone() {
         // make sure the node is in the same attribute as conditional
-        if node.contents
-            == function_net.bf.as_ref().unwrap()[(cond_box - 1) as usize]
-                .contents
-                .unwrap()
-        {
+        if node.contents == cond_box as usize {
             // make sure box matches correctly
-            if (node.nbox as u32) == cond_box {
-                // make sure we don't pick up ports by only getting the predicate
-                if node.n_type == *"Predicate" {
-                    condition_id = node.node_id.clone();
-                }
+            // make sure we don't pick up ports by only getting the predicate
+            if node.n_type == *"Predicate" {
+                condition_id = node.node_id.clone();
             }
         }
-        if node.contents
-            == function_net.bf.as_ref().unwrap()[(body_if_box - 1) as usize]
-                .contents
-                .unwrap()
-        {
+        if node.contents == body_if_box as usize {
             // make sure box matches correctly
-            if (node.nbox as u32) == body_if_box {
-                // make sure we don't pick up ports by only getting the predicate
-                if node.n_type == *"Function" {
-                    body_if_id = node.node_id.clone();
-                }
+            // make sure we don't pick up ports by only getting the predicate
+            if node.n_type == *"Function" {
+                body_if_id = node.node_id.clone();
             }
         }
-        if node.contents
-            == function_net.bf.as_ref().unwrap()[(body_else_box - 1) as usize]
-                .contents
-                .unwrap()
-        {
+        if node.contents == body_else_box as usize {
             // make sure box matches correctly
-            if (node.nbox as u32) == body_else_box {
-                // make sure we don't pick up ports by only getting the predicate
-                if node.n_type == *"Function" {
-                    body_else_id = node.node_id.clone();
-                }
+            // make sure we don't pick up ports by only getting the predicate
+            if node.n_type == *"Function" {
+                body_else_id = node.node_id.clone();
             }
         }
     }
@@ -2136,7 +1740,7 @@ pub fn create_conditional(
     for wire in function_net.wfc.as_ref().unwrap().iter() {
         // collect info to identify the opi src node
         let src_idx = wire.src; // port index
-        let src_att = idx_in + 1; // attribute index of submodule (also opi contents value)
+        let src_att = att_idx; // attribute index of submodule (also opi contents value)
         let src_nbox = bf_counter; // nbox value of src opi
         let src_pic_idx = src_idx;
 
@@ -2174,9 +1778,9 @@ pub fn create_conditional(
         }
         for node in nodes.iter() {
             // make sure in correct box
-            if tgt_nbox == node.nbox {
+            if tgt_nbox == node.nbox as u8 {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if tgt_att == node.contents {
+                if tgt_att == node.contents as u32 {
                     // only opo's
                     if node.n_type == "Opo" {
                         // iterate through port to check for tgt
@@ -2201,79 +1805,13 @@ pub fn create_conditional(
         }
     }
     // now to perform the wl_cargs wiring which is a connection from the condition's pif/opi to the pic
-    for wire in function_net.wl_cargs.as_ref().unwrap().iter() {
-        // collect info to identify the opi src node
-        let src_idx = wire.src; // port index, this points into bc I think
-        let src_pif = function_net.pif.as_ref().unwrap()[(src_idx - 1) as usize].clone(); // attribute index of submodule (also opi contents value)
-        let src_opi_idx = src_pif.id.unwrap();
-        let src_box = src_pif.r#box; // nbox value of src opi
-        let src_att = function_net.bf.as_ref().unwrap()[(src_box - 1) as usize]
-            .contents
-            .unwrap();
-        let src_nbox = src_box;
-
-        let tgt_idx = wire.tgt; // port index
-        let _tgt_pic = function_net.pic.as_ref().unwrap()[(tgt_idx - 1) as usize].clone(); // tgt port
-        let tgt_box = bf_counter; // tgt sub module box number
-
-        let tgt_att = idx_in + 1; // attribute index of submodule (also opo contents value)
-        let tgt_nbox = tgt_box; // nbox value of tgt opo
-        let tgt_pic_idx = tgt_idx;
-
-        // now to construct the wire
-        let mut wl_cargs_src_tgt: Vec<String> = vec![];
-        // find the src node
-        for node in nodes.iter() {
-            // make sure in correct box
-            if src_nbox == node.nbox {
-                // make sure only looking in current attribute nodes for srcs and tgts
-                if src_att == node.contents {
-                    // only opo's
-                    if node.n_type == "Opi" {
-                        // iterate through port to check for tgt
-                        for p in node.in_indx.as_ref().unwrap().iter() {
-                            // push the src first, being pif
-                            if (src_opi_idx as u32) == *p {
-                                wl_cargs_src_tgt.push(node.node_id.clone());
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        for node in nodes.iter() {
-            // make sure in correct box
-            if tgt_nbox == node.nbox {
-                // make sure only looking in current attribute nodes for srcs and tgts
-                if tgt_att == node.contents {
-                    // only opo's
-                    if node.n_type == "Pic" {
-                        // iterate through port to check for tgt
-                        for p in node.in_indx.as_ref().unwrap().iter() {
-                            // push the src first, being pif
-                            if (tgt_pic_idx as u32) == *p {
-                                wl_cargs_src_tgt.push(node.node_id.clone());
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        if wl_cargs_src_tgt.len() == 2 {
-            let e9 = Edge {
-                src: wl_cargs_src_tgt[0].clone(),
-                tgt: wl_cargs_src_tgt[1].clone(),
-                e_type: String::from("Wire"),
-                prop: None,
-            };
-            edges.push(e9);
-        }
-    }
+    // This needs to be redone, as wl_cargs no longer exist.
 
     // now to make the implicit wires that go from pics -> pifs/opis and pofs/opos -> pocs.
     // first we will iterate through the pics
+    // This will need to be redone as well, due to contents refering to attributes instead of bf entries.
     let mut pic_counter = 1;
-    for _pic in function_net.pic.as_ref().unwrap().iter() {
+    /*  for _pic in function_net.pic.as_ref().unwrap().iter() {
         // collect info to identify the opi src node
         // Each pic is the target there will then be 2 srcs one for each wire, one going to "if" and one to "else"
 
@@ -2302,7 +1840,7 @@ pub fn create_conditional(
         // setting up the pic is straight forward
         let tgt_idx = pic_counter; // port index
         let tgt_box = bf_counter; // tgt sub module box number
-        let tgt_att = idx_in + 1; // attribute index of submodule (also opo contents value)
+        let tgt_att = att_idx; // attribute index of submodule (also opo contents value)
         let tgt_nbox = tgt_box; // nbox value of tgt opo
         let tgt_pic_idx = tgt_idx;
 
@@ -2314,7 +1852,7 @@ pub fn create_conditional(
             // make sure in correct box
             if src_if_nbox == (node.nbox as u32) {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if src_if_att == node.contents {
+                if src_if_att == node.contents as u32 {
                     // only opo's
                     if node.n_type == "Opi" {
                         // iterate through port to check for tgt
@@ -2333,7 +1871,7 @@ pub fn create_conditional(
             // make sure in correct box
             if src_else_nbox == (node.nbox as u32) {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if src_else_att == node.contents {
+                if src_else_att == node.contents as u32 {
                     // only opo's
                     if node.n_type == "Opi" {
                         // iterate through port to check for tgt
@@ -2386,11 +1924,12 @@ pub fn create_conditional(
             edges.push(e11);
         }
         pic_counter += 1;
-    }
+    }*/
 
     // now we construct the output wires for the bodies
+    // this also needs to be fixed for referencing the attributes and not the bf entries
     let mut poc_counter = 1;
-    for _poc in function_net.poc.as_ref().unwrap().iter() {
+    /*  for _poc in function_net.poc.as_ref().unwrap().iter() {
         // collect info to identify the opi src node
         // Each pic is the target there will then be 2 srcs one for each wire, one going to "if" and one to "else"
 
@@ -2419,7 +1958,7 @@ pub fn create_conditional(
         // setting up the pic is straight forward
         let src_idx = poc_counter; // port index
         let src_box = bf_counter; // tgt sub module box number
-        let src_att = idx_in + 1; // attribute index of submodule (also opo contents value)
+        let src_att = att_idx; // attribute index of submodule (also opo contents value)
         let src_nbox = src_box; // nbox value of tgt opo
         let src_poc_idx = src_idx;
 
@@ -2451,7 +1990,7 @@ pub fn create_conditional(
             // make sure in correct box
             if tgt_else_nbox == (node.nbox as u32) {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if tgt_else_att == node.contents {
+                if tgt_else_att == node.contents as u32 {
                     // only opo's
                     if node.n_type == "Opo" {
                         // iterate through port to check for tgt
@@ -2470,7 +2009,7 @@ pub fn create_conditional(
             // make sure in correct box
             if tgt_if_nbox == (node.nbox as u32) {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if tgt_if_att == node.contents {
+                if tgt_if_att == node.contents as u32 {
                     // only opo's
                     if node.n_type == "Opo" {
                         // iterate through port to check for tgt
@@ -2503,24 +2042,25 @@ pub fn create_conditional(
             edges.push(e13);
         }
         poc_counter += 1;
-    }
+    } */
     // might still need pass through wiring??
-
-    (nodes, edges, start, meta_nodes)
 }
 #[allow(unused_assignments)]
 pub fn create_while_loop(
-    gromet: &ModuleCollection,
-    function_net: FunctionNet, // This is gromet but is more generalizable based on scope
-    mut nodes: Vec<Node>,
-    mut edges: Vec<Edge>,
-    parent_node: Node,
-    idx_in: u32,       // This will index the attribute the conditional is in, if any
-    cond_counter: u32, // This indexes the conditional in the list of conditionals (bc)
-    bf_counter: u8,    // This indexes which box the conditional is under, if any
-    mut start: u32,
-    mut meta_nodes: Vec<MetadataNode>,
-) -> (Vec<Node>, Vec<Edge>, u32, Vec<MetadataNode>) {
+    gromet: &ModuleCollection, // needed still for metadata unfortunately
+    nodes: &mut Vec<Node>,
+    edges: &mut Vec<Edge>,
+    meta_nodes: &mut Vec<MetadataNode>,
+    start: &mut u32, // for node and edge indexing
+    c_args: ConstructorArgs,
+    cond_counter: u32,
+) -> () {
+    let mut bf_counter = c_args.bf_counter;
+    let mut att_idx = c_args.att_idx;
+    let mut att_bf_idx = c_args.att_bf_idx;
+    let mut box_counter = c_args.box_counter;
+    let function_net = c_args.att_box.clone();
+
     let mut metadata_idx = 0;
     let n1 = Node {
         n_type: String::from("While_Loop"),
@@ -2529,15 +2069,16 @@ pub fn create_while_loop(
         node_id: format!("n{}", start),
         out_idx: None,
         in_indx: None,
-        contents: idx_in + 1,
+        contents: att_idx,
         nbox: bf_counter,
         att_bf_idx: 0,
+        box_counter: 0,
     };
     let e1 = Edge {
-        src: parent_node.node_id,
+        src: c_args.parent_node.node_id.clone(),
         tgt: format!("n{}", start),
         e_type: String::from("Contains"),
-        prop: Some(cond_counter),
+        prop: Some(cond_counter as usize),
     };
     nodes.push(n1.clone());
     edges.push(e1);
@@ -2567,7 +2108,7 @@ pub fn create_while_loop(
         }
     }
 
-    start += 1;
+    *start += 1;
 
     // now we make the pic and poc ports and connect them to the conditional node
     if function_net.pil.is_some() {
@@ -2587,9 +2128,10 @@ pub fn create_while_loop(
                 node_id: format!("n{}", start),
                 out_idx: None,
                 in_indx: Some([port_count].to_vec()),
-                contents: idx_in + 1,
+                contents: att_idx,
                 nbox: bf_counter,
                 att_bf_idx: 0,
+                box_counter: 0,
             };
             let e3 = Edge {
                 src: n1.node_id.clone(),
@@ -2619,7 +2161,7 @@ pub fn create_while_loop(
                 }
             }
             port_count += 1;
-            start += 1;
+            *start += 1;
         }
     }
     if function_net.pol.is_some() {
@@ -2638,9 +2180,10 @@ pub fn create_while_loop(
                 node_id: format!("n{}", start),
                 out_idx: Some([port_count].to_vec()),
                 in_indx: None,
-                contents: idx_in + 1,
+                contents: att_idx,
                 nbox: bf_counter,
                 att_bf_idx: 0,
+                box_counter: 0,
             };
             let e5 = Edge {
                 src: n1.node_id.clone(),
@@ -2670,7 +2213,7 @@ pub fn create_while_loop(
                 }
             }
             port_count += 1;
-            start += 1;
+            *start += 1;
         }
     }
 
@@ -2685,32 +2228,37 @@ pub fn create_while_loop(
     let body_if_box = function_net.bl.as_ref().unwrap()[cond_counter as usize]
         .body
         .unwrap();
+
+    let cond_att_box = gromet.modules[0].attributes[(cond_box - 1) as usize].clone();
+    let body_att_box = gromet.modules[0].attributes[(body_if_box - 1) as usize].clone();
+
+    let mut new_c_args = c_args.clone();
+    // creating the conditional
+    new_c_args.att_box = cond_att_box.clone();
+    new_c_args.att_idx = cond_box as usize;
+    create_att_predicate(&gromet, nodes, edges, meta_nodes, start, new_c_args.clone());
+    *start += 1;
+
+    // creating the if body
+    new_c_args.att_box = body_att_box.clone();
+    new_c_args.att_idx = body_if_box as usize;
+    create_function(&gromet, nodes, edges, meta_nodes, start, new_c_args.clone());
+    *start += 1;
+
     for node in nodes.clone() {
         // make sure the node is in the same attribute as conditional
-        if node.contents
-            == function_net.bf.as_ref().unwrap()[(cond_box - 1) as usize]
-                .contents
-                .unwrap()
-        {
-            // make sure box matches correctly
-            if (node.nbox as u32) == cond_box {
-                // make sure we don't pick up ports by only getting the predicate
-                if node.n_type == *"Predicate" {
-                    condition_id = node.node_id.clone();
-                }
+        // make sure box matches correctly
+        if node.contents == cond_box as usize {
+            // make sure we don't pick up ports by only getting the predicate
+            if node.n_type == *"Predicate" {
+                condition_id = node.node_id.clone();
             }
         }
-        if node.contents
-            == function_net.bf.as_ref().unwrap()[(body_if_box - 1) as usize]
-                .contents
-                .unwrap()
-        {
-            // make sure box matches correctly
-            if (node.nbox as u32) == body_if_box {
-                // make sure we don't pick up ports by only getting the predicate
-                if node.n_type == *"Function" {
-                    body_if_id = node.node_id.clone();
-                }
+        // make sure box matches correctly
+        if node.contents == body_if_box as usize {
+            // make sure we don't pick up ports by only getting the predicate
+            if node.n_type == *"Function" {
+                body_if_id = node.node_id.clone();
             }
         }
     }
@@ -2735,7 +2283,7 @@ pub fn create_while_loop(
     for wire in function_net.wfl.as_ref().unwrap().iter() {
         // collect info to identify the opi src node
         let src_idx = wire.src; // port index
-        let src_att = idx_in + 1; // attribute index of submodule (also opi contents value)
+        let src_att = att_idx; // attribute index of submodule (also opi contents value)
         let src_nbox = bf_counter; // nbox value of src opi
         let src_pil_idx = src_idx;
 
@@ -2773,9 +2321,9 @@ pub fn create_while_loop(
         }
         for node in nodes.iter() {
             // make sure in correct box
-            if tgt_nbox == node.nbox {
+            if tgt_nbox == node.nbox as u8 {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if tgt_att == node.contents {
+                if tgt_att == node.contents as u32 {
                     // only opo's
                     if node.n_type == "Opo" {
                         // iterate through port to check for tgt
@@ -2800,7 +2348,7 @@ pub fn create_while_loop(
         }
     }
     // now to perform the wl_cargs wiring which is a connection from the condition's pif/opi to the pic
-    for wire in function_net.wl_cargs.as_ref().unwrap().iter() {
+    /*for wire in function_net.wl_cargs.as_ref().unwrap().iter() {
         // collect info to identify the opi src node
         let src_idx = wire.src; // port index, this points into bc I think
         let src_pif = function_net.pif.as_ref().unwrap()[(src_idx - 1) as usize].clone(); // attribute index of submodule (also opi contents value)
@@ -2815,7 +2363,7 @@ pub fn create_while_loop(
         let _tgt_pil = function_net.pil.as_ref().unwrap()[(tgt_idx - 1) as usize].clone(); // tgt port
         let tgt_box = bf_counter; // tgt sub module box number
 
-        let tgt_att = idx_in + 1; // attribute index of submodule (also opo contents value)
+        let tgt_att = att_idx; // attribute index of submodule (also opo contents value)
         let tgt_nbox = tgt_box; // nbox value of tgt opo
         let tgt_pil_idx = tgt_idx;
 
@@ -2824,9 +2372,9 @@ pub fn create_while_loop(
         // find the src node
         for node in nodes.iter() {
             // make sure in correct box
-            if src_nbox == node.nbox {
+            if src_nbox == node.nbox as u8 {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if src_att == node.contents {
+                if src_att == node.contents as u32 {
                     // only opo's
                     if node.n_type == "Opi" {
                         // iterate through port to check for tgt
@@ -2867,12 +2415,12 @@ pub fn create_while_loop(
             };
             edges.push(e9);
         }
-    }
+    } */
 
     // now to make the implicit wires that go from pics -> pifs/opis and pofs/opos -> pocs.
     // first we will iterate through the pics
     let mut pic_counter = 1;
-    for _pic in function_net.pil.as_ref().unwrap().iter() {
+    /*for _pic in function_net.pil.as_ref().unwrap().iter() {
         // collect info to identify the opi src node
         // Each pic is the target there will then be 2 srcs one for each wire, one going to "if" and one to "else"
 
@@ -2892,7 +2440,7 @@ pub fn create_while_loop(
         // setting up the pic is straight forward
         let tgt_idx = pic_counter; // port index
         let tgt_box = bf_counter; // tgt sub module box number
-        let tgt_att = idx_in + 1; // attribute index of submodule (also opo contents value)
+        let tgt_att = att_idx; // attribute index of submodule (also opo contents value)
         let tgt_nbox = tgt_box; // nbox value of tgt opo
         let tgt_pil_idx = tgt_idx;
 
@@ -2903,7 +2451,7 @@ pub fn create_while_loop(
             // make sure in correct box
             if src_if_nbox == (node.nbox as u32) {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if src_if_att == node.contents {
+                if src_if_att == node.contents as u32 {
                     // only opo's
                     if node.n_type == "Opi" {
                         // iterate through port to check for tgt
@@ -2946,11 +2494,11 @@ pub fn create_while_loop(
             edges.push(e10);
         }
         pic_counter += 1;
-    }
+    } */
 
     // now we construct the output wires for the bodies
     let mut poc_counter = 1;
-    for _poc in function_net.pol.as_ref().unwrap().iter() {
+    /*for _poc in function_net.pol.as_ref().unwrap().iter() {
         // collect info to identify the opi src node
         // Each pic is the target there will then be 2 srcs one for each wire, one going to "if" and one to "else"
 
@@ -2970,7 +2518,7 @@ pub fn create_while_loop(
         // setting up the pic is straight forward
         let src_idx = poc_counter; // port index
         let src_box = bf_counter; // tgt sub module box number
-        let src_att = idx_in + 1; // attribute index of submodule (also opo contents value)
+        let src_att = att_idx; // attribute index of submodule (also opo contents value)
         let src_nbox = src_box; // nbox value of tgt opo
         let src_pol_idx = src_idx;
 
@@ -3000,7 +2548,7 @@ pub fn create_while_loop(
             // make sure in correct box
             if tgt_if_nbox == (node.nbox as u32) {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if tgt_if_att == node.contents {
+                if tgt_if_att == node.contents as u32 {
                     // only opo's
                     if node.n_type == "Opo" {
                         // iterate through port to check for tgt
@@ -3024,30 +2572,21 @@ pub fn create_while_loop(
             edges.push(e12);
         }
         poc_counter += 1;
-    }
+    } */
     // might still need pass through wiring??
-
-    (nodes, edges, start, meta_nodes)
 }
 
 // This needs to be updated to handle the new node structure and remove the overloaded contents field which will mess with the wiring alot
 #[allow(unused_assignments)]
 pub fn create_att_expression(
-    gromet: &ModuleCollection,
-    _eeboxf: FunctionNet,
-    ssboxf: GrometBox,
-    mut nodes: Vec<Node>,
-    mut edges: Vec<Edge>,
-    parent_node: Node,
-    _idx_in: u32,
-    box_counter: u8,
-    bf_counter: u8,
-    mut start: u32,
-    mut meta_nodes: Vec<MetadataNode>,
-) -> (Vec<Node>, Vec<Edge>, u32, Vec<MetadataNode>) {
+    gromet: &ModuleCollection, // needed still for metadata unfortunately
+    nodes: &mut Vec<Node>,
+    edges: &mut Vec<Edge>,
+    meta_nodes: &mut Vec<MetadataNode>,
+    start: &mut u32, // for node and edge indexing
+    c_args: ConstructorArgs,
+) -> () {
     let mut metadata_idx = 0;
-
-    let idx = ssboxf.contents.unwrap(); // this is the attribute index of the current expression
 
     let n1 = Node {
         n_type: String::from("Expression"),
@@ -3056,20 +2595,24 @@ pub fn create_att_expression(
         node_id: format!("n{}", start),
         out_idx: None,
         in_indx: None,
-        contents: idx, // this is the attribute index, USED TO BE PARENT ATTRIBUTE INDEX
-        nbox: bf_counter, // inherited from parent
-        att_bf_idx: box_counter as u32, // This is a reference to the index in parent function
+        contents: c_args.att_idx, // this is the attribute index, USED TO BE PARENT ATTRIBUTE INDEX
+        nbox: c_args.bf_counter,  // inherited from parent
+        att_bf_idx: c_args.att_bf_idx, // This is a reference to the index in parent function
+        box_counter: c_args.box_counter,
     };
     let e1 = Edge {
-        src: parent_node.node_id.clone(),
+        src: c_args.parent_node.node_id.clone(),
         tgt: format!("n{}", start),
         e_type: String::from("Contains"),
-        prop: Some(idx),
+        prop: Some(c_args.att_idx),
     };
     nodes.push(n1.clone());
     edges.push(e1);
-    if ssboxf.metadata.as_ref().is_some() {
-        metadata_idx = ssboxf.metadata.unwrap();
+
+    let exp_box = c_args.cur_box.clone();
+
+    if exp_box.metadata.as_ref().is_some() {
+        metadata_idx = exp_box.metadata.unwrap();
         let mut repeat_meta = false;
         for node in meta_nodes.iter() {
             if node.metadata_idx == metadata_idx {
@@ -3090,31 +2633,44 @@ pub fn create_att_expression(
     // now travel to contents index of the attribute list (note it is 1 index,
     // so contents=1 => attribute[0])
     // create nodes and edges for this entry, include opo's and opi's
-    start += 1;
+    *start += 1;
 
-    let eboxf = gromet.modules[0].attributes[(idx - 1) as usize].clone(); // current expression attribute
-    let pboxf = gromet.modules[0].attributes[(parent_node.contents - 1) as usize].clone(); // parent attribute
+    let att_box = gromet.modules[0].attributes[(c_args.att_idx - 1) as usize].clone(); // current expression attribute
+    let parent_att_box =
+        gromet.modules[0].attributes[(c_args.parent_node.contents) as usize].clone(); // parent attribute
 
     // construct opo nodes, if not none
     // not calling the opo port constuctors since they are based on grabbing in the name from top level of the gromet,
     // not a parental attribute
-    if eboxf.opo.is_some() {
+    if att_box.opo.is_some() {
         // grab name which is one level up and based on indexing
         // this can be done by the parent nodes contents field should give the index of the attributes
         // this constructs a vec for the names in the pofs, if any.
         let mut opo_name: Vec<String> = vec![];
-        for port in pboxf.pof.as_ref().unwrap().iter() {
-            if port.r#box == box_counter {
-                if port.name.is_some() {
-                    opo_name.push(port.name.as_ref().unwrap().clone());
-                } else {
-                    opo_name.push(String::from("un-named"));
+        if c_args.att_bf_idx != 0 {
+            for port in parent_att_box.pof.as_ref().unwrap().iter() {
+                if port.r#box == c_args.box_counter as u8 {
+                    if port.name.is_some() {
+                        opo_name.push(port.name.as_ref().unwrap().clone());
+                    } else {
+                        opo_name.push(String::from("un-named"));
+                    }
+                }
+            }
+        } else {
+            for port in gromet.modules[0].r#fn.pof.as_ref().unwrap().iter() {
+                if port.r#box == c_args.bf_counter as u8 {
+                    if port.name.is_some() {
+                        opo_name.push(port.name.as_ref().unwrap().clone());
+                    } else {
+                        opo_name.push(String::from("un-named"));
+                    }
                 }
             }
         }
         if !opo_name.clone().is_empty() {
             let mut oport: u32 = 0;
-            for _op in eboxf.opo.as_ref().unwrap().iter() {
+            for _op in att_box.opo.as_ref().unwrap().iter() {
                 let n2 = Node {
                     n_type: String::from("Opo"),
                     value: None,
@@ -3122,9 +2678,10 @@ pub fn create_att_expression(
                     node_id: format!("n{}", start),
                     out_idx: Some([oport + 1].to_vec()),
                     in_indx: None,
-                    contents: idx,                  // current att idx
-                    nbox: bf_counter,               // top parent top level idx
-                    att_bf_idx: box_counter as u32, // current box idx of parent
+                    contents: c_args.att_idx,      // current att idx
+                    nbox: c_args.bf_counter,       // top parent top level idx
+                    att_bf_idx: c_args.att_bf_idx, // current box idx of parent
+                    box_counter: c_args.box_counter,
                 };
                 nodes.push(n2.clone());
                 // construct edge: expression -> Opo
@@ -3135,13 +2692,15 @@ pub fn create_att_expression(
                     prop: None,
                 };
                 edges.push(e3);
-                if eboxf.opo.clone().as_ref().unwrap()[oport as usize]
+                if att_box.opo.clone().as_ref().unwrap()[oport as usize]
                     .metadata
                     .clone()
                     .as_ref()
                     .is_some()
                 {
-                    metadata_idx = eboxf.opo.clone().unwrap()[oport as usize].metadata.unwrap();
+                    metadata_idx = att_box.opo.clone().unwrap()[oport as usize]
+                        .metadata
+                        .unwrap();
                     let mut repeat_meta = false;
                     for node in meta_nodes.iter() {
                         if node.metadata_idx == metadata_idx {
@@ -3160,16 +2719,16 @@ pub fn create_att_expression(
                     }
                 }
                 // construct any metadata edges
-                start += 1;
+                *start += 1;
                 oport += 1;
             }
         }
     }
     // construct opi nodes, in not none
-    if eboxf.opi.is_some() {
+    if att_box.opi.is_some() {
         // grab name which is NOT one level up as in opo
         let mut opi_name: Vec<String> = vec![];
-        for port in eboxf.opi.as_ref().unwrap().iter() {
+        for port in att_box.opi.as_ref().unwrap().iter() {
             if port.name.is_some() {
                 opi_name.push(port.name.as_ref().unwrap().clone());
             } else {
@@ -3177,7 +2736,7 @@ pub fn create_att_expression(
             }
         }
         let mut iport: u32 = 0;
-        for _op in eboxf.opi.as_ref().unwrap().iter() {
+        for _op in att_box.opi.as_ref().unwrap().iter() {
             let n2 = Node {
                 n_type: String::from("Opi"),
                 value: None,
@@ -3185,9 +2744,10 @@ pub fn create_att_expression(
                 node_id: format!("n{}", start),
                 out_idx: None,
                 in_indx: Some([iport + 1].to_vec()),
-                contents: idx,
-                nbox: bf_counter,
-                att_bf_idx: box_counter as u32,
+                contents: c_args.att_idx,
+                nbox: c_args.bf_counter,
+                att_bf_idx: c_args.att_bf_idx,
+                box_counter: c_args.box_counter,
             };
             nodes.push(n2.clone());
             // construct edge: expression -> Opo
@@ -3198,13 +2758,15 @@ pub fn create_att_expression(
                 prop: None,
             };
             edges.push(e3);
-            if eboxf.opi.clone().as_ref().unwrap()[iport as usize]
+            if att_box.opi.clone().as_ref().unwrap()[iport as usize]
                 .metadata
                 .clone()
                 .as_ref()
                 .is_some()
             {
-                metadata_idx = eboxf.opi.clone().unwrap()[iport as usize].metadata.unwrap();
+                metadata_idx = att_box.opi.clone().unwrap()[iport as usize]
+                    .metadata
+                    .unwrap();
                 let mut repeat_meta = false;
                 for node in meta_nodes.iter() {
                     if node.metadata_idx == metadata_idx {
@@ -3222,80 +2784,73 @@ pub fn create_att_expression(
                     edges.push(me1);
                 }
             }
-            start += 1;
+            *start += 1;
             iport += 1;
         }
     }
     // now to construct the nodes inside the expression, Literal and Primitives
-    let mut box_counter: u8 = 1;
-    if eboxf.bf.is_some() {
-        for sboxf in eboxf.bf.as_ref().unwrap().iter() {
-            match sboxf.function_type {
+    let mut new_c_args = c_args.clone();
+    let mut box_counter: usize = 1;
+    if att_box.bf.is_some() {
+        for att_sub_box in att_box.bf.as_ref().unwrap().iter() {
+            new_c_args.box_counter = box_counter;
+            new_c_args.cur_box = att_sub_box.clone();
+            if att_sub_box.contents.is_some() {
+                new_c_args.att_idx = att_sub_box.contents.unwrap() as usize;
+            }
+            match att_sub_box.function_type {
                 FunctionType::Literal => {
-                    (nodes, edges, meta_nodes) = create_att_literal(
-                        &gromet.clone(),
-                        eboxf.clone(),
-                        sboxf.clone(),
-                        nodes.clone(),
-                        edges.clone(),
-                        n1.clone(),
-                        idx,
-                        box_counter,
-                        bf_counter,
+                    create_att_literal(
+                        &gromet, // gromet for metadata
+                        nodes,   // nodes
+                        edges,
+                        meta_nodes,
                         start,
-                        meta_nodes.clone(),
+                        new_c_args.clone(),
                     );
                 }
                 FunctionType::Primitive => {
-                    (nodes, edges, meta_nodes) = create_att_primitive(
-                        &gromet.clone(),
-                        eboxf.clone(),
-                        sboxf.clone(),
-                        nodes.clone(),
-                        edges.clone(),
-                        n1.clone(),
-                        idx,
-                        box_counter,
-                        bf_counter,
+                    create_att_primitive(
+                        &gromet, // gromet for metadata
+                        nodes,   // nodes
+                        edges,
+                        meta_nodes,
                         start,
-                        meta_nodes.clone(),
+                        new_c_args.clone(),
                     );
                 }
                 _ => {}
             }
             box_counter += 1;
-            start += 1;
+            *start += 1;
         }
     }
     // Now we perform the internal wiring of this branch
-    edges = internal_wiring(eboxf.clone(), nodes.clone(), edges, idx, bf_counter);
+    internal_wiring(
+        att_box.clone(),
+        nodes,
+        edges,
+        c_args.att_idx,
+        c_args.bf_counter,
+    );
 
     // Now we also perform wopio wiring in case there is an empty expression
-    if eboxf.wopio.is_some() {
-        edges = wopio_wiring(eboxf, nodes.clone(), edges, idx - 1, bf_counter);
+    if att_box.wopio.is_some() {
+        wopio_wiring(att_box, nodes, edges, c_args.att_idx - 1, c_args.bf_counter);
     }
-
-    (nodes, edges, start, meta_nodes)
 }
 
 // This needs to be updated to handle the new node structure and remove the overloaded contents field which will mess with the wiring alot
 #[allow(unused_assignments)]
 pub fn create_att_predicate(
-    gromet: &ModuleCollection,
-    _eeboxf: FunctionNet,
-    ssboxf: GrometBox,
-    mut nodes: Vec<Node>,
-    mut edges: Vec<Edge>,
-    parent_node: Node,
-    _idx_in: u32,
-    box_counter: u8,
-    bf_counter: u8,
-    mut start: u32,
-    mut meta_nodes: Vec<MetadataNode>,
-) -> (Vec<Node>, Vec<Edge>, u32, Vec<MetadataNode>) {
+    gromet: &ModuleCollection, // needed still for metadata unfortunately
+    nodes: &mut Vec<Node>,
+    edges: &mut Vec<Edge>,
+    meta_nodes: &mut Vec<MetadataNode>,
+    start: &mut u32, // for node and edge indexing
+    c_args: ConstructorArgs,
+) -> () {
     let mut metadata_idx = 0;
-
-    let idx = ssboxf.contents.unwrap(); // this is the attribute index of the current expression
 
     let n1 = Node {
         n_type: String::from("Predicate"),
@@ -3304,20 +2859,24 @@ pub fn create_att_predicate(
         node_id: format!("n{}", start),
         out_idx: None,
         in_indx: None,
-        contents: idx, // this is the attribute index, USED TO BE PARENT ATTRIBUTE INDEX
-        nbox: bf_counter, // inherited from parent
-        att_bf_idx: box_counter as u32, // This is a reference to the index in parent function
+        contents: c_args.att_idx,        // this is the attribute index
+        nbox: c_args.bf_counter,         // inherited from parent, should be 0
+        att_bf_idx: c_args.att_bf_idx,   // inherited and should be 0
+        box_counter: c_args.box_counter, // inherited and should be 0
     };
     let e1 = Edge {
-        src: parent_node.node_id.clone(),
+        src: c_args.parent_node.node_id.clone(),
         tgt: format!("n{}", start),
         e_type: String::from("Contains"),
-        prop: Some(idx),
+        prop: Some(c_args.att_idx),
     };
     nodes.push(n1.clone());
     edges.push(e1);
-    if ssboxf.metadata.as_ref().is_some() {
-        metadata_idx = ssboxf.metadata.unwrap();
+
+    let in_att_box = c_args.cur_box.clone();
+
+    if in_att_box.metadata.as_ref().is_some() {
+        metadata_idx = in_att_box.metadata.unwrap();
         let mut repeat_meta = false;
         for node in meta_nodes.iter() {
             if node.metadata_idx == metadata_idx {
@@ -3338,30 +2897,23 @@ pub fn create_att_predicate(
     // now travel to contents index of the attribute list (note it is 1 index,
     // so contents=1 => attribute[0])
     // create nodes and edges for this entry, include opo's and opi's
-    start += 1;
+    *start += 1;
 
-    let eboxf = gromet.modules[0].attributes[(idx - 1) as usize].clone(); // current expression attribute
-    let pboxf = gromet.modules[0].attributes[(parent_node.contents - 1) as usize].clone(); // parent attribute
+    let att_box = gromet.modules[0].attributes[(c_args.att_idx - 1) as usize].clone(); // current expression attribute
 
     // construct opo nodes, if not none
     // not calling the opo port constuctors since they are based on grabbing in the name from top level of the gromet,
     // not a parental attribute
-    if eboxf.opo.is_some() {
+    if att_box.opo.is_some() {
         // grab name which is one level up and based on indexing
         // this can be done by the parent nodes contents field should give the index of the attributes
         // this constructs a vec for the names in the pofs, if any.
         let mut opo_name: Vec<String> = vec![];
-        for port in pboxf.pof.as_ref().unwrap().iter() {
-            if port.r#box == box_counter {
-                if port.name.is_some() {
-                    opo_name.push(port.name.as_ref().unwrap().clone());
-                } else {
-                    opo_name.push(String::from("un-named"));
-                }
-            }
+        for port in att_box.opo.as_ref().unwrap().iter() {
+            opo_name.push(String::from("un-named"));
         }
         let mut oport: u32 = 0;
-        for _op in eboxf.opo.as_ref().unwrap().iter() {
+        for _op in att_box.opo.as_ref().unwrap().iter() {
             let n2 = Node {
                 n_type: String::from("Opo"),
                 value: None,
@@ -3369,9 +2921,10 @@ pub fn create_att_predicate(
                 node_id: format!("n{}", start),
                 out_idx: Some([oport + 1].to_vec()),
                 in_indx: None,
-                contents: idx,                  // current att idx
-                nbox: bf_counter,               // top parent top level idx
-                att_bf_idx: box_counter as u32, // current box idx of parent
+                contents: c_args.att_idx,      // this is the attribute index
+                nbox: c_args.bf_counter,       // inherited from parent
+                att_bf_idx: c_args.att_bf_idx, // This is a reference to the index in parent function
+                box_counter: c_args.box_counter, // current box idx of parent
             };
             nodes.push(n2.clone());
             // construct edge: expression -> Opo
@@ -3382,13 +2935,15 @@ pub fn create_att_predicate(
                 prop: None,
             };
             edges.push(e3);
-            if eboxf.opo.clone().as_ref().unwrap()[oport as usize]
+            if att_box.opo.clone().as_ref().unwrap()[oport as usize]
                 .metadata
                 .clone()
                 .as_ref()
                 .is_some()
             {
-                metadata_idx = eboxf.opo.clone().unwrap()[oport as usize].metadata.unwrap();
+                metadata_idx = att_box.opo.clone().unwrap()[oport as usize]
+                    .metadata
+                    .unwrap();
                 let mut repeat_meta = false;
                 for node in meta_nodes.iter() {
                     if node.metadata_idx == metadata_idx {
@@ -3407,23 +2962,19 @@ pub fn create_att_predicate(
                 }
             }
             // construct any metadata edges
-            start += 1;
+            *start += 1;
             oport += 1;
         }
     }
     // construct opi nodes, in not none
-    if eboxf.opi.is_some() {
+    if att_box.opi.is_some() {
         // grab name which is NOT one level up as in opo
         let mut opi_name: Vec<String> = vec![];
-        for port in eboxf.opi.as_ref().unwrap().iter() {
-            if port.name.is_some() {
-                opi_name.push(port.name.as_ref().unwrap().clone());
-            } else {
-                opi_name.push(String::from("un-named"));
-            }
+        for port in att_box.opi.as_ref().unwrap().iter() {
+            opi_name.push(String::from("un-named"));
         }
         let mut iport: u32 = 0;
-        for _op in eboxf.opi.as_ref().unwrap().iter() {
+        for _op in att_box.opi.as_ref().unwrap().iter() {
             let n2 = Node {
                 n_type: String::from("Opi"),
                 value: None,
@@ -3431,9 +2982,10 @@ pub fn create_att_predicate(
                 node_id: format!("n{}", start),
                 out_idx: None,
                 in_indx: Some([iport + 1].to_vec()),
-                contents: idx,
-                nbox: bf_counter,
-                att_bf_idx: box_counter as u32,
+                contents: c_args.att_idx,      // this is the attribute index
+                nbox: c_args.bf_counter,       // inherited from parent
+                att_bf_idx: c_args.att_bf_idx, // This is a reference to the index in parent function
+                box_counter: c_args.box_counter,
             };
             nodes.push(n2.clone());
             // construct edge: expression -> Opo
@@ -3444,13 +2996,15 @@ pub fn create_att_predicate(
                 prop: None,
             };
             edges.push(e3);
-            if eboxf.opi.clone().as_ref().unwrap()[iport as usize]
+            if att_box.opi.clone().as_ref().unwrap()[iport as usize]
                 .metadata
                 .clone()
                 .as_ref()
                 .is_some()
             {
-                metadata_idx = eboxf.opi.clone().unwrap()[iport as usize].metadata.unwrap();
+                metadata_idx = att_box.opi.clone().unwrap()[iport as usize]
+                    .metadata
+                    .unwrap();
                 let mut repeat_meta = false;
                 for node in meta_nodes.iter() {
                     if node.metadata_idx == metadata_idx {
@@ -3468,75 +3022,69 @@ pub fn create_att_predicate(
                     edges.push(me1);
                 }
             }
-            start += 1;
+            *start += 1;
             iport += 1;
         }
     }
     // now to construct the nodes inside the expression, Literal and Primitives
-    let mut box_counter: u8 = 1;
-    for sboxf in eboxf.bf.as_ref().unwrap().iter() {
-        match sboxf.function_type {
+    let mut new_c_args = c_args.clone();
+    let mut box_counter: usize = 1;
+    for sub_att_box in att_box.bf.as_ref().unwrap().iter() {
+        new_c_args.box_counter = box_counter;
+        new_c_args.cur_box = sub_att_box.clone();
+        match sub_att_box.function_type {
             FunctionType::Literal => {
-                (nodes, edges, meta_nodes) = create_att_literal(
-                    &gromet.clone(),
-                    eboxf.clone(),
-                    sboxf.clone(),
-                    nodes.clone(),
-                    edges.clone(),
-                    n1.clone(),
-                    idx,
-                    box_counter,
-                    bf_counter,
+                create_att_literal(
+                    &gromet, // gromet for metadata
+                    nodes,   // nodes
+                    edges,
+                    meta_nodes,
                     start,
-                    meta_nodes.clone(),
+                    new_c_args.clone(),
                 );
             }
             FunctionType::Primitive => {
-                (nodes, edges, meta_nodes) = create_att_primitive(
-                    &gromet.clone(),
-                    eboxf.clone(),
-                    sboxf.clone(),
-                    nodes.clone(),
-                    edges.clone(),
-                    n1.clone(),
-                    idx,
-                    box_counter,
-                    bf_counter,
+                create_att_primitive(
+                    &gromet, // gromet for metadata
+                    nodes,   // nodes
+                    edges,
+                    meta_nodes,
                     start,
-                    meta_nodes.clone(),
+                    new_c_args.clone(),
                 );
             }
             _ => {}
         }
         box_counter += 1;
-        start += 1;
+        *start += 1;
     }
     // Now we perform the internal wiring of this branch
-    edges = internal_wiring(eboxf, nodes.clone(), edges, idx, bf_counter);
-    (nodes, edges, start, meta_nodes)
+    internal_wiring(
+        att_box,
+        &mut nodes.clone(),
+        edges,
+        c_args.att_idx,
+        c_args.bf_counter,
+    );
 }
 #[allow(unused_assignments)]
 pub fn create_att_literal(
-    gromet: &ModuleCollection,
-    eboxf: FunctionNet,
-    sboxf: GrometBox,
-    mut nodes: Vec<Node>,
-    mut edges: Vec<Edge>,
-    parent_node: Node,
-    idx: u32,
-    box_counter: u8,
-    bf_counter: u8,
-    start: u32,
-    mut meta_nodes: Vec<MetadataNode>,
-) -> (Vec<Node>, Vec<Edge>, Vec<MetadataNode>) {
+    gromet: &ModuleCollection, // needed still for metadata unfortunately
+    nodes: &mut Vec<Node>,
+    edges: &mut Vec<Edge>,
+    meta_nodes: &mut Vec<MetadataNode>,
+    start: &mut u32, // for node and edge indexing
+    c_args: ConstructorArgs,
+) -> () {
+    let att_box = c_args.att_box;
+    let lit_box = c_args.cur_box;
     // first find the pof's for box
     let mut pof: Vec<u32> = vec![];
-    if eboxf.pof.is_some() {
+    if att_box.pof.is_some() {
         let mut po_idx: u32 = 1;
-        for port in eboxf.pof.unwrap().iter() {
-            if port.r#box == box_counter {
+        for port in att_box.pof.unwrap().iter() {
+            if port.r#box == c_args.box_counter as u8 {
                 pof.push(po_idx);
-            } else {
             }
             po_idx += 1;
         }
@@ -3545,26 +3093,27 @@ pub fn create_att_literal(
     let mut metadata_idx = 0;
     let n3 = Node {
         n_type: String::from("Literal"),
-        value: Some(format!("{:?}", sboxf.value.as_ref().unwrap())),
+        value: Some(format!("{:?}", lit_box.value.as_ref().unwrap())),
         name: None,
         node_id: format!("n{}", start),
         out_idx: Some(pof),
         in_indx: None, // literals should only have out ports
-        contents: idx,
-        nbox: bf_counter,
-        att_bf_idx: box_counter as u32,
+        contents: c_args.att_idx,
+        nbox: c_args.bf_counter,
+        att_bf_idx: c_args.att_bf_idx,
+        box_counter: c_args.box_counter,
     };
     nodes.push(n3.clone());
     // make edge connecting to expression
     let e4 = Edge {
-        src: parent_node.node_id,
+        src: c_args.parent_node.node_id,
         tgt: n3.node_id.clone(),
         e_type: String::from("Contains"),
         prop: None,
     };
     edges.push(e4);
-    if sboxf.metadata.is_some() {
-        metadata_idx = sboxf.metadata.unwrap();
+    if lit_box.metadata.is_some() {
+        metadata_idx = lit_box.metadata.unwrap();
         let mut repeat_meta = false;
         for node in meta_nodes.iter() {
             if node.metadata_idx == metadata_idx {
@@ -3582,28 +3131,23 @@ pub fn create_att_literal(
             edges.push(me1);
         }
     }
-    (nodes, edges, meta_nodes)
+    *start += 1;
 }
 #[allow(unused_assignments)]
 pub fn create_att_primitive(
-    gromet: &ModuleCollection,
-    eboxf: FunctionNet,
-    sboxf: GrometBox,
-    mut nodes: Vec<Node>,
-    mut edges: Vec<Edge>,
-    parent_node: Node,
-    idx: u32,
-    box_counter: u8,
-    bf_counter: u8,
-    start: u32,
-    mut meta_nodes: Vec<MetadataNode>,
-) -> (Vec<Node>, Vec<Edge>, Vec<MetadataNode>) {
+    gromet: &ModuleCollection, // needed still for metadata unfortunately
+    nodes: &mut Vec<Node>,
+    edges: &mut Vec<Edge>,
+    meta_nodes: &mut Vec<MetadataNode>,
+    start: &mut u32, // for node and edge indexing
+    c_args: ConstructorArgs,
+) -> () {
     // first find the pof's for box
     let mut pof: Vec<u32> = vec![];
-    if eboxf.pof.is_some() {
+    if c_args.att_box.pof.is_some() {
         let mut po_idx: u32 = 1;
-        for port in eboxf.pof.clone().unwrap().iter() {
-            if port.r#box == box_counter {
+        for port in c_args.att_box.pof.clone().unwrap().iter() {
+            if port.r#box == c_args.box_counter as u8 {
                 pof.push(po_idx);
             }
             po_idx += 1;
@@ -3611,10 +3155,10 @@ pub fn create_att_primitive(
     }
     // then find pif's for box
     let mut pif: Vec<u32> = vec![];
-    if eboxf.pif.is_some() {
+    if c_args.att_box.pif.is_some() {
         let mut pi_idx: u32 = 1;
-        for port in eboxf.pif.unwrap().iter() {
-            if port.r#box == box_counter {
+        for port in c_args.att_box.pif.unwrap().iter() {
+            if port.r#box == c_args.box_counter as u8 {
                 pif.push(pi_idx);
             }
             pi_idx += 1;
@@ -3625,25 +3169,26 @@ pub fn create_att_primitive(
     let n3 = Node {
         n_type: String::from("Primitive"),
         value: None,
-        name: sboxf.name.clone(),
+        name: c_args.cur_box.name.clone(),
         node_id: format!("n{}", start),
         out_idx: Some(pof),
         in_indx: Some(pif),
-        contents: idx,
-        nbox: bf_counter,
-        att_bf_idx: box_counter as u32,
+        contents: c_args.att_idx,
+        nbox: c_args.bf_counter,
+        att_bf_idx: c_args.att_bf_idx,
+        box_counter: c_args.box_counter,
     };
     nodes.push(n3.clone());
     // make edge connecting to expression
     let e4 = Edge {
-        src: parent_node.node_id,
+        src: c_args.parent_node.node_id,
         tgt: n3.node_id.clone(),
         e_type: String::from("Contains"),
         prop: None,
     };
     edges.push(e4);
-    if sboxf.metadata.is_some() {
-        metadata_idx = sboxf.metadata.unwrap();
+    if c_args.cur_box.metadata.is_some() {
+        metadata_idx = c_args.cur_box.metadata.unwrap();
         let mut repeat_meta = false;
         for node in meta_nodes.iter() {
             if node.metadata_idx == metadata_idx {
@@ -3661,29 +3206,28 @@ pub fn create_att_primitive(
             edges.push(me1);
         }
     }
-    (nodes, edges, meta_nodes)
+    *start += 1;
 }
 // This is for the construction of Opo's
 #[allow(unused_assignments)]
 pub fn create_opo(
-    mut nodes: Vec<Node>,
-    mut edges: Vec<Edge>,
-    mut meta_nodes: Vec<MetadataNode>,
-    gromet: &ModuleCollection,
-    mut start: u32,
-    parent_node: Node,
-    box_counter: u8,
-) -> (Vec<Node>, Vec<Edge>, Vec<MetadataNode>, u32) {
-    let eboxf = gromet.modules[0].attributes[(parent_node.contents - 1) as usize].clone();
+    gromet: &ModuleCollection, // needed still for metadata unfortunately
+    nodes: &mut Vec<Node>,
+    edges: &mut Vec<Edge>,
+    meta_nodes: &mut Vec<MetadataNode>,
+    start: &mut u32, // for node and edge indexing
+    c_args: ConstructorArgs,
+) -> () {
+    let att_box = gromet.modules[0].attributes[(c_args.parent_node.contents - 1) as usize].clone();
     // construct opo nodes, if not none
-    if eboxf.opo.is_some() {
+    if att_box.opo.is_some() {
         // grab name which is one level up and based on indexing
         let mut opo_name = "un-named";
         let mut oport: u32 = 0;
-        for op in eboxf.opo.as_ref().unwrap().iter() {
+        for op in att_box.opo.as_ref().unwrap().iter() {
             if op.name.as_ref().is_none() && gromet.modules[0].r#fn.pof.as_ref().is_some() {
                 for port in gromet.modules[0].r#fn.pof.as_ref().unwrap().iter() {
-                    if port.r#box == box_counter
+                    if port.r#box == c_args.bf_counter as u8
                         && oport == (port.id.unwrap() as u32 - 1)
                         && port.name.is_some()
                     {
@@ -3700,14 +3244,15 @@ pub fn create_opo(
                 node_id: format!("n{}", start),
                 out_idx: Some([oport + 1].to_vec()),
                 in_indx: None,
-                contents: parent_node.contents,
-                nbox: box_counter,
-                att_bf_idx: parent_node.att_bf_idx,
+                contents: c_args.parent_node.contents,
+                nbox: c_args.bf_counter,
+                att_bf_idx: c_args.parent_node.att_bf_idx,
+                box_counter: c_args.parent_node.box_counter,
             };
             nodes.push(n2.clone());
             // construct edge: expression -> Opo
             let e3 = Edge {
-                src: parent_node.node_id.clone(),
+                src: c_args.parent_node.node_id.clone(),
                 tgt: n2.node_id.clone(),
                 e_type: String::from("Port_Of"),
                 prop: None,
@@ -3715,13 +3260,15 @@ pub fn create_opo(
             edges.push(e3);
 
             let mut metadata_idx = 0;
-            if eboxf.opo.clone().as_ref().unwrap()[oport as usize]
+            if att_box.opo.clone().as_ref().unwrap()[oport as usize]
                 .metadata
                 .clone()
                 .as_ref()
                 .is_some()
             {
-                metadata_idx = eboxf.opo.clone().unwrap()[oport as usize].metadata.unwrap();
+                metadata_idx = att_box.opo.clone().unwrap()[oport as usize]
+                    .metadata
+                    .unwrap();
                 let mut repeat_meta = false;
                 for node in meta_nodes.iter() {
                     if node.metadata_idx == metadata_idx {
@@ -3740,34 +3287,32 @@ pub fn create_opo(
                 }
             }
             // construct any metadata edges
-            start += 1;
+            *start += 1;
             oport += 1;
         }
     }
-    (nodes, edges, meta_nodes, start)
 }
 
 // This is for the construction of Opi's
 #[allow(unused_assignments)]
 pub fn create_opi(
-    mut nodes: Vec<Node>,
-    mut edges: Vec<Edge>,
-    mut meta_nodes: Vec<MetadataNode>,
-    gromet: &ModuleCollection,
-    mut start: u32,
-    parent_node: Node,
-    box_counter: u8,
-) -> (Vec<Node>, Vec<Edge>, Vec<MetadataNode>, u32) {
-    let eboxf = gromet.modules[0].attributes[(parent_node.contents - 1) as usize].clone();
+    gromet: &ModuleCollection, // needed still for metadata unfortunately
+    nodes: &mut Vec<Node>,
+    edges: &mut Vec<Edge>,
+    meta_nodes: &mut Vec<MetadataNode>,
+    start: &mut u32, // for node and edge indexing
+    c_args: ConstructorArgs,
+) -> () {
+    let att_box = gromet.modules[0].attributes[(c_args.parent_node.contents - 1) as usize].clone();
     // construct opi nodes, if not none
-    if eboxf.opi.is_some() {
+    if att_box.opi.is_some() {
         // grab name which is one level up and based on indexing
         let mut opi_name = "un-named";
         let mut oport: u32 = 0;
-        for op in eboxf.opi.as_ref().unwrap().iter() {
+        for op in att_box.opi.as_ref().unwrap().iter() {
             if op.name.as_ref().is_none() && gromet.modules[0].r#fn.pif.as_ref().is_some() {
                 for port in gromet.modules[0].r#fn.pif.as_ref().unwrap().iter() {
-                    if port.r#box == box_counter
+                    if port.r#box == c_args.bf_counter as u8
                         && oport == (port.id.unwrap() as u32 - 1)
                         && port.name.is_some()
                     {
@@ -3784,14 +3329,15 @@ pub fn create_opi(
                 node_id: format!("n{}", start),
                 out_idx: None,
                 in_indx: Some([oport + 1].to_vec()),
-                contents: parent_node.contents,
-                nbox: box_counter,
-                att_bf_idx: parent_node.att_bf_idx,
+                contents: c_args.parent_node.contents,
+                nbox: c_args.bf_counter,
+                att_bf_idx: c_args.parent_node.att_bf_idx,
+                box_counter: c_args.parent_node.box_counter,
             };
             nodes.push(n2.clone());
             // construct edge: expression <- Opi
             let e3 = Edge {
-                src: parent_node.node_id.clone(),
+                src: c_args.parent_node.node_id.clone(),
                 tgt: n2.node_id.clone(),
                 e_type: String::from("Port_Of"),
                 prop: None,
@@ -3799,13 +3345,15 @@ pub fn create_opi(
             edges.push(e3);
 
             let mut metadata_idx = 0;
-            if eboxf.opi.clone().as_ref().unwrap()[oport as usize]
+            if att_box.opi.clone().as_ref().unwrap()[oport as usize]
                 .metadata
                 .clone()
                 .as_ref()
                 .is_some()
             {
-                metadata_idx = eboxf.opi.clone().unwrap()[oport as usize].metadata.unwrap();
+                metadata_idx = att_box.opi.clone().unwrap()[oport as usize]
+                    .metadata
+                    .unwrap();
                 let mut repeat_meta = false;
                 for node in meta_nodes.iter() {
                     if node.metadata_idx == metadata_idx {
@@ -3824,21 +3372,20 @@ pub fn create_opi(
                 }
             }
             // construct any metadata edges
-            start += 1;
+            *start += 1;
             oport += 1;
         }
     }
-    (nodes, edges, meta_nodes, start)
 }
 // having issues with deeply nested structure, it is breaking in the internal wiring of the function level.
 // wfopi: pif -> opi
 pub fn wfopi_wiring(
     eboxf: FunctionNet,
-    nodes: Vec<Node>,
-    mut edges: Vec<Edge>,
+    nodes: &mut Vec<Node>,
+    edges: &mut Vec<Edge>,
     idx: u32,
     bf_counter: u8,
-) -> Vec<Edge> {
+) -> () {
     // iterate through all wires of type
     for wire in eboxf.wfopi.unwrap().iter() {
         let mut prop = None;
@@ -3846,9 +3393,9 @@ pub fn wfopi_wiring(
         // find the src node
         for node in nodes.iter() {
             // make sure in correct box
-            if bf_counter == node.nbox {
+            if bf_counter == node.nbox as u8 {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if (idx) == node.contents {
+                if (idx) == node.contents as u32 {
                     // only include nodes with pifs
                     if node.in_indx.is_some() {
                         // exclude opi's
@@ -3869,9 +3416,9 @@ pub fn wfopi_wiring(
         // find the tgt node
         for node in nodes.iter() {
             // make sure in correct box
-            if bf_counter == node.nbox {
+            if bf_counter == node.nbox as u8 {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if (idx) == node.contents {
+                if (idx) == node.contents as u32 {
                     // only opi's
                     if node.n_type == "Opi" {
                         // iterate through port to check for tgt
@@ -3890,30 +3437,29 @@ pub fn wfopi_wiring(
                 src: wfopi_src_tgt[0].clone(),
                 tgt: wfopi_src_tgt[1].clone(),
                 e_type: String::from("Wire"),
-                prop,
+                prop: Some(prop.unwrap() as usize),
             };
             edges.push(e6);
         }
     }
-    edges
 }
 
 pub fn wfopo_wiring(
     eboxf: FunctionNet,
-    nodes: Vec<Node>,
-    mut edges: Vec<Edge>,
+    nodes: &mut Vec<Node>,
+    edges: &mut Vec<Edge>,
     idx: u32,
     bf_counter: u8,
-) -> Vec<Edge> {
+) -> () {
     // iterate through all wires of type
     for wire in eboxf.wfopo.unwrap().iter() {
         let mut wfopo_src_tgt: Vec<String> = vec![];
         // find the src node
         for node in nodes.iter() {
             // make sure in correct box
-            if bf_counter == node.nbox {
+            if bf_counter == node.nbox as u8 {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if (idx) == node.contents {
+                if (idx) == node.contents as u32 {
                     // only opo's
                     if node.n_type == "Opo" {
                         // iterate through port to check for tgt
@@ -3930,9 +3476,9 @@ pub fn wfopo_wiring(
         // finding the tgt node
         for node in nodes.iter() {
             // make sure in correct box
-            if bf_counter == node.nbox {
+            if bf_counter == node.nbox as u8 {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if (idx) == node.contents {
+                if (idx) == node.contents as u32 {
                     // only include nodes with pofs
                     if node.out_idx.is_some() {
                         // exclude opo's
@@ -3959,18 +3505,17 @@ pub fn wfopo_wiring(
             edges.push(e7);
         }
     }
-    edges
 }
 // this is duplicating wires a ton. (01/15/23)
 // shouldn't use bf_counter, should use att_bf_idx since that is variable for functions, need to pull the box from the
 // ports the wires point to.
 pub fn wff_wiring(
     eboxf: FunctionNet,
-    nodes: Vec<Node>,
-    mut edges: Vec<Edge>,
+    nodes: &mut Vec<Node>,
+    edges: &mut Vec<Edge>,
     idx: u32,
     bf_counter: u8,
-) -> Vec<Edge> {
+) -> () {
     // iterate through all wires of type
     for wire in eboxf.wff.unwrap().iter() {
         let mut wff_src_tgt: Vec<String> = vec![];
@@ -3995,11 +3540,11 @@ pub fn wff_wiring(
         // find the src node
         for node in nodes.iter() {
             // make sure in correct box
-            if src_nbox == node.nbox {
+            if src_nbox == node.nbox as u8 {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if src_att == node.contents {
+                if src_att == node.contents as u32 {
                     // matche the box
-                    if (src_box as u32) == node.att_bf_idx {
+                    if (src_box as u32) == node.att_bf_idx as u32 {
                         // only include nodes with pifs
                         if node.in_indx.is_some() {
                             // exclude opo's
@@ -4021,11 +3566,11 @@ pub fn wff_wiring(
         // finding the tgt node
         for node in nodes.iter() {
             // make sure in correct box
-            if tgt_nbox == node.nbox {
+            if tgt_nbox == node.nbox as u8 {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if tgt_att == node.contents {
+                if tgt_att == node.contents as u32 {
                     // match internal box
-                    if (tgt_box as u32) == node.att_bf_idx {
+                    if (tgt_box as u32) == node.att_bf_idx as u32 {
                         // only include nodes with pofs
                         if node.out_idx.is_some() {
                             // exclude opo's
@@ -4048,21 +3593,20 @@ pub fn wff_wiring(
                 src: wff_src_tgt[0].clone(),
                 tgt: wff_src_tgt[1].clone(),
                 e_type: String::from("Wire"),
-                prop,
+                prop: Some(prop.unwrap() as usize),
             };
             edges.push(e8);
         }
     }
-    edges
 }
 
 pub fn wopio_wiring(
     eboxf: FunctionNet,
-    nodes: Vec<Node>,
-    mut edges: Vec<Edge>,
-    idx: u32,
-    bf_counter: u8,
-) -> Vec<Edge> {
+    nodes: &mut Vec<Node>,
+    edges: &mut Vec<Edge>,
+    idx: usize,
+    bf_counter: usize,
+) -> () {
     // iterate through all wires of type
     for wire in eboxf.wopio.unwrap().iter() {
         let mut wopio_src_tgt: Vec<String> = vec![];
@@ -4117,16 +3661,15 @@ pub fn wopio_wiring(
             edges.push(e7);
         }
     }
-    edges
 }
 
 pub fn internal_wiring(
     eboxf: FunctionNet,
-    nodes: Vec<Node>,
-    mut edges: Vec<Edge>,
-    idx: u32,
-    bf_counter: u8,
-) -> Vec<Edge> {
+    nodes: &mut Vec<Node>,
+    edges: &mut Vec<Edge>,
+    idx: usize,
+    bf_counter: usize,
+) -> () {
     // first lets wire the wfopi, note we need to first limit ourselves
     // to only nodes in the current attribute by checking the contents field
     // and then we run find the ports that match the wire src and tgt.
@@ -4137,38 +3680,53 @@ pub fn internal_wiring(
 
     // check if wire exists, wfopi
     if eboxf.wfopi.is_some() {
-        edges = wfopi_wiring(eboxf.clone(), nodes.clone(), edges, idx, bf_counter);
+        wfopi_wiring(
+            eboxf.clone(),
+            &mut nodes.clone(),
+            edges,
+            idx as u32,
+            bf_counter as u8,
+        );
     }
 
     // check if wire exists, wfopo
     if eboxf.wfopo.is_some() {
-        edges = wfopo_wiring(eboxf.clone(), nodes.clone(), edges, idx, bf_counter);
+        wfopo_wiring(
+            eboxf.clone(),
+            &mut nodes.clone(),
+            edges,
+            idx as u32,
+            bf_counter as u8,
+        );
     }
 
     // check if wire exists, wff
     if eboxf.wff.is_some() {
-        edges = wff_wiring(eboxf.clone(), nodes.clone(), edges, idx, bf_counter);
+        wff_wiring(
+            eboxf.clone(),
+            &mut nodes.clone(),
+            edges,
+            idx as u32,
+            bf_counter as u8,
+        );
     }
 
     // check if wire exists, wopio
     if eboxf.wopio.is_some() {
-        edges = wopio_wiring(eboxf, nodes, edges, idx, bf_counter);
+        wopio_wiring(eboxf, nodes, edges, idx, bf_counter);
     }
-
-    edges
 }
 
 // now for the wiring used for imports
 // needs to handle top level and function level wiring that uses the function net at the call of the import.
 pub fn import_wiring(
     gromet: &ModuleCollection,
-    _eboxf: FunctionNet,
-    nodes: Vec<Node>,
-    mut edges: Vec<Edge>,
-    idx: u32,
-    bf_counter: u8,
+    nodes: &mut Vec<Node>,
+    edges: &mut Vec<Edge>,
+    idx: usize,
+    bf_counter: usize,
     parent_node: Node,
-) -> Vec<Edge> {
+) -> () {
     // first based on the parent_node determine if we need to grab the pof's from the top scope or a sub scope
     if parent_node.att_bf_idx == 0 {
         // this means top level wiring
@@ -4200,7 +3758,7 @@ pub fn import_wiring(
                     // make sure only looking in current attribute nodes for srcs and tgts
                     if src_att == node.contents {
                         // matche the box
-                        if (src_box as u32) == node.att_bf_idx {
+                        if (src_box as u32) == node.att_bf_idx as u32 {
                             // only include nodes with pifs
                             if node.in_indx.is_some() {
                                 // exclude opo's
@@ -4225,7 +3783,7 @@ pub fn import_wiring(
                     // make sure only looking in current attribute nodes for srcs and tgts
                     if tgt_att == node.contents {
                         // match internal box
-                        if (tgt_box as u32) == node.att_bf_idx {
+                        if (tgt_box as u32) == node.att_bf_idx as u32 {
                             // only include nodes with pofs
                             if node.out_idx.is_some() {
                                 // exclude opo's
@@ -4253,7 +3811,6 @@ pub fn import_wiring(
                 edges.push(e8);
             }
         }
-        edges
     } else {
         // this means we are in function scope, concerns on if this is cross attributal or just internal wiring...
         let eboxf = gromet.modules[0].attributes[(parent_node.contents - 1) as usize].clone();
@@ -4285,7 +3842,7 @@ pub fn import_wiring(
                     // make sure only looking in current attribute nodes for srcs and tgts
                     if src_att == node.contents {
                         // matche the box
-                        if (src_box as u32) == node.att_bf_idx {
+                        if (src_box as u32) == node.att_bf_idx as u32 {
                             // only include nodes with pifs
                             if node.in_indx.is_some() {
                                 // exclude opo's
@@ -4310,7 +3867,7 @@ pub fn import_wiring(
                     // make sure only looking in current attribute nodes for srcs and tgts
                     if tgt_att == node.contents {
                         // match internal box
-                        if (tgt_box as u32) == node.att_bf_idx {
+                        if (tgt_box as u32) == node.att_bf_idx as u32 {
                             // only include nodes with pofs
                             if node.out_idx.is_some() {
                                 // exclude opo's
@@ -4338,17 +3895,16 @@ pub fn import_wiring(
                 edges.push(e8);
             }
         }
-        edges
     }
 }
 
 pub fn cross_att_wiring(
     eboxf: FunctionNet, // This is the current attribute
-    mut nodes: Vec<Node>,
-    mut edges: Vec<Edge>,
-    idx: u32,       // this +1 is the current attribute index
-    bf_counter: u8, // this is the current box
-) -> (Vec<Edge>, Vec<Node>) {
+    nodes: &mut Vec<Node>,
+    edges: &mut Vec<Edge>,
+    idx: usize,        // this +1 is the current attribute index
+    bf_counter: usize, // this is the current box
+) -> () {
     // wire id corresponds to subport index in list so ex: wff src.id="1" means the first opi in the list of the src.box, this is the in_idx in the opi or out_indx in opo.
     // This will have to run wfopo wfopi and wff all in order to get the cross attribual wiring that can exist in all of them, refactoring won't do much in code saving space though.
     // for cross attributal wiring they will construct the following types of wires
@@ -4358,29 +3914,46 @@ pub fn cross_att_wiring(
 
     // check if wire exists, wfopi
     if eboxf.wfopi.is_some() {
-        edges = wfopi_cross_att_wiring(eboxf.clone(), nodes.clone(), edges, idx, bf_counter);
+        wfopi_cross_att_wiring(
+            eboxf.clone(),
+            &mut nodes.clone(),
+            edges,
+            idx as u32,
+            bf_counter as u8,
+        );
     }
 
     // check if wire exists, wfopo
     if eboxf.wfopo.is_some() {
-        edges = wfopo_cross_att_wiring(eboxf.clone(), nodes.clone(), edges, idx, bf_counter);
+        wfopo_cross_att_wiring(
+            eboxf.clone(),
+            &mut nodes.clone(),
+            edges,
+            idx as u32,
+            bf_counter as u8,
+        );
     }
 
     // check if wire exists, wff
     if eboxf.wff.is_some() {
-        (edges, nodes) = wff_cross_att_wiring(eboxf, nodes.clone(), edges, idx, bf_counter);
+        wff_cross_att_wiring(
+            eboxf,
+            &mut nodes.clone(),
+            edges,
+            idx as u32,
+            bf_counter as u8,
+        );
     }
-    (edges, nodes)
 }
 // this will construct connections from the function opi's to the sub module opi's, tracing inputs through the function
 // opi(sub)->opi(fun)
 pub fn wfopi_cross_att_wiring(
     eboxf: FunctionNet, // This is the current attribute
-    nodes: Vec<Node>,
-    mut edges: Vec<Edge>,
+    nodes: &mut Vec<Node>,
+    edges: &mut Vec<Edge>,
     idx: u32,       // this is the current attribute index
     bf_counter: u8, // this is the current box
-) -> Vec<Edge> {
+) -> () {
     for wire in eboxf.wfopi.as_ref().unwrap().iter() {
         // collect info to identify the opi src node
         let src_idx = wire.src; // port index
@@ -4408,11 +3981,11 @@ pub fn wfopi_cross_att_wiring(
             // find the src node
             for node in nodes.iter() {
                 // make sure in correct box
-                if src_nbox == node.nbox {
+                if src_nbox == node.nbox as u8 {
                     // make sure only looking in current attribute nodes for srcs and tgts
-                    if src_att == node.contents {
+                    if src_att == node.contents as u32 {
                         // make sure box index also lines up
-                        if (src_box as u32) == node.att_bf_idx {
+                        if (src_box as u32) == node.att_bf_idx as u32 {
                             // only opo's
                             if node.n_type == "Opi" {
                                 // iterate through port to check for tgt
@@ -4429,9 +4002,9 @@ pub fn wfopi_cross_att_wiring(
             }
             for node in nodes.iter() {
                 // make sure in correct box
-                if tgt_nbox == node.nbox {
+                if tgt_nbox == node.nbox as u8 {
                     // make sure only looking in current attribute nodes for srcs and tgts
-                    if tgt_att == node.contents {
+                    if tgt_att == node.contents as u32 {
                         // only opo's
                         if node.n_type == "Opi" {
                             // iterate through port to check for tgt
@@ -4456,17 +4029,16 @@ pub fn wfopi_cross_att_wiring(
             }
         }
     }
-    edges
 }
 // this will construct connections from the function opo's to the sub module opo's, tracing outputs through the function
 // opo(fun)->opo(sub)
 pub fn wfopo_cross_att_wiring(
     eboxf: FunctionNet, // This is the current attribute
-    nodes: Vec<Node>,
-    mut edges: Vec<Edge>,
+    nodes: &mut Vec<Node>,
+    edges: &mut Vec<Edge>,
     idx: u32,       // this +1 is the current attribute index
     bf_counter: u8, // this is the current box
-) -> Vec<Edge> {
+) -> () {
     for wire in eboxf.wfopo.as_ref().unwrap().iter() {
         // collect info to identify the opo tgt node
         let tgt_idx = wire.tgt; // port index
@@ -4494,9 +4066,9 @@ pub fn wfopo_cross_att_wiring(
             // find the src node
             for node in nodes.iter() {
                 // make sure in correct box
-                if src_nbox == node.nbox {
+                if src_nbox == node.nbox as u8 {
                     // make sure only looking in current attribute nodes for srcs and tgts
-                    if src_att == node.contents {
+                    if src_att == node.contents as u32 {
                         // only opo's
                         if node.n_type == "Opo" {
                             // iterate through port to check for tgt
@@ -4512,9 +4084,10 @@ pub fn wfopo_cross_att_wiring(
             }
             for node in nodes.iter() {
                 // make sure in correct box
-                if tgt_nbox == node.nbox {
+                if tgt_nbox == node.nbox as u8 {
                     // make sure only looking in current attribute nodes for srcs and tgts
-                    if tgt_att == node.contents && (tgt_box as u32) == node.att_bf_idx {
+                    if tgt_att == node.contents as u32 && (tgt_box as u32) == node.att_bf_idx as u32
+                    {
                         // only opo's
                         if node.n_type == "Opo" {
                             // iterate through port to check for tgt
@@ -4539,18 +4112,17 @@ pub fn wfopo_cross_att_wiring(
             }
         }
     }
-    edges
 }
 // this will construct connections from the sub function modules opi's to another sub module opo's, tracing data inside the function
 // opi(sub)->opo(sub)
 #[allow(unused_assignments)]
 pub fn wff_cross_att_wiring(
     eboxf: FunctionNet, // This is the current attribute, should be the function if in a function
-    mut nodes: Vec<Node>,
-    mut edges: Vec<Edge>,
+    nodes: &mut Vec<Node>,
+    edges: &mut Vec<Edge>,
     idx: u32,       // this +1 is the current attribute index
     bf_counter: u8, // this is the current box
-) -> (Vec<Edge>, Vec<Node>) {
+) -> () {
     for wire in eboxf.wff.as_ref().unwrap().iter() {
         // collect info to identify the opi src node
         let src_idx = wire.src; // port index
@@ -4595,9 +4167,11 @@ pub fn wff_cross_att_wiring(
                 // internal wiring module as well?
                 for node in nodes.iter() {
                     // make sure in correct box
-                    if src_nbox == node.nbox {
+                    if src_nbox == node.nbox as u8 {
                         // make sure only looking in current attribute nodes for srcs and tgts
-                        if src_att == node.contents && (src_box as u32) == node.att_bf_idx {
+                        if src_att == node.contents as u32
+                            && (src_box as u32) == node.att_bf_idx as u32
+                        {
                             // only opo's
                             if node.n_type == "Opi" {
                                 // iterate through port to check for tgt
@@ -4613,9 +4187,11 @@ pub fn wff_cross_att_wiring(
                 }
                 for node in nodes.iter() {
                     // make sure in correct box
-                    if tgt_nbox == node.nbox {
+                    if tgt_nbox == node.nbox as u8 {
                         // make sure only looking in current attribute nodes for srcs and tgts
-                        if tgt_att == node.contents && (tgt_box as u32) == node.att_bf_idx {
+                        if tgt_att == node.contents as u32
+                            && (tgt_box as u32) == node.att_bf_idx as u32
+                        {
                             // only opo's
                             if node.n_type == "Opo" {
                                 // iterate through port to check for tgt
@@ -4645,9 +4221,11 @@ pub fn wff_cross_att_wiring(
                 // find the src node
                 for node in nodes.iter() {
                     // make sure in correct box
-                    if src_nbox == node.nbox {
+                    if src_nbox == node.nbox as u8 {
                         // make sure only looking in current attribute nodes for srcs and tgts
-                        if src_att == node.contents && (src_box as u32) == node.att_bf_idx {
+                        if src_att == node.contents as u32
+                            && (src_box as u32) == node.att_bf_idx as u32
+                        {
                             // only opo's
                             if node.n_type == "Opi" {
                                 // iterate through port to check for tgt
@@ -4663,9 +4241,11 @@ pub fn wff_cross_att_wiring(
                 }
                 for node in nodes.iter() {
                     // make sure in correct box
-                    if tgt_nbox == node.nbox {
+                    if tgt_nbox == node.nbox as u8 {
                         // make sure only looking in current attribute nodes for srcs and tgts
-                        if tgt_att == node.contents && (tgt_box as u32) == node.att_bf_idx {
+                        if tgt_att == node.contents as u32
+                            && (tgt_box as u32) == node.att_bf_idx as u32
+                        {
                             // only opo's
                             if node.n_type == "Primitive" || node.n_type == "Literal" {
                                 // iterate through port to check for tgt
@@ -4727,9 +4307,11 @@ pub fn wff_cross_att_wiring(
                 // find the src node
                 for node in nodes.iter() {
                     // make sure in correct box
-                    if src_nbox == node.nbox {
+                    if src_nbox == node.nbox as u8 {
                         // make sure only looking in current attribute nodes for srcs and tgts
-                        if src_att == node.contents && (src_box as u32) == node.att_bf_idx {
+                        if src_att == node.contents as u32
+                            && (src_box as u32) == node.att_bf_idx as u32
+                        {
                             // only opo's
                             if node.n_type == "Primitive" {
                                 // iterate through port to check for tgt
@@ -4745,9 +4327,11 @@ pub fn wff_cross_att_wiring(
                 }
                 for node in nodes.iter() {
                     // make sure in correct box
-                    if tgt_nbox == node.nbox {
+                    if tgt_nbox == node.nbox as u8 {
                         // make sure only looking in current attribute nodes for srcs and tgts
-                        if tgt_att == node.contents && (tgt_box as u32) == node.att_bf_idx {
+                        if tgt_att == node.contents as u32
+                            && (tgt_box as u32) == node.att_bf_idx as u32
+                        {
                             // only opo's
                             if node.n_type == "Opo" {
                                 // iterate through port to check for tgt
@@ -4773,14 +4357,13 @@ pub fn wff_cross_att_wiring(
             }
         }
     }
-    (edges, nodes)
 }
 // external wiring is the wiring between boxes at the module level
 pub fn external_wiring(
     gromet: &ModuleCollection,
-    nodes: Vec<Node>,
-    mut edges: Vec<Edge>,
-) -> Vec<Edge> {
+    nodes: &mut Vec<Node>,
+    edges: &mut Vec<Edge>,
+) -> () {
     if gromet.modules[0].r#fn.wff.as_ref().is_some() {
         for wire in gromet.modules[0].r#fn.wff.as_ref().unwrap().iter() {
             let src_idx = wire.src; // pif wire connects to
@@ -4827,8 +4410,8 @@ pub fn external_wiring(
             // This is double counting since only check is name and box, check on attributes?
             // find the src
             for node in nodes.iter() {
-                if node.nbox == src_box
-                    && src_att == node.contents
+                if node.nbox == src_box as usize
+                    && src_att == node.contents as u32
                     && (node.n_type == "Opi" || node.n_type == "Import")
                 {
                     for p in node.in_indx.as_ref().unwrap().iter() {
@@ -4844,8 +4427,8 @@ pub fn external_wiring(
             for node in nodes.iter() {
                 // check this field
                 if node.n_type == "Opo" {
-                    if node.nbox == tgt_box {
-                        if tgt_att == node.contents {
+                    if node.nbox == tgt_box as usize {
+                        if tgt_att == node.contents as u32 {
                             for p in node.out_idx.as_ref().unwrap().iter() {
                                 // push the tgt
                                 if (tgt_id as u32) == *p {
@@ -4855,7 +4438,7 @@ pub fn external_wiring(
                             }
                         } else {
                             // perform extra check to make sure not getting sub cross attributal nodes
-                            if tgt_att == node.contents {
+                            if tgt_att == node.contents as u32 {
                                 for p in node.out_idx.as_ref().unwrap().iter() {
                                     // push the tgt
                                     if (tgt_id as u32) == *p {
@@ -4868,7 +4451,7 @@ pub fn external_wiring(
                     }
                 } else if (node.n_type == "Literal" || node.n_type == "Import")
                     && !tgt_found
-                    && node.nbox == tgt_box
+                    && node.nbox == tgt_box as usize
                 {
                     for p in node.out_idx.as_ref().unwrap().iter() {
                         // push the tgt
@@ -4878,6 +4461,8 @@ pub fn external_wiring(
                     }
                 }
             }
+            println!("{:?}", wff_src_tgt.clone());
+            println!("{:?}", tgt_found);
             if wff_src_tgt.len() == 2 {
                 let e9 = Edge {
                     src: wff_src_tgt[0].clone(),
@@ -4889,7 +4474,6 @@ pub fn external_wiring(
             }
         }
     }
-    edges
 }
 
 pub fn parse_gromet_queries(gromet: ModuleCollection) -> Vec<String> {
