+++ skema/skema-rs/skema/src/database_old.rs
@@ -55,16 +55,15 @@ pub struct MetadataNode {
 
 #[derive(Debug, Clone)]
 pub struct Node {
-    pub n_type: String, // name of type of node, function or expression
+    pub n_type: String,
     pub value: Option<String>,
     pub name: Option<String>,
     pub node_id: String,
     pub out_idx: Option<Vec<u32>>, // opo or pof index, directly matching the wire src/tgt notation
     pub in_indx: Option<Vec<u32>>, // opi or pif index
-    pub contents: usize, // This indexes which index this node has inside the fn_array list, but for functions this is overloaded
-    pub nbox: usize, // this indexs the top level box call for the node, if any, default for none is 0
-    pub att_bf_idx: usize, // this indexes the attribute index of one higher scope if any, default for none is 0
-    pub box_counter: usize, // this indexes the box call for the node one scope up, matches nbox if higher scope is top level
+    pub contents: u32, // This indexes which index this node has inside the attribute list, but for functions this is overloaded
+    pub nbox: u8,
+    pub att_bf_idx: u32,
 }
 
 #[derive(Debug, Clone)]
@@ -72,18 +71,7 @@ pub struct Edge {
     pub src: String,
     pub tgt: String,
     pub e_type: String,
-    pub prop: Option<usize>, // option because of opo's and opi's
-}
-
-#[derive(Debug, Clone)]
-pub struct ConstructorArgs {
-    pub att_box: FunctionNet, // current attribute object is in
-    pub cur_box: GrometBox,   // current box object is in
-    pub parent_node: Node,    // parent node of the current constructor
-    pub att_idx: usize,       // This will index the attribute the function is in, 0 if not
-    pub bf_counter: usize, // This indexes which top level box the function is under, inherited from parent if not explicit, 0 if not
-    pub att_bf_idx: usize, // This indexes if the function is a subscope of a larger function, 0 if not
-    pub box_counter: usize, // this is the index of the box if called inside another function, 0 if not
+    pub prop: Option<u32>, // option because of opo's and opi's
 }
 
 pub fn execute_query(query: &str, host: &str) -> Result<(), MgError> {
@@ -402,9 +390,9 @@ fn create_function_net_lib(gromet: &ModuleCollection, mut start: u32) -> Vec<Str
     for node in nodes.clone() {
         if (1 == node.contents) && (node.n_type == "Function") {
             function_call_repeat = true;
-            if node.nbox < original_bf as usize {
-                original_bf = node.nbox as u8; // This grabs the first instance of bf that the function was called
-                                               // and thus is the nbox value of the nodes of the original contents
+            if node.nbox < original_bf {
+                original_bf = node.nbox; // This grabs the first instance of bf that the function was called
+                                         // and thus is the nbox value of the nodes of the original contents
             }
         }
     }
@@ -426,9 +414,8 @@ fn create_function_net_lib(gromet: &ModuleCollection, mut start: u32) -> Vec<Str
             out_idx: None,
             in_indx: None,
             contents: 1,
-            nbox: bf_counter as usize,
+            nbox: bf_counter,
             att_bf_idx: 0,
-            box_counter: 0,
         };
         let e1 = Edge {
             src: String::from("mod"),
@@ -462,37 +449,28 @@ fn create_function_net_lib(gromet: &ModuleCollection, mut start: u32) -> Vec<Str
         start += 1;
         let idx = 0;
         let eboxf = gromet.modules[0].attributes[idx as usize].clone();
-
-        let mut c_args = ConstructorArgs {
-            att_box: eboxf.clone(),
-            cur_box: eboxf.bf.unwrap()[0].clone(),
-            parent_node: n1.clone(),
-            att_idx: 1,
-            bf_counter: 1,
-            att_bf_idx: 0,
-            box_counter: 0,
-        };
-
         // create the ports
-        create_opo(
-            &gromet,    // gromet for metadata
-            &mut nodes, // nodes
-            &mut edges,
-            &mut meta_nodes,
+        (nodes, edges, meta_nodes, start) = create_opo(
+            nodes.clone(),
+            edges.clone(),
+            meta_nodes.clone(),
+            &gromet.clone(),
             start,
-            c_args.clone(),
+            n1.clone(),
+            bf_counter,
         );
-        create_opi(
-            &gromet,    // gromet for metadata
-            &mut nodes, // nodes
-            &mut edges,
-            &mut meta_nodes,
+        (nodes, edges, meta_nodes, start) = create_opi(
+            nodes.clone(),
+            edges.clone(),
+            meta_nodes.clone(),
+            &gromet.clone(),
             start,
-            c_args.clone(),
+            n1.clone(),
+            bf_counter,
         );
         // now to add the contains wires for the additional function call onto the original contents nodes:
         for node in nodes.clone() {
-            if (node.nbox == original_bf as usize)
+            if (node.nbox == original_bf)
                 && (node.contents == (idx + 1))
                 && ((node.n_type == "Literal")
                     || (node.n_type == "Primitive")
@@ -518,7 +496,7 @@ fn create_function_net_lib(gromet: &ModuleCollection, mut start: u32) -> Vec<Str
             // find the src node
             for node in nodes.iter() {
                 // make sure in correct box
-                if original_bf == node.nbox as u8 {
+                if original_bf == node.nbox {
                     // make sure only looking in current attribute nodes for srcs and tgts
                     if (idx + 1) == node.contents {
                         // only include nodes with pifs
@@ -540,7 +518,7 @@ fn create_function_net_lib(gromet: &ModuleCollection, mut start: u32) -> Vec<Str
             // find the tgt node
             for node in nodes.iter() {
                 // make sure in correct box
-                if bf_counter == node.nbox as u8 {
+                if bf_counter == node.nbox {
                     // make sure only looking in current attribute nodes for srcs and tgts
                     if (idx + 1) == node.contents {
                         // only opi's
@@ -571,7 +549,7 @@ fn create_function_net_lib(gromet: &ModuleCollection, mut start: u32) -> Vec<Str
             // find the src node
             for node in nodes.iter() {
                 // make sure in correct box
-                if bf_counter == node.nbox as u8 {
+                if bf_counter == node.nbox {
                     // make sure only looking in current attribute nodes for srcs and tgts
                     if (idx + 1) == node.contents {
                         // only opo's
@@ -590,7 +568,7 @@ fn create_function_net_lib(gromet: &ModuleCollection, mut start: u32) -> Vec<Str
             // finding the tgt node
             for node in nodes.iter() {
                 // make sure in correct box
-                if original_bf == node.nbox as u8 {
+                if original_bf == node.nbox {
                     // make sure only looking in current attribute nodes for srcs and tgts
                     if (idx + 1) == node.contents {
                         // only include nodes with pofs
@@ -623,46 +601,178 @@ fn create_function_net_lib(gromet: &ModuleCollection, mut start: u32) -> Vec<Str
         /////////////////////////
         // This is for no function repeat in a library representation
         /////////////////////////
-        let att_box = gromet.modules[0].attributes[0_usize].clone();
-
-        let mut temp_node = Node {
-            n_type: String::from("Temp"),
+        let eboxf = gromet.modules[0].attributes[0_usize].clone();
+        let n1 = Node {
+            n_type: String::from("Function"),
             value: None,
-            name: None,
-            node_id: "mod".to_string(),
+            name: Some(
+                eboxf.b.as_ref().unwrap()[0]
+                    .name
+                    .clone()
+                    .map_or_else(|| format!("Function{}", start), |x| x),
+            ),
+            node_id: format!("n{}", start),
             out_idx: None,
             in_indx: None,
-            contents: 0,
-            nbox: 0,
+            contents: 1,
+            nbox: bf_counter,
             att_bf_idx: 0,
-            box_counter: 0,
         };
-
-        // this represents the default state for this struct
-        let mut c_args = ConstructorArgs {
-            att_box: att_box.clone(),
-            cur_box: att_box.bf.unwrap()[0].clone(),
-            parent_node: temp_node,
-            att_idx: 1,
-            bf_counter: 0,
-            att_bf_idx: 0,
-            box_counter: 0,
+        let e1 = Edge {
+            src: String::from("mod"),
+            tgt: format!("n{}", start),
+            e_type: String::from("Contains"),
+            prop: Some(1),
         };
+        nodes.push(n1.clone());
+        edges.push(e1);
 
-        create_function(
-            &gromet,    // gromet for metadata
-            &mut nodes, // nodes
-            &mut edges,
-            &mut meta_nodes,
+        if eboxf.b.as_ref().unwrap()[0].metadata.as_ref().is_some() {
+            metadata_idx = eboxf.b.unwrap()[0].metadata.unwrap();
+            let mut repeat_meta = false;
+            for node in meta_nodes.iter() {
+                if node.metadata_idx == metadata_idx {
+                    repeat_meta = true;
+                }
+            }
+            if !repeat_meta {
+                meta_nodes.append(&mut create_metadata_node(&gromet.clone(), metadata_idx));
+                // adding the metadata edge
+                let me1 = Edge {
+                    src: n1.node_id.clone(),
+                    tgt: format!("m{}", metadata_idx),
+                    e_type: String::from("Metadata"),
+                    prop: None,
+                };
+                edges.push(me1);
+            }
+        }
+        // now travel to contents index of the attribute list (note it is 1 index,
+        // so contents=1 => attribute[0])
+        // create nodes and edges for this entry, include opo's and opi's
+        start += 1;
+        let idx = 1;
+        let eboxf = gromet.modules[0].attributes[(idx - 1) as usize].clone();
+        // create the ports
+        (nodes, edges, meta_nodes, start) = create_opo(
+            nodes.clone(),
+            edges.clone(),
+            meta_nodes.clone(),
+            &gromet.clone(),
             start,
-            c_args.clone(),
+            n1.clone(),
+            bf_counter,
         );
+        (nodes, edges, meta_nodes, start) = create_opi(
+            nodes.clone(),
+            edges.clone(),
+            meta_nodes.clone(),
+            &gromet.clone(),
+            start,
+            n1.clone(),
+            bf_counter,
+        );
+        // now to construct the nodes inside the function, currently supported Literals and Primitives
+        // first include an Expression for increased depth
+        let mut box_counter: u8 = 1;
+        if eboxf.bf.is_some() {
+            for sboxf in eboxf.bf.as_ref().unwrap().iter() {
+                match sboxf.function_type {
+                    FunctionType::Predicate => {
+                        (nodes, edges, start, meta_nodes) = create_att_predicate(
+                            &gromet.clone(),
+                            eboxf.clone(),
+                            sboxf.clone(),
+                            nodes.clone(),
+                            edges.clone(),
+                            n1.clone(),
+                            idx,
+                            box_counter,
+                            bf_counter,
+                            start,
+                            meta_nodes.clone(),
+                        );
+                    }
+                    FunctionType::Expression => {
+                        (nodes, edges, start, meta_nodes) = create_att_expression(
+                            &gromet.clone(),
+                            eboxf.clone(),
+                            sboxf.clone(),
+                            nodes.clone(),
+                            edges.clone(),
+                            n1.clone(),
+                            idx,
+                            box_counter,
+                            bf_counter,
+                            start,
+                            meta_nodes.clone(),
+                        );
+                    }
+                    FunctionType::Literal => {
+                        (nodes, edges, meta_nodes) = create_att_literal(
+                            &gromet.clone(),
+                            eboxf.clone(),
+                            sboxf.clone(),
+                            nodes.clone(),
+                            edges.clone(),
+                            n1.clone(),
+                            idx,
+                            box_counter,
+                            bf_counter,
+                            start,
+                            meta_nodes.clone(),
+                        );
+                    }
+                    FunctionType::Primitive => {
+                        (nodes, edges, meta_nodes) = create_att_primitive(
+                            &gromet.clone(),
+                            eboxf.clone(),
+                            sboxf.clone(),
+                            nodes.clone(),
+                            edges.clone(),
+                            n1.clone(),
+                            idx,
+                            box_counter,
+                            bf_counter,
+                            start,
+                            meta_nodes.clone(),
+                        );
+                    }
+                    FunctionType::Abstract => {
+                        (nodes, edges, meta_nodes) = create_att_primitive(
+                            &gromet.clone(),
+                            eboxf.clone(),
+                            sboxf.clone(),
+                            nodes.clone(),
+                            edges.clone(),
+                            n1.clone(),
+                            idx,
+                            box_counter,
+                            bf_counter,
+                            start,
+                            meta_nodes.clone(),
+                        );
+                    }
+                    _ => {}
+                }
+                box_counter += 1;
+                start += 1;
+            }
+        }
+
+        // Now we perform the internal wiring of this branch
+        edges = internal_wiring(eboxf.clone(), nodes.clone(), edges, idx, bf_counter);
+        // perform cross attributal wiring of function
+        (edges, nodes) = cross_att_wiring(eboxf, nodes.clone(), edges, idx, bf_counter);
     }
+    start += 1;
+    bf_counter += 1;
+
     // add wires for inbetween attribute level boxes, so opo's, opi's and module level literals
     // between attributes
     // get wired through module level wff field, will require reading through node list to
     // match contents field to box field on wff entries
-    external_wiring(gromet, &mut nodes.clone(), &mut edges);
+    edges = external_wiring(gromet, nodes.clone(), edges);
 
     // make conditionals if they exist
     if gromet.modules[0].r#fn.bc.as_ref().is_some() {
@@ -677,27 +787,20 @@ fn create_function_net_lib(gromet: &ModuleCollection, mut start: u32) -> Vec<Str
             contents: 0,
             nbox: 0,
             att_bf_idx: 0,
-            box_counter: 0,
-        };
-        let mut c_args = ConstructorArgs {
-            att_box: gromet.modules[0].r#fn.clone(),
-            cur_box: gromet.modules[0].attributes[0].clone().bf.unwrap()[0].clone(), // this is a temp fill
-            parent_node: temp_mod_node,
-            att_idx: 1,
-            bf_counter: 0,
-            att_bf_idx: 0,
-            box_counter: 0,
         };
         for _cond in gromet.modules[0].r#fn.bc.as_ref().unwrap().iter() {
             // now lets check for and setup any conditionals at this level
-            create_conditional(
-                &gromet,    // gromet for metadata
-                &mut nodes, // nodes
-                &mut edges,
-                &mut meta_nodes,
+            (nodes, edges, start, meta_nodes) = create_conditional(
+                &gromet.clone(),
+                gromet.modules[0].r#fn.clone(), // This is gromet but is more generalizable based on scope
+                nodes.clone(),
+                edges.clone(),
+                temp_mod_node.clone(),
+                0,            // because top level
+                cond_counter, // This indexes the conditional in the list of conditionals (bc)
+                0,            // because top level
                 start,
-                c_args.clone(),
-                cond_counter,
+                meta_nodes.clone(),
             );
             cond_counter += 1;
         }
@@ -715,27 +818,20 @@ fn create_function_net_lib(gromet: &ModuleCollection, mut start: u32) -> Vec<Str
             contents: 0,
             nbox: 0,
             att_bf_idx: 0,
-            box_counter: 0,
-        };
-        let mut c_args = ConstructorArgs {
-            att_box: gromet.modules[0].r#fn.clone(),
-            cur_box: gromet.modules[0].attributes[0].clone().bf.unwrap()[0].clone(),
-            parent_node: temp_mod_node,
-            att_idx: 1,
-            bf_counter: 0,
-            att_bf_idx: 0,
-            box_counter: 0,
         };
         for _while_l in gromet.modules[0].r#fn.bl.as_ref().unwrap().iter() {
             // now lets check for and setup any conditionals at this level
-            create_while_loop(
-                &gromet,    // gromet for metadata
-                &mut nodes, // nodes
-                &mut edges,
-                &mut meta_nodes,
+            (nodes, edges, start, meta_nodes) = create_while_loop(
+                &gromet.clone(),
+                gromet.modules[0].r#fn.clone(), // This is gromet but is more generalizable based on scope
+                nodes.clone(),
+                edges.clone(),
+                temp_mod_node.clone(),
+                0,             // because top level
+                while_counter, // This indexes the conditional in the list of conditionals (bc)
+                0,             // because top level
                 start,
-                c_args.clone(),
-                while_counter,
+                meta_nodes.clone(),
             );
             while_counter += 1;
         }
@@ -790,82 +886,283 @@ fn create_function_net(gromet: &ModuleCollection, mut start: u32) -> Vec<String>
     let mut metadata_idx = 0;
     let mut edges: Vec<Edge> = vec![];
 
-    // temp node for calling the module as parent for top level
-    let mut temp_node = Node {
-        n_type: String::from("Temp"),
-        value: None,
-        name: None,
-        node_id: "mod".to_string(),
-        out_idx: None,
-        in_indx: None,
-        contents: 0,
-        nbox: 0,
-        att_bf_idx: 0,
-        box_counter: 0,
-    };
-
-    // this represents the default state for this struct
-    let mut c_args = ConstructorArgs {
-        att_box: gromet.modules[0].r#fn.clone(),
-        cur_box: gromet.modules[0].r#fn.bf.as_ref().unwrap()[0].clone(),
-        parent_node: temp_node.clone(),
-        att_idx: 0,
-        bf_counter: 0,
-        att_bf_idx: 0,
-        box_counter: 0,
-    };
-
-    let mut bf_counter: usize = 1;
+    let mut bf_counter: u8 = 1;
     for boxf in gromet.modules[0].r#fn.bf.as_ref().unwrap().iter() {
-        c_args.bf_counter = bf_counter;
-        c_args.cur_box = boxf.clone();
-        c_args.att_idx = 0; // reset incase we hit a function and then literal for example
-                            // construct the sub module level boxes along with their metadata and connection to module
-        c_args.parent_node = temp_node.clone(); // for if overwritten in function constructor
+        // construct the sub module level boxes along with their metadata and connection to module
         match boxf.function_type {
             FunctionType::Literal => {
-                create_att_literal(
-                    &gromet,    // gromet for metadata
-                    &mut nodes, // nodes
-                    &mut edges,
-                    &mut meta_nodes,
-                    start,
-                    c_args.clone(),
-                );
+                // first we find the pof value for the literal box
+                let mut pof: Vec<u32> = vec![];
+                if gromet.modules[0].r#fn.pof.clone().is_some() {
+                    let mut po_idx: u32 = 1;
+                    for port in gromet.modules[0].r#fn.pof.clone().unwrap().iter() {
+                        if port.r#box == bf_counter {
+                            pof.push(po_idx);
+                        }
+                        po_idx += 1;
+                    }
+                }
+                let n1 = Node {
+                    n_type: String::from("Literal"),
+                    value: Some(format!("{:?}", boxf.value.clone().as_ref().unwrap())),
+                    name: Some("Literal".to_string()),
+                    node_id: format!("n{}", start),
+                    out_idx: Some(pof),
+                    in_indx: None,
+                    contents: 0,
+                    nbox: bf_counter,
+                    att_bf_idx: 0,
+                };
+                let e1 = Edge {
+                    src: String::from("mod"),
+                    tgt: format!("n{}", start),
+                    e_type: String::from("Contains"),
+                    prop: boxf.contents,
+                };
+                nodes.push(n1.clone());
+                edges.push(e1);
+                if boxf.metadata.as_ref().is_some() {
+                    metadata_idx = boxf.metadata.unwrap();
+                    let mut repeat_meta = false;
+                    for node in meta_nodes.iter() {
+                        if node.metadata_idx == metadata_idx {
+                            repeat_meta = true;
+                        }
+                    }
+                    if !repeat_meta {
+                        meta_nodes.append(&mut create_metadata_node(&gromet.clone(), metadata_idx));
+                        let me1 = Edge {
+                            src: n1.node_id.clone(),
+                            tgt: format!("m{}", metadata_idx),
+                            e_type: String::from("Metadata"),
+                            prop: None,
+                        };
+                        edges.push(me1);
+                    }
+                }
             }
             FunctionType::Predicate => {
-                c_args.att_idx = boxf.contents.unwrap() as usize;
-                create_att_predicate(
-                    &gromet,    // gromet for metadata
-                    &mut nodes, // nodes
-                    &mut edges,
-                    &mut meta_nodes,
+                let n1 = Node {
+                    n_type: String::from("Predicate"),
+                    value: None,
+                    name: Some(format!("Predicate{}", start)),
+                    node_id: format!("n{}", start),
+                    out_idx: None,
+                    in_indx: None,
+                    contents: boxf.contents.unwrap(),
+                    nbox: bf_counter,
+                    att_bf_idx: 0,
+                };
+                let e1 = Edge {
+                    src: String::from("mod"),
+                    tgt: format!("n{}", start),
+                    e_type: String::from("Contains"),
+                    prop: boxf.contents,
+                };
+                nodes.push(n1.clone());
+                edges.push(e1);
+
+                if boxf.metadata.as_ref().is_some() {
+                    metadata_idx = boxf.metadata.unwrap();
+                    let mut repeat_meta = false;
+                    for node in meta_nodes.iter() {
+                        if node.metadata_idx == metadata_idx {
+                            repeat_meta = true;
+                        }
+                    }
+                    if !repeat_meta {
+                        meta_nodes.append(&mut create_metadata_node(&gromet.clone(), metadata_idx));
+                        let me1 = Edge {
+                            src: n1.node_id.clone(),
+                            tgt: format!("m{}", metadata_idx),
+                            e_type: String::from("Metadata"),
+                            prop: None,
+                        };
+                        edges.push(me1);
+                    }
+                }
+
+                // now travel to contents index of the attribute list (note it is 1 index,
+                // so contents=1 => attribute[0])
+                // create nodes and edges for this entry, include opo's and opi's
+                start += 1;
+                let idx = boxf.contents.unwrap() - 1;
+                let eboxf = gromet.modules[0].attributes[idx as usize].clone();
+                // create the ports
+                (nodes, edges, meta_nodes, start) = create_opo(
+                    nodes.clone(),
+                    edges.clone(),
+                    meta_nodes.clone(),
+                    &gromet.clone(),
                     start,
-                    c_args.clone(),
+                    n1.clone(),
+                    bf_counter,
                 );
+                (nodes, edges, meta_nodes, start) = create_opi(
+                    nodes.clone(),
+                    edges.clone(),
+                    meta_nodes.clone(),
+                    &gromet.clone(),
+                    start,
+                    n1.clone(),
+                    bf_counter,
+                );
+                // now to construct the nodes inside the expression, Literal and Primitives
+                let mut box_counter: u8 = 1;
+                for sboxf in eboxf.bf.clone().as_ref().unwrap().iter() {
+                    match sboxf.function_type {
+                        FunctionType::Literal => {
+                            (nodes, edges, meta_nodes) = create_att_literal(
+                                &gromet.clone(),
+                                eboxf.clone(),
+                                sboxf.clone(),
+                                nodes.clone(),
+                                edges.clone(),
+                                n1.clone(),
+                                idx,
+                                box_counter,
+                                bf_counter,
+                                start,
+                                meta_nodes.clone(),
+                            );
+                        }
+                        FunctionType::Primitive => {
+                            (nodes, edges, meta_nodes) = create_att_primitive(
+                                &gromet.clone(),
+                                eboxf.clone(),
+                                sboxf.clone(),
+                                nodes.clone(),
+                                edges.clone(),
+                                n1.clone(),
+                                idx,
+                                box_counter,
+                                bf_counter,
+                                start,
+                                meta_nodes.clone(),
+                            );
+                        }
+                        _ => {}
+                    }
+                    box_counter += 1;
+                    start += 1;
+                }
+                // Now we perform the internal wiring of this branch
+                edges = internal_wiring(eboxf.clone(), nodes.clone(), edges, idx, bf_counter);
             }
             FunctionType::Expression => {
-                c_args.att_idx = boxf.contents.unwrap() as usize;
-                create_att_expression(
-                    &gromet,    // gromet for metadata
-                    &mut nodes, // nodes
-                    &mut edges,
-                    &mut meta_nodes,
+                let n1 = Node {
+                    n_type: String::from("Expression"),
+                    value: None,
+                    name: Some(format!("Expression{}", start)),
+                    node_id: format!("n{}", start),
+                    out_idx: None,
+                    in_indx: None,
+                    contents: boxf.contents.unwrap(),
+                    nbox: bf_counter,
+                    att_bf_idx: 0,
+                };
+                let e1 = Edge {
+                    src: String::from("mod"),
+                    tgt: format!("n{}", start),
+                    e_type: String::from("Contains"),
+                    prop: boxf.contents,
+                };
+                nodes.push(n1.clone());
+                edges.push(e1);
+                if boxf.metadata.as_ref().is_some() {
+                    metadata_idx = boxf.metadata.unwrap();
+                    let mut repeat_meta = false;
+                    for node in meta_nodes.iter() {
+                        if node.metadata_idx == metadata_idx {
+                            repeat_meta = true;
+                        }
+                    }
+                    if !repeat_meta {
+                        meta_nodes.append(&mut create_metadata_node(&gromet.clone(), metadata_idx));
+                        let me1 = Edge {
+                            src: n1.node_id.clone(),
+                            tgt: format!("m{}", metadata_idx),
+                            e_type: String::from("Metadata"),
+                            prop: None,
+                        };
+                        edges.push(me1);
+                    }
+                }
+
+                // now travel to contents index of the attribute list (note it is 1 index,
+                // so contents=1 => attribute[0])
+                // create nodes and edges for this entry, include opo's and opi's
+                start += 1;
+                let idx = boxf.contents.unwrap() - 1;
+
+                let eboxf = gromet.modules[0].attributes[idx as usize].clone();
+                (nodes, edges, meta_nodes, start) = create_opo(
+                    nodes.clone(),
+                    edges.clone(),
+                    meta_nodes.clone(),
+                    &gromet.clone(),
                     start,
-                    c_args.clone(),
+                    n1.clone(),
+                    bf_counter,
                 );
+                (nodes, edges, meta_nodes, start) = create_opi(
+                    nodes.clone(),
+                    edges.clone(),
+                    meta_nodes.clone(),
+                    &gromet.clone(),
+                    start,
+                    n1.clone(),
+                    bf_counter,
+                );
+                // now to construct the nodes inside the expression, Literal and Primitives
+                let mut box_counter: u8 = 1;
+                for sboxf in eboxf.bf.clone().as_ref().unwrap().iter() {
+                    match sboxf.function_type {
+                        FunctionType::Literal => {
+                            (nodes, edges, meta_nodes) = create_att_literal(
+                                &gromet.clone(),
+                                eboxf.clone(),
+                                sboxf.clone(),
+                                nodes.clone(),
+                                edges.clone(),
+                                n1.clone(),
+                                idx,
+                                box_counter,
+                                bf_counter,
+                                start,
+                                meta_nodes.clone(),
+                            );
+                        }
+                        FunctionType::Primitive => {
+                            (nodes, edges, meta_nodes) = create_att_primitive(
+                                &gromet.clone(),
+                                eboxf.clone(),
+                                sboxf.clone(),
+                                nodes.clone(),
+                                edges.clone(),
+                                n1.clone(),
+                                idx,
+                                box_counter,
+                                bf_counter,
+                                start,
+                                meta_nodes.clone(),
+                            );
+                        }
+                        _ => {}
+                    }
+                    box_counter += 1;
+                    start += 1;
+                }
+                // Now we perform the internal wiring of this branch
+                edges = internal_wiring(eboxf.clone(), nodes.clone(), edges, idx, bf_counter);
             }
             FunctionType::Function => {
                 // in order to have less repetition for multiple function calls and to setup support for recursive functions
                 // We check if the function node and thus contents were already made, and not duplicate the contents if already made
-                c_args.att_idx = boxf.contents.unwrap() as usize;
-
                 let mut function_call_repeat = false;
                 let mut original_bf = bf_counter;
                 for node in nodes.clone() {
-                    if (boxf.contents.unwrap() == node.contents as u32)
-                        && (node.n_type == "Function")
-                    {
+                    if (boxf.contents.unwrap() == node.contents) && (node.n_type == "Function") {
                         function_call_repeat = true;
                         if node.nbox < original_bf {
                             original_bf = node.nbox; // This grabs the first instance of bf that the function was called
@@ -893,16 +1190,15 @@ fn create_function_net(gromet: &ModuleCollection, mut start: u32) -> Vec<String>
                         node_id: format!("n{}", start),
                         out_idx: None,
                         in_indx: None,
-                        contents: idx as usize + 1,
+                        contents: idx + 1,
                         nbox: bf_counter,
                         att_bf_idx: 0,
-                        box_counter: 0,
                     };
                     let e1 = Edge {
                         src: String::from("mod"),
                         tgt: format!("n{}", start),
                         e_type: String::from("Contains"),
-                        prop: Some(boxf.contents.unwrap() as usize),
+                        prop: boxf.contents,
                     };
                     nodes.push(n1.clone());
                     edges.push(e1);
@@ -950,30 +1246,29 @@ fn create_function_net(gromet: &ModuleCollection, mut start: u32) -> Vec<String>
                     }
                     // we still construct unique ports for this function, however the contents will not be repeated
                     start += 1;
-
-                    c_args.parent_node = n1.clone();
-
                     // construct opo nodes, if not none
-                    create_opo(
-                        &gromet,    // gromet for metadata
-                        &mut nodes, // nodes
-                        &mut edges,
-                        &mut meta_nodes,
+                    (nodes, edges, meta_nodes, start) = create_opo(
+                        nodes.clone(),
+                        edges.clone(),
+                        meta_nodes.clone(),
+                        &gromet.clone(),
                         start,
-                        c_args.clone(),
+                        n1.clone(),
+                        bf_counter,
                     );
-                    create_opi(
-                        &gromet,    // gromet for metadata
-                        &mut nodes, // nodes
-                        &mut edges,
-                        &mut meta_nodes,
+                    (nodes, edges, meta_nodes, start) = create_opi(
+                        nodes.clone(),
+                        edges.clone(),
+                        meta_nodes.clone(),
+                        &gromet.clone(),
                         start,
-                        c_args.clone(),
+                        n1.clone(),
+                        bf_counter,
                     );
                     // now to add the contains wires for the additional function call onto the original contents nodes:
                     for node in nodes.clone() {
                         if (node.nbox == original_bf)
-                            && (node.contents == (idx as usize + 1))
+                            && (node.contents == (idx + 1))
                             && ((node.n_type == "Literal")
                                 || (node.n_type == "Primitive")
                                 || (node.n_type == "Predicate")
@@ -1000,7 +1295,7 @@ fn create_function_net(gromet: &ModuleCollection, mut start: u32) -> Vec<String>
                             // make sure in correct box
                             if original_bf == node.nbox {
                                 // make sure only looking in current attribute nodes for srcs and tgts
-                                if (idx + 1) == node.contents as u32 {
+                                if (idx + 1) == node.contents {
                                     // only include nodes with pifs
                                     if node.in_indx.is_some() {
                                         // exclude opi's
@@ -1022,7 +1317,7 @@ fn create_function_net(gromet: &ModuleCollection, mut start: u32) -> Vec<String>
                             // make sure in correct box
                             if bf_counter == node.nbox {
                                 // make sure only looking in current attribute nodes for srcs and tgts
-                                if (idx + 1) == node.contents as u32 {
+                                if (idx + 1) == node.contents {
                                     // only opi's
                                     if node.n_type == "Opi" {
                                         // iterate through port to check for tgt
@@ -1053,7 +1348,7 @@ fn create_function_net(gromet: &ModuleCollection, mut start: u32) -> Vec<String>
                             // make sure in correct box
                             if bf_counter == node.nbox {
                                 // make sure only looking in current attribute nodes for srcs and tgts
-                                if (idx + 1) == node.contents as u32 {
+                                if (idx + 1) == node.contents {
                                     // only opo's
                                     if node.n_type == "Opo" {
                                         // iterate through port to check for tgt
@@ -1072,7 +1367,7 @@ fn create_function_net(gromet: &ModuleCollection, mut start: u32) -> Vec<String>
                             // make sure in correct box
                             if original_bf == node.nbox {
                                 // make sure only looking in current attribute nodes for srcs and tgts
-                                if (idx + 1) == node.contents as u32 {
+                                if (idx + 1) == node.contents {
                                     // only include nodes with pofs
                                     if node.out_idx.is_some() {
                                         // exclude opo's
@@ -1100,15 +1395,35 @@ fn create_function_net(gromet: &ModuleCollection, mut start: u32) -> Vec<String>
                         }
                     }
                 } else {
-                    // new function call
-                    c_args.att_idx = boxf.contents.unwrap() as usize;
-                    create_function(
-                        &gromet,    // gromet for metadata
-                        &mut nodes, // nodes
-                        &mut edges,
-                        &mut meta_nodes,
-                        start,
-                        c_args.clone(),
+                    // mark
+                    // get attribute function net to pass into function
+                    let pfn =
+                        gromet.modules[0].attributes[(boxf.contents.unwrap() - 1) as usize].clone();
+
+                    // need a parent node for the module at this level
+                    let temp_mod_node = Node {
+                        n_type: String::from("module"),
+                        value: None,
+                        name: None,
+                        node_id: "mod".to_string(),
+                        out_idx: None,
+                        in_indx: None,
+                        contents: 0,
+                        nbox: 0,
+                        att_bf_idx: 0,
+                    };
+
+                    (nodes, edges, start, meta_nodes) = create_function(
+                        &gromet.clone(),        // gromet for metadata
+                        pfn.clone(),            // function net for attribute
+                        nodes.clone(),          // nodes
+                        edges.clone(),          // edges
+                        temp_mod_node.clone(),  // parent node
+                        boxf.contents.unwrap(), // attribute idx
+                        bf_counter,             // top level bf idx
+                        0,                      // sub box idx, 0 for top level
+                        start,                  // start
+                        meta_nodes.clone(),     // metadata nodes
                     );
                 }
             }
@@ -1122,7 +1437,7 @@ fn create_function_net(gromet: &ModuleCollection, mut start: u32) -> Vec<String>
     // between attributes
     // get wired through module level wff field, will require reading through node list to
     // match contents field to box field on wff entries
-    external_wiring(gromet, &mut nodes.clone(), &mut edges);
+    edges = external_wiring(gromet, nodes.clone(), edges);
 
     // make conditionals if they exist
     if gromet.modules[0].r#fn.bc.as_ref().is_some() {
@@ -1137,18 +1452,20 @@ fn create_function_net(gromet: &ModuleCollection, mut start: u32) -> Vec<String>
             contents: 0,
             nbox: 0,
             att_bf_idx: 0,
-            box_counter: 0,
         };
         for _cond in gromet.modules[0].r#fn.bc.as_ref().unwrap().iter() {
             // now lets check for and setup any conditionals at this level
-            create_conditional(
-                &gromet,    // gromet for metadata
-                &mut nodes, // nodes
-                &mut edges,
-                &mut meta_nodes,
+            (nodes, edges, start, meta_nodes) = create_conditional(
+                &gromet.clone(),
+                gromet.modules[0].r#fn.clone(), // This is gromet but is more generalizable based on scope
+                nodes.clone(),
+                edges.clone(),
+                temp_mod_node.clone(),
+                0,            // because top level
+                cond_counter, // This indexes the conditional in the list of conditionals (bc)
+                0,            // because top level
                 start,
-                c_args.clone(),
-                cond_counter,
+                meta_nodes.clone(),
             );
             cond_counter += 1;
         }
@@ -1166,18 +1483,20 @@ fn create_function_net(gromet: &ModuleCollection, mut start: u32) -> Vec<String>
             contents: 0,
             nbox: 0,
             att_bf_idx: 0,
-            box_counter: 0,
         };
         for _while_l in gromet.modules[0].r#fn.bl.as_ref().unwrap().iter() {
             // now lets check for and setup any conditionals at this level
-            create_while_loop(
-                &gromet,    // gromet for metadata
-                &mut nodes, // nodes
-                &mut edges,
-                &mut meta_nodes,
+            (nodes, edges, start, meta_nodes) = create_while_loop(
+                &gromet.clone(),
+                gromet.modules[0].r#fn.clone(), // This is gromet but is more generalizable based on scope
+                nodes.clone(),
+                edges.clone(),
+                temp_mod_node.clone(),
+                0,             // because top level
+                while_counter, // This indexes the conditional in the list of conditionals (bc)
+                0,             // because top level
                 start,
-                c_args.clone(),
-                while_counter,
+                meta_nodes.clone(),
             );
             while_counter += 1;
         }
@@ -1223,25 +1542,27 @@ fn create_function_net(gromet: &ModuleCollection, mut start: u32) -> Vec<String>
     queries
 }
 // this method creates an import type function
-// currently assumes top level call
 #[allow(unused_assignments)]
 pub fn create_import(
     gromet: &ModuleCollection, // needed still for metadata unfortunately
-    nodes: &mut Vec<Node>,
-    edges: &mut Vec<Edge>,
-    meta_nodes: &mut Vec<MetadataNode>,
-    mut start: u32, // for node and edge indexing
-    c_args: ConstructorArgs,
-) -> () {
+    mut nodes: Vec<Node>,
+    mut edges: Vec<Edge>,
+    parent_node: Node, // used for contains construction
+    att_idx: u32,      // This will index the attribute the function is in
+    bf_counter: u8, // This indexes which box the function is under, inherited from parent if not explicit
+    att_bf_idx: u32, // This indexes if the function is a subscope of a larger function, 0 if not
+    start: u32,     // for node and edge indexing
+    mut meta_nodes: Vec<MetadataNode>,
+) -> (Vec<Node>, Vec<Edge>, u32, Vec<MetadataNode>) {
     let eboxf = gromet.modules[0].clone();
-    let sboxf = gromet.modules[0].attributes[(c_args.att_idx - 1) as usize].clone();
-    let mboxf = eboxf.r#fn.bf.unwrap()[(c_args.bf_counter - 1) as usize].clone();
+    let sboxf = gromet.modules[0].attributes[(att_idx - 1) as usize].clone();
+    let mboxf = eboxf.r#fn.bf.unwrap()[(bf_counter - 1) as usize].clone();
 
     let mut pof: Vec<u32> = vec![];
     if eboxf.r#fn.pof.is_some() {
         let mut po_idx: u32 = 1;
         for port in eboxf.r#fn.pof.clone().unwrap().iter() {
-            if port.r#box == c_args.bf_counter as u8 {
+            if port.r#box == bf_counter {
                 pof.push(po_idx);
             }
             po_idx += 1;
@@ -1252,7 +1573,7 @@ pub fn create_import(
     if eboxf.r#fn.pif.is_some() {
         let mut pi_idx: u32 = 1;
         for port in eboxf.r#fn.pif.unwrap().iter() {
-            if port.r#box == c_args.bf_counter as u8 {
+            if port.r#box == bf_counter {
                 pif.push(pi_idx);
             }
             pi_idx += 1;
@@ -1267,15 +1588,14 @@ pub fn create_import(
         node_id: format!("n{}", start),
         out_idx: Some(pof),
         in_indx: Some(pif),
-        contents: c_args.att_idx,
-        nbox: c_args.bf_counter,
-        att_bf_idx: c_args.att_bf_idx,
-        box_counter: c_args.box_counter,
+        contents: att_idx,
+        nbox: bf_counter,
+        att_bf_idx,
     };
     nodes.push(n3.clone());
     // make edge connecting to expression
     let e4 = Edge {
-        src: c_args.parent_node.node_id,
+        src: parent_node.node_id,
         tgt: n3.node_id.clone(),
         e_type: String::from("Contains"),
         prop: None,
@@ -1300,44 +1620,61 @@ pub fn create_import(
             edges.push(me1);
         }
     }
-    start += 1;
+
+    (nodes, edges, start, meta_nodes)
 }
 
 // this creates a function node including all the contents included in it, including additional functions
-// CHANGEs: removed function_net arg and added box_counter arg
 #[allow(unused_assignments)]
 pub fn create_function(
     gromet: &ModuleCollection, // needed still for metadata unfortunately
-    nodes: &mut Vec<Node>,
-    edges: &mut Vec<Edge>,
-    meta_nodes: &mut Vec<MetadataNode>,
+    function_net: FunctionNet, // This is gromet but is more generalizable based on scope
+    mut nodes: Vec<Node>,
+    mut edges: Vec<Edge>,
+    parent_node: Node, // used for contains construction
+    att_idx: u32,      // This will index the attribute the function is in
+    bf_counter: u8, // This indexes which box the function is under, inherited from parent if not explicit
+    att_bf_idx: u32, // This indexes if the function is a subscope of a larger function, 0 if not
     mut start: u32, // for node and edge indexing
-    c_args: ConstructorArgs,
-) -> () {
+    mut meta_nodes: Vec<MetadataNode>,
+) -> (Vec<Node>, Vec<Edge>, u32, Vec<MetadataNode>) {
     // function is not repeated
-    let att_box = gromet.modules[0].attributes[(c_args.att_idx - 1) as usize].clone(); // current expression attribute
+    let eboxf = gromet.modules[0].attributes[(att_idx - 1) as usize].clone();
 
     // now we add a check for if this is an imported function
-    match att_box.b.clone().unwrap()[0].function_type.clone() {
+    match eboxf.b.clone().unwrap()[0].function_type.clone() {
         FunctionType::Imported => {
-            create_import(&gromet, nodes, edges, meta_nodes, start, c_args.clone());
+            (nodes, edges, start, meta_nodes) = create_import(
+                &gromet.clone(),
+                nodes.clone(),
+                edges.clone(),
+                parent_node.clone(),
+                att_idx,
+                bf_counter,
+                att_bf_idx,
+                start,
+                meta_nodes.clone(),
+            );
             start += 1;
             // now to implement wiring
-            import_wiring(
+            edges = import_wiring(
                 &gromet.clone(),
-                nodes,
+                eboxf,
+                nodes.clone(),
                 edges,
-                c_args.att_idx,
-                c_args.bf_counter,
-                c_args.parent_node,
+                att_idx,
+                bf_counter,
+                parent_node,
             );
+
+            (nodes, edges, start, meta_nodes)
         }
         _ => {
             let n1 = Node {
                 n_type: String::from("Function"),
                 value: None,
                 name: Some(
-                    att_box.b.as_ref().unwrap()[0]
+                    eboxf.b.as_ref().unwrap()[0]
                         .name
                         .clone()
                         .map_or_else(|| format!("Function{}", start), |x| x),
@@ -1345,23 +1682,22 @@ pub fn create_function(
                 node_id: format!("n{}", start),
                 out_idx: None,
                 in_indx: None,
-                contents: c_args.att_idx,
-                nbox: c_args.bf_counter,
-                att_bf_idx: c_args.att_bf_idx,
-                box_counter: c_args.box_counter,
+                contents: att_idx,
+                nbox: bf_counter,
+                att_bf_idx,
             };
             let e1 = Edge {
-                src: c_args.parent_node.node_id.clone(),
+                src: parent_node.node_id,
                 tgt: n1.node_id.clone(),
                 e_type: String::from("Contains"),
-                prop: Some(c_args.att_idx),
+                prop: Some(att_idx),
             };
             nodes.push(n1.clone());
             edges.push(e1);
             let mut metadata_idx = 0;
             // attribute b level metadata reference
-            if att_box.b.as_ref().unwrap()[0].metadata.as_ref().is_some() {
-                metadata_idx = att_box.b.as_ref().unwrap()[0].metadata.unwrap();
+            if eboxf.b.as_ref().unwrap()[0].metadata.as_ref().is_some() {
+                metadata_idx = eboxf.b.as_ref().unwrap()[0].metadata.unwrap();
                 let mut repeat_meta = false;
                 for node in meta_nodes.iter() {
                     if node.metadata_idx == metadata_idx {
@@ -1386,86 +1722,103 @@ pub fn create_function(
             // create nodes and edges for this entry, include opo's and opi's
             start += 1;
 
-            let mut new_c_args = c_args.clone();
-            new_c_args.parent_node = n1.clone();
-
             // construct opo nodes, if not none, might need to
-            create_opo(
-                &gromet, // gromet for metadata
-                nodes,   // nodes
-                edges,
-                meta_nodes,
+            (nodes, edges, meta_nodes, start) = create_opo(
+                nodes.clone(),
+                edges.clone(),
+                meta_nodes.clone(),
+                &gromet.clone(),
                 start,
-                new_c_args.clone(),
+                n1.clone(),
+                bf_counter,
             );
-            create_opi(
-                &gromet, // gromet for metadata
-                nodes,   // nodes
-                edges,
-                meta_nodes,
+            (nodes, edges, meta_nodes, start) = create_opi(
+                nodes.clone(),
+                edges.clone(),
+                meta_nodes.clone(),
+                &gromet.clone(),
                 start,
-                new_c_args.clone(),
+                n1.clone(),
+                bf_counter,
             );
             // now to construct the nodes inside the function, currently supported Literals and Primitives
             // first include an Expression for increased depth
-            let mut box_counter: usize = 1;
-            for att_sub_box in att_box.bf.as_ref().unwrap().iter() {
-                new_c_args.box_counter = box_counter;
-                new_c_args.cur_box = att_sub_box.clone();
-                new_c_args.att_idx = c_args.att_idx;
-                match att_sub_box.function_type {
+            let mut box_counter: u8 = 1;
+            for sboxf in eboxf.bf.as_ref().unwrap().iter() {
+                match sboxf.function_type {
                     FunctionType::Function => {
-                        new_c_args.att_idx = att_sub_box.contents.unwrap() as usize;
-                        create_function(
-                            &gromet, // gromet for metadata
-                            nodes,   // nodes
-                            edges,
-                            meta_nodes,
-                            start,
-                            new_c_args.clone(),
+                        (nodes, edges, start, meta_nodes) = create_function(
+                            &gromet.clone(),         // gromet for metadata
+                            eboxf.clone(),           // function net for attribute
+                            nodes.clone(),           // nodes
+                            edges.clone(),           // edges
+                            n1.clone(),              // parent node
+                            sboxf.contents.unwrap(), // attribute idx
+                            bf_counter,              // top level bf idx
+                            box_counter.into(),      // sub box idx, 0 for top level
+                            start,                   // start
+                            meta_nodes.clone(),      // metadata nodes
                         );
                     }
 
                     FunctionType::Predicate => {
-                        new_c_args.att_idx = att_sub_box.contents.unwrap() as usize;
-                        create_att_predicate(
-                            &gromet, // gromet for metadata
-                            nodes,   // nodes
-                            edges,
-                            meta_nodes,
+                        (nodes, edges, start, meta_nodes) = create_att_predicate(
+                            &gromet.clone(),
+                            eboxf.clone(),
+                            sboxf.clone(),
+                            nodes.clone(),
+                            edges.clone(),
+                            n1.clone(),
+                            att_idx,
+                            box_counter,
+                            bf_counter,
                             start,
-                            new_c_args.clone(),
+                            meta_nodes.clone(),
                         );
                     }
                     FunctionType::Expression => {
-                        new_c_args.att_idx = att_sub_box.contents.unwrap() as usize;
-                        create_att_expression(
-                            &gromet, // gromet for metadata
-                            nodes,   // nodes
-                            edges,
-                            meta_nodes,
+                        (nodes, edges, start, meta_nodes) = create_att_expression(
+                            &gromet.clone(),
+                            eboxf.clone(),
+                            sboxf.clone(),
+                            nodes.clone(),
+                            edges.clone(),
+                            n1.clone(),
+                            att_idx,
+                            box_counter,
+                            bf_counter,
                             start,
-                            new_c_args.clone(),
+                            meta_nodes.clone(),
                         );
                     }
                     FunctionType::Literal => {
-                        create_att_literal(
-                            &gromet, // gromet for metadata
-                            nodes,   // nodes
-                            edges,
-                            meta_nodes,
+                        (nodes, edges, meta_nodes) = create_att_literal(
+                            &gromet.clone(),
+                            eboxf.clone(),
+                            sboxf.clone(),
+                            nodes.clone(),
+                            edges.clone(),
+                            n1.clone(),
+                            att_idx,
+                            box_counter,
+                            bf_counter,
                             start,
-                            new_c_args.clone(),
+                            meta_nodes.clone(),
                         );
                     }
                     FunctionType::Primitive => {
-                        create_att_primitive(
-                            &gromet, // gromet for metadata
-                            nodes,   // nodes
-                            edges,
-                            meta_nodes,
+                        (nodes, edges, meta_nodes) = create_att_primitive(
+                            &gromet.clone(),
+                            eboxf.clone(),
+                            sboxf.clone(),
+                            nodes.clone(),
+                            edges.clone(),
+                            n1.clone(),
+                            att_idx,
+                            box_counter,
+                            bf_counter,
                             start,
-                            new_c_args.clone(),
+                            meta_nodes.clone(),
                         );
                     }
                     // stub for where to put in create function for nested functions
@@ -1476,21 +1829,10 @@ pub fn create_function(
             }
 
             // Now we perform the internal wiring of this branch
-            internal_wiring(
-                att_box.clone(),
-                nodes,
-                edges,
-                c_args.att_idx,
-                c_args.bf_counter,
-            );
+            edges = internal_wiring(eboxf.clone(), nodes.clone(), edges, att_idx, bf_counter);
             // perform cross attributal wiring of function
-            cross_att_wiring(
-                att_box.clone(),
-                nodes,
-                edges,
-                c_args.att_idx,
-                c_args.bf_counter,
-            );
+            (edges, nodes) =
+                cross_att_wiring(eboxf.clone(), nodes.clone(), edges, att_idx, bf_counter);
 
             // make conditionals if they exist
             // might need to make wiring more robust
@@ -1498,14 +1840,17 @@ pub fn create_function(
                 let mut cond_counter = 0;
                 for _cond in gromet.modules[0].r#fn.bc.as_ref().unwrap().iter() {
                     // now lets check for and setup any conditionals at this level
-                    create_conditional(
-                        &gromet, // gromet for metadata
-                        nodes,   // nodes
-                        edges,
-                        meta_nodes,
+                    (nodes, edges, start, meta_nodes) = create_conditional(
+                        &gromet.clone(),
+                        eboxf.clone(), // This is gromet but is more generalizable based on scope
+                        nodes.clone(),
+                        edges.clone(),
+                        n1.clone(),
+                        att_idx - 1,  // because top level
+                        cond_counter, // This indexes the conditional in the list of conditionals (bc)
+                        bf_counter,   // because top level
                         start,
-                        c_args.clone(),
-                        cond_counter,
+                        meta_nodes.clone(),
                     );
                     cond_counter += 1;
                 }
@@ -1516,18 +1861,22 @@ pub fn create_function(
                 let mut while_counter = 0;
                 for _while_l in gromet.modules[0].r#fn.bl.as_ref().unwrap().iter() {
                     // now lets check for and setup any conditionals at this level
-                    create_while_loop(
-                        &gromet, // gromet for metadata
-                        nodes,   // nodes
-                        edges,
-                        meta_nodes,
+                    (nodes, edges, start, meta_nodes) = create_while_loop(
+                        &gromet.clone(),
+                        eboxf.clone(), // This is gromet but is more generalizable based on scope
+                        nodes.clone(),
+                        edges.clone(),
+                        n1.clone(),
+                        att_idx - 1,   // because top level
+                        while_counter, // This indexes the conditional in the list of conditionals (bc)
+                        bf_counter,    // because top level
                         start,
-                        c_args.clone(),
-                        while_counter,
+                        meta_nodes.clone(),
                     );
                     while_counter += 1;
                 }
             }
+            (nodes, edges, start, meta_nodes)
         }
     }
 }
@@ -1537,20 +1886,17 @@ pub fn create_function(
 // The iterator through the conditionals will need to be outside this function
 #[allow(unused_assignments)]
 pub fn create_conditional(
-    gromet: &ModuleCollection, // needed still for metadata unfortunately
-    nodes: &mut Vec<Node>,
-    edges: &mut Vec<Edge>,
-    meta_nodes: &mut Vec<MetadataNode>,
-    mut start: u32, // for node and edge indexing
-    c_args: ConstructorArgs,
-    cond_counter: u32,
-) -> () {
-    let mut bf_counter = c_args.bf_counter;
-    let mut att_idx = c_args.att_idx;
-    let mut att_bf_idx = c_args.att_bf_idx;
-    let mut box_counter = c_args.box_counter;
-    let function_net = c_args.att_box;
-
+    gromet: &ModuleCollection,
+    function_net: FunctionNet, // This is gromet but is more generalizable based on scope
+    mut nodes: Vec<Node>,
+    mut edges: Vec<Edge>,
+    parent_node: Node,
+    idx_in: u32,       // This will index the attribute the conditional is in, if any
+    cond_counter: u32, // This indexes the conditional in the list of conditionals (bc)
+    bf_counter: u8,    // This indexes which box the conditional is under, if any
+    mut start: u32,
+    mut meta_nodes: Vec<MetadataNode>,
+) -> (Vec<Node>, Vec<Edge>, u32, Vec<MetadataNode>) {
     let mut metadata_idx = 0;
     let n1 = Node {
         n_type: String::from("Conditional"),
@@ -1559,16 +1905,15 @@ pub fn create_conditional(
         node_id: format!("n{}", start),
         out_idx: None,
         in_indx: None,
-        contents: att_idx,
+        contents: idx_in + 1,
         nbox: bf_counter,
-        att_bf_idx: att_bf_idx,
-        box_counter: box_counter,
+        att_bf_idx: 0,
     };
     let e1 = Edge {
-        src: c_args.parent_node.node_id,
+        src: parent_node.node_id,
         tgt: format!("n{}", start),
         e_type: String::from("Contains"),
-        prop: Some(cond_counter as usize),
+        prop: Some(cond_counter),
     };
     nodes.push(n1.clone());
     edges.push(e1);
@@ -1618,10 +1963,9 @@ pub fn create_conditional(
                 node_id: format!("n{}", start),
                 out_idx: None,
                 in_indx: Some([port_count].to_vec()),
-                contents: att_idx,
+                contents: idx_in + 1,
                 nbox: bf_counter,
                 att_bf_idx: 0,
-                box_counter: 0,
             };
             let e3 = Edge {
                 src: n1.node_id.clone(),
@@ -1671,10 +2015,9 @@ pub fn create_conditional(
                 node_id: format!("n{}", start),
                 out_idx: Some([port_count].to_vec()),
                 in_indx: None,
-                contents: att_idx,
+                contents: idx_in + 1,
                 nbox: bf_counter,
                 att_bf_idx: 0,
-                box_counter: 0,
             };
             let e5 = Edge {
                 src: n1.node_id.clone(),
@@ -1728,7 +2071,7 @@ pub fn create_conditional(
         if node.contents
             == function_net.bf.as_ref().unwrap()[(cond_box - 1) as usize]
                 .contents
-                .unwrap() as usize
+                .unwrap()
         {
             // make sure box matches correctly
             if (node.nbox as u32) == cond_box {
@@ -1741,7 +2084,7 @@ pub fn create_conditional(
         if node.contents
             == function_net.bf.as_ref().unwrap()[(body_if_box - 1) as usize]
                 .contents
-                .unwrap() as usize
+                .unwrap()
         {
             // make sure box matches correctly
             if (node.nbox as u32) == body_if_box {
@@ -1754,7 +2097,7 @@ pub fn create_conditional(
         if node.contents
             == function_net.bf.as_ref().unwrap()[(body_else_box - 1) as usize]
                 .contents
-                .unwrap() as usize
+                .unwrap()
         {
             // make sure box matches correctly
             if (node.nbox as u32) == body_else_box {
@@ -1793,7 +2136,7 @@ pub fn create_conditional(
     for wire in function_net.wfc.as_ref().unwrap().iter() {
         // collect info to identify the opi src node
         let src_idx = wire.src; // port index
-        let src_att = att_idx; // attribute index of submodule (also opi contents value)
+        let src_att = idx_in + 1; // attribute index of submodule (also opi contents value)
         let src_nbox = bf_counter; // nbox value of src opi
         let src_pic_idx = src_idx;
 
@@ -1831,9 +2174,9 @@ pub fn create_conditional(
         }
         for node in nodes.iter() {
             // make sure in correct box
-            if tgt_nbox == node.nbox as u8 {
+            if tgt_nbox == node.nbox {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if tgt_att == node.contents as u32 {
+                if tgt_att == node.contents {
                     // only opo's
                     if node.n_type == "Opo" {
                         // iterate through port to check for tgt
@@ -1873,7 +2216,7 @@ pub fn create_conditional(
         let _tgt_pic = function_net.pic.as_ref().unwrap()[(tgt_idx - 1) as usize].clone(); // tgt port
         let tgt_box = bf_counter; // tgt sub module box number
 
-        let tgt_att = att_idx; // attribute index of submodule (also opo contents value)
+        let tgt_att = idx_in + 1; // attribute index of submodule (also opo contents value)
         let tgt_nbox = tgt_box; // nbox value of tgt opo
         let tgt_pic_idx = tgt_idx;
 
@@ -1882,9 +2225,9 @@ pub fn create_conditional(
         // find the src node
         for node in nodes.iter() {
             // make sure in correct box
-            if src_nbox == node.nbox as u8 {
+            if src_nbox == node.nbox {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if src_att == node.contents as u32 {
+                if src_att == node.contents {
                     // only opo's
                     if node.n_type == "Opi" {
                         // iterate through port to check for tgt
@@ -1959,7 +2302,7 @@ pub fn create_conditional(
         // setting up the pic is straight forward
         let tgt_idx = pic_counter; // port index
         let tgt_box = bf_counter; // tgt sub module box number
-        let tgt_att = att_idx; // attribute index of submodule (also opo contents value)
+        let tgt_att = idx_in + 1; // attribute index of submodule (also opo contents value)
         let tgt_nbox = tgt_box; // nbox value of tgt opo
         let tgt_pic_idx = tgt_idx;
 
@@ -1971,7 +2314,7 @@ pub fn create_conditional(
             // make sure in correct box
             if src_if_nbox == (node.nbox as u32) {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if src_if_att == node.contents as u32 {
+                if src_if_att == node.contents {
                     // only opo's
                     if node.n_type == "Opi" {
                         // iterate through port to check for tgt
@@ -1990,7 +2333,7 @@ pub fn create_conditional(
             // make sure in correct box
             if src_else_nbox == (node.nbox as u32) {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if src_else_att == node.contents as u32 {
+                if src_else_att == node.contents {
                     // only opo's
                     if node.n_type == "Opi" {
                         // iterate through port to check for tgt
@@ -2076,7 +2419,7 @@ pub fn create_conditional(
         // setting up the pic is straight forward
         let src_idx = poc_counter; // port index
         let src_box = bf_counter; // tgt sub module box number
-        let src_att = att_idx; // attribute index of submodule (also opo contents value)
+        let src_att = idx_in + 1; // attribute index of submodule (also opo contents value)
         let src_nbox = src_box; // nbox value of tgt opo
         let src_poc_idx = src_idx;
 
@@ -2108,7 +2451,7 @@ pub fn create_conditional(
             // make sure in correct box
             if tgt_else_nbox == (node.nbox as u32) {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if tgt_else_att == node.contents as u32 {
+                if tgt_else_att == node.contents {
                     // only opo's
                     if node.n_type == "Opo" {
                         // iterate through port to check for tgt
@@ -2127,7 +2470,7 @@ pub fn create_conditional(
             // make sure in correct box
             if tgt_if_nbox == (node.nbox as u32) {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if tgt_if_att == node.contents as u32 {
+                if tgt_if_att == node.contents {
                     // only opo's
                     if node.n_type == "Opo" {
                         // iterate through port to check for tgt
@@ -2162,23 +2505,22 @@ pub fn create_conditional(
         poc_counter += 1;
     }
     // might still need pass through wiring??
+
+    (nodes, edges, start, meta_nodes)
 }
 #[allow(unused_assignments)]
 pub fn create_while_loop(
-    gromet: &ModuleCollection, // needed still for metadata unfortunately
-    nodes: &mut Vec<Node>,
-    edges: &mut Vec<Edge>,
-    meta_nodes: &mut Vec<MetadataNode>,
-    mut start: u32, // for node and edge indexing
-    c_args: ConstructorArgs,
-    cond_counter: u32,
-) -> () {
-    let mut bf_counter = c_args.bf_counter;
-    let mut att_idx = c_args.att_idx;
-    let mut att_bf_idx = c_args.att_bf_idx;
-    let mut box_counter = c_args.box_counter;
-    let function_net = c_args.att_box;
-
+    gromet: &ModuleCollection,
+    function_net: FunctionNet, // This is gromet but is more generalizable based on scope
+    mut nodes: Vec<Node>,
+    mut edges: Vec<Edge>,
+    parent_node: Node,
+    idx_in: u32,       // This will index the attribute the conditional is in, if any
+    cond_counter: u32, // This indexes the conditional in the list of conditionals (bc)
+    bf_counter: u8,    // This indexes which box the conditional is under, if any
+    mut start: u32,
+    mut meta_nodes: Vec<MetadataNode>,
+) -> (Vec<Node>, Vec<Edge>, u32, Vec<MetadataNode>) {
     let mut metadata_idx = 0;
     let n1 = Node {
         n_type: String::from("While_Loop"),
@@ -2187,16 +2529,15 @@ pub fn create_while_loop(
         node_id: format!("n{}", start),
         out_idx: None,
         in_indx: None,
-        contents: att_idx,
+        contents: idx_in + 1,
         nbox: bf_counter,
         att_bf_idx: 0,
-        box_counter: 0,
     };
     let e1 = Edge {
-        src: c_args.parent_node.node_id,
+        src: parent_node.node_id,
         tgt: format!("n{}", start),
         e_type: String::from("Contains"),
-        prop: Some(cond_counter as usize),
+        prop: Some(cond_counter),
     };
     nodes.push(n1.clone());
     edges.push(e1);
@@ -2246,10 +2587,9 @@ pub fn create_while_loop(
                 node_id: format!("n{}", start),
                 out_idx: None,
                 in_indx: Some([port_count].to_vec()),
-                contents: att_idx,
+                contents: idx_in + 1,
                 nbox: bf_counter,
                 att_bf_idx: 0,
-                box_counter: 0,
             };
             let e3 = Edge {
                 src: n1.node_id.clone(),
@@ -2298,10 +2638,9 @@ pub fn create_while_loop(
                 node_id: format!("n{}", start),
                 out_idx: Some([port_count].to_vec()),
                 in_indx: None,
-                contents: att_idx,
+                contents: idx_in + 1,
                 nbox: bf_counter,
                 att_bf_idx: 0,
-                box_counter: 0,
             };
             let e5 = Edge {
                 src: n1.node_id.clone(),
@@ -2351,7 +2690,7 @@ pub fn create_while_loop(
         if node.contents
             == function_net.bf.as_ref().unwrap()[(cond_box - 1) as usize]
                 .contents
-                .unwrap() as usize
+                .unwrap()
         {
             // make sure box matches correctly
             if (node.nbox as u32) == cond_box {
@@ -2364,7 +2703,7 @@ pub fn create_while_loop(
         if node.contents
             == function_net.bf.as_ref().unwrap()[(body_if_box - 1) as usize]
                 .contents
-                .unwrap() as usize
+                .unwrap()
         {
             // make sure box matches correctly
             if (node.nbox as u32) == body_if_box {
@@ -2396,7 +2735,7 @@ pub fn create_while_loop(
     for wire in function_net.wfl.as_ref().unwrap().iter() {
         // collect info to identify the opi src node
         let src_idx = wire.src; // port index
-        let src_att = att_idx; // attribute index of submodule (also opi contents value)
+        let src_att = idx_in + 1; // attribute index of submodule (also opi contents value)
         let src_nbox = bf_counter; // nbox value of src opi
         let src_pil_idx = src_idx;
 
@@ -2434,9 +2773,9 @@ pub fn create_while_loop(
         }
         for node in nodes.iter() {
             // make sure in correct box
-            if tgt_nbox == node.nbox as u8 {
+            if tgt_nbox == node.nbox {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if tgt_att == node.contents as u32 {
+                if tgt_att == node.contents {
                     // only opo's
                     if node.n_type == "Opo" {
                         // iterate through port to check for tgt
@@ -2476,7 +2815,7 @@ pub fn create_while_loop(
         let _tgt_pil = function_net.pil.as_ref().unwrap()[(tgt_idx - 1) as usize].clone(); // tgt port
         let tgt_box = bf_counter; // tgt sub module box number
 
-        let tgt_att = att_idx; // attribute index of submodule (also opo contents value)
+        let tgt_att = idx_in + 1; // attribute index of submodule (also opo contents value)
         let tgt_nbox = tgt_box; // nbox value of tgt opo
         let tgt_pil_idx = tgt_idx;
 
@@ -2485,9 +2824,9 @@ pub fn create_while_loop(
         // find the src node
         for node in nodes.iter() {
             // make sure in correct box
-            if src_nbox == node.nbox as u8 {
+            if src_nbox == node.nbox {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if src_att == node.contents as u32 {
+                if src_att == node.contents {
                     // only opo's
                     if node.n_type == "Opi" {
                         // iterate through port to check for tgt
@@ -2553,7 +2892,7 @@ pub fn create_while_loop(
         // setting up the pic is straight forward
         let tgt_idx = pic_counter; // port index
         let tgt_box = bf_counter; // tgt sub module box number
-        let tgt_att = att_idx; // attribute index of submodule (also opo contents value)
+        let tgt_att = idx_in + 1; // attribute index of submodule (also opo contents value)
         let tgt_nbox = tgt_box; // nbox value of tgt opo
         let tgt_pil_idx = tgt_idx;
 
@@ -2564,7 +2903,7 @@ pub fn create_while_loop(
             // make sure in correct box
             if src_if_nbox == (node.nbox as u32) {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if src_if_att == node.contents as u32 {
+                if src_if_att == node.contents {
                     // only opo's
                     if node.n_type == "Opi" {
                         // iterate through port to check for tgt
@@ -2631,7 +2970,7 @@ pub fn create_while_loop(
         // setting up the pic is straight forward
         let src_idx = poc_counter; // port index
         let src_box = bf_counter; // tgt sub module box number
-        let src_att = att_idx; // attribute index of submodule (also opo contents value)
+        let src_att = idx_in + 1; // attribute index of submodule (also opo contents value)
         let src_nbox = src_box; // nbox value of tgt opo
         let src_pol_idx = src_idx;
 
@@ -2661,7 +3000,7 @@ pub fn create_while_loop(
             // make sure in correct box
             if tgt_if_nbox == (node.nbox as u32) {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if tgt_if_att == node.contents as u32 {
+                if tgt_if_att == node.contents {
                     // only opo's
                     if node.n_type == "Opo" {
                         // iterate through port to check for tgt
@@ -2687,19 +3026,28 @@ pub fn create_while_loop(
         poc_counter += 1;
     }
     // might still need pass through wiring??
+
+    (nodes, edges, start, meta_nodes)
 }
 
 // This needs to be updated to handle the new node structure and remove the overloaded contents field which will mess with the wiring alot
 #[allow(unused_assignments)]
 pub fn create_att_expression(
-    gromet: &ModuleCollection, // needed still for metadata unfortunately
-    nodes: &mut Vec<Node>,
-    edges: &mut Vec<Edge>,
-    meta_nodes: &mut Vec<MetadataNode>,
-    mut start: u32, // for node and edge indexing
-    c_args: ConstructorArgs,
-) -> () {
+    gromet: &ModuleCollection,
+    _eeboxf: FunctionNet,
+    ssboxf: GrometBox,
+    mut nodes: Vec<Node>,
+    mut edges: Vec<Edge>,
+    parent_node: Node,
+    _idx_in: u32,
+    box_counter: u8,
+    bf_counter: u8,
+    mut start: u32,
+    mut meta_nodes: Vec<MetadataNode>,
+) -> (Vec<Node>, Vec<Edge>, u32, Vec<MetadataNode>) {
     let mut metadata_idx = 0;
+
+    let idx = ssboxf.contents.unwrap(); // this is the attribute index of the current expression
 
     let n1 = Node {
         n_type: String::from("Expression"),
@@ -2708,24 +3056,20 @@ pub fn create_att_expression(
         node_id: format!("n{}", start),
         out_idx: None,
         in_indx: None,
-        contents: c_args.att_idx, // this is the attribute index, USED TO BE PARENT ATTRIBUTE INDEX
-        nbox: c_args.bf_counter,  // inherited from parent
-        att_bf_idx: c_args.att_bf_idx, // This is a reference to the index in parent function
-        box_counter: c_args.box_counter,
+        contents: idx, // this is the attribute index, USED TO BE PARENT ATTRIBUTE INDEX
+        nbox: bf_counter, // inherited from parent
+        att_bf_idx: box_counter as u32, // This is a reference to the index in parent function
     };
     let e1 = Edge {
-        src: c_args.parent_node.node_id.clone(),
+        src: parent_node.node_id.clone(),
         tgt: format!("n{}", start),
         e_type: String::from("Contains"),
-        prop: Some(c_args.att_idx),
+        prop: Some(idx),
     };
     nodes.push(n1.clone());
     edges.push(e1);
-
-    let exp_box = c_args.cur_box.clone();
-
-    if exp_box.metadata.as_ref().is_some() {
-        metadata_idx = exp_box.metadata.unwrap();
+    if ssboxf.metadata.as_ref().is_some() {
+        metadata_idx = ssboxf.metadata.unwrap();
         let mut repeat_meta = false;
         for node in meta_nodes.iter() {
             if node.metadata_idx == metadata_idx {
@@ -2748,20 +3092,19 @@ pub fn create_att_expression(
     // create nodes and edges for this entry, include opo's and opi's
     start += 1;
 
-    let att_box = gromet.modules[0].attributes[(c_args.att_idx - 1) as usize].clone(); // current expression attribute
-    let parent_att_box =
-        gromet.modules[0].attributes[(c_args.parent_node.contents - 1) as usize].clone(); // parent attribute
+    let eboxf = gromet.modules[0].attributes[(idx - 1) as usize].clone(); // current expression attribute
+    let pboxf = gromet.modules[0].attributes[(parent_node.contents - 1) as usize].clone(); // parent attribute
 
     // construct opo nodes, if not none
     // not calling the opo port constuctors since they are based on grabbing in the name from top level of the gromet,
     // not a parental attribute
-    if att_box.opo.is_some() {
+    if eboxf.opo.is_some() {
         // grab name which is one level up and based on indexing
         // this can be done by the parent nodes contents field should give the index of the attributes
         // this constructs a vec for the names in the pofs, if any.
         let mut opo_name: Vec<String> = vec![];
-        for port in parent_att_box.pof.as_ref().unwrap().iter() {
-            if port.r#box == c_args.box_counter as u8 {
+        for port in pboxf.pof.as_ref().unwrap().iter() {
+            if port.r#box == box_counter {
                 if port.name.is_some() {
                     opo_name.push(port.name.as_ref().unwrap().clone());
                 } else {
@@ -2771,7 +3114,7 @@ pub fn create_att_expression(
         }
         if !opo_name.clone().is_empty() {
             let mut oport: u32 = 0;
-            for _op in att_box.opo.as_ref().unwrap().iter() {
+            for _op in eboxf.opo.as_ref().unwrap().iter() {
                 let n2 = Node {
                     n_type: String::from("Opo"),
                     value: None,
@@ -2779,10 +3122,9 @@ pub fn create_att_expression(
                     node_id: format!("n{}", start),
                     out_idx: Some([oport + 1].to_vec()),
                     in_indx: None,
-                    contents: c_args.att_idx,      // current att idx
-                    nbox: c_args.bf_counter,       // top parent top level idx
-                    att_bf_idx: c_args.att_bf_idx, // current box idx of parent
-                    box_counter: c_args.box_counter,
+                    contents: idx,                  // current att idx
+                    nbox: bf_counter,               // top parent top level idx
+                    att_bf_idx: box_counter as u32, // current box idx of parent
                 };
                 nodes.push(n2.clone());
                 // construct edge: expression -> Opo
@@ -2793,15 +3135,13 @@ pub fn create_att_expression(
                     prop: None,
                 };
                 edges.push(e3);
-                if att_box.opo.clone().as_ref().unwrap()[oport as usize]
+                if eboxf.opo.clone().as_ref().unwrap()[oport as usize]
                     .metadata
                     .clone()
                     .as_ref()
                     .is_some()
                 {
-                    metadata_idx = att_box.opo.clone().unwrap()[oport as usize]
-                        .metadata
-                        .unwrap();
+                    metadata_idx = eboxf.opo.clone().unwrap()[oport as usize].metadata.unwrap();
                     let mut repeat_meta = false;
                     for node in meta_nodes.iter() {
                         if node.metadata_idx == metadata_idx {
@@ -2826,10 +3166,10 @@ pub fn create_att_expression(
         }
     }
     // construct opi nodes, in not none
-    if att_box.opi.is_some() {
+    if eboxf.opi.is_some() {
         // grab name which is NOT one level up as in opo
         let mut opi_name: Vec<String> = vec![];
-        for port in att_box.opi.as_ref().unwrap().iter() {
+        for port in eboxf.opi.as_ref().unwrap().iter() {
             if port.name.is_some() {
                 opi_name.push(port.name.as_ref().unwrap().clone());
             } else {
@@ -2837,7 +3177,7 @@ pub fn create_att_expression(
             }
         }
         let mut iport: u32 = 0;
-        for _op in att_box.opi.as_ref().unwrap().iter() {
+        for _op in eboxf.opi.as_ref().unwrap().iter() {
             let n2 = Node {
                 n_type: String::from("Opi"),
                 value: None,
@@ -2845,10 +3185,9 @@ pub fn create_att_expression(
                 node_id: format!("n{}", start),
                 out_idx: None,
                 in_indx: Some([iport + 1].to_vec()),
-                contents: c_args.att_idx,
-                nbox: c_args.bf_counter,
-                att_bf_idx: c_args.att_bf_idx,
-                box_counter: c_args.box_counter,
+                contents: idx,
+                nbox: bf_counter,
+                att_bf_idx: box_counter as u32,
             };
             nodes.push(n2.clone());
             // construct edge: expression -> Opo
@@ -2859,15 +3198,13 @@ pub fn create_att_expression(
                 prop: None,
             };
             edges.push(e3);
-            if att_box.opi.clone().as_ref().unwrap()[iport as usize]
+            if eboxf.opi.clone().as_ref().unwrap()[iport as usize]
                 .metadata
                 .clone()
                 .as_ref()
                 .is_some()
             {
-                metadata_idx = att_box.opi.clone().unwrap()[iport as usize]
-                    .metadata
-                    .unwrap();
+                metadata_idx = eboxf.opi.clone().unwrap()[iport as usize].metadata.unwrap();
                 let mut repeat_meta = false;
                 for node in meta_nodes.iter() {
                     if node.metadata_idx == metadata_idx {
@@ -2890,32 +3227,38 @@ pub fn create_att_expression(
         }
     }
     // now to construct the nodes inside the expression, Literal and Primitives
-    let mut new_c_args = c_args.clone();
-    let mut box_counter: usize = 1;
-    if att_box.bf.is_some() {
-        for att_sub_box in att_box.bf.as_ref().unwrap().iter() {
-            new_c_args.box_counter = box_counter;
-            new_c_args.cur_box = att_sub_box.clone();
-            new_c_args.att_idx = att_sub_box.contents.unwrap() as usize;
-            match att_sub_box.function_type {
+    let mut box_counter: u8 = 1;
+    if eboxf.bf.is_some() {
+        for sboxf in eboxf.bf.as_ref().unwrap().iter() {
+            match sboxf.function_type {
                 FunctionType::Literal => {
-                    create_att_literal(
-                        &gromet, // gromet for metadata
-                        nodes,   // nodes
-                        edges,
-                        meta_nodes,
+                    (nodes, edges, meta_nodes) = create_att_literal(
+                        &gromet.clone(),
+                        eboxf.clone(),
+                        sboxf.clone(),
+                        nodes.clone(),
+                        edges.clone(),
+                        n1.clone(),
+                        idx,
+                        box_counter,
+                        bf_counter,
                         start,
-                        new_c_args.clone(),
+                        meta_nodes.clone(),
                     );
                 }
                 FunctionType::Primitive => {
-                    create_att_primitive(
-                        &gromet, // gromet for metadata
-                        nodes,   // nodes
-                        edges,
-                        meta_nodes,
+                    (nodes, edges, meta_nodes) = create_att_primitive(
+                        &gromet.clone(),
+                        eboxf.clone(),
+                        sboxf.clone(),
+                        nodes.clone(),
+                        edges.clone(),
+                        n1.clone(),
+                        idx,
+                        box_counter,
+                        bf_counter,
                         start,
-                        new_c_args.clone(),
+                        meta_nodes.clone(),
                     );
                 }
                 _ => {}
@@ -2925,31 +3268,34 @@ pub fn create_att_expression(
         }
     }
     // Now we perform the internal wiring of this branch
-    internal_wiring(
-        att_box.clone(),
-        nodes,
-        edges,
-        c_args.att_idx,
-        c_args.bf_counter,
-    );
+    edges = internal_wiring(eboxf.clone(), nodes.clone(), edges, idx, bf_counter);
 
     // Now we also perform wopio wiring in case there is an empty expression
-    if att_box.wopio.is_some() {
-        wopio_wiring(att_box, nodes, edges, c_args.att_idx - 1, c_args.bf_counter);
+    if eboxf.wopio.is_some() {
+        edges = wopio_wiring(eboxf, nodes.clone(), edges, idx - 1, bf_counter);
     }
+
+    (nodes, edges, start, meta_nodes)
 }
 
 // This needs to be updated to handle the new node structure and remove the overloaded contents field which will mess with the wiring alot
 #[allow(unused_assignments)]
 pub fn create_att_predicate(
-    gromet: &ModuleCollection, // needed still for metadata unfortunately
-    nodes: &mut Vec<Node>,
-    edges: &mut Vec<Edge>,
-    meta_nodes: &mut Vec<MetadataNode>,
-    mut start: u32, // for node and edge indexing
-    c_args: ConstructorArgs,
-) -> () {
+    gromet: &ModuleCollection,
+    _eeboxf: FunctionNet,
+    ssboxf: GrometBox,
+    mut nodes: Vec<Node>,
+    mut edges: Vec<Edge>,
+    parent_node: Node,
+    _idx_in: u32,
+    box_counter: u8,
+    bf_counter: u8,
+    mut start: u32,
+    mut meta_nodes: Vec<MetadataNode>,
+) -> (Vec<Node>, Vec<Edge>, u32, Vec<MetadataNode>) {
     let mut metadata_idx = 0;
+
+    let idx = ssboxf.contents.unwrap(); // this is the attribute index of the current expression
 
     let n1 = Node {
         n_type: String::from("Predicate"),
@@ -2958,24 +3304,20 @@ pub fn create_att_predicate(
         node_id: format!("n{}", start),
         out_idx: None,
         in_indx: None,
-        contents: c_args.att_idx,      // this is the attribute index
-        nbox: c_args.bf_counter,       // inherited from parent
-        att_bf_idx: c_args.att_bf_idx, // This is a reference to the index in parent function
-        box_counter: c_args.box_counter,
+        contents: idx, // this is the attribute index, USED TO BE PARENT ATTRIBUTE INDEX
+        nbox: bf_counter, // inherited from parent
+        att_bf_idx: box_counter as u32, // This is a reference to the index in parent function
     };
     let e1 = Edge {
-        src: c_args.parent_node.node_id.clone(),
+        src: parent_node.node_id.clone(),
         tgt: format!("n{}", start),
         e_type: String::from("Contains"),
-        prop: Some(c_args.att_idx),
+        prop: Some(idx),
     };
     nodes.push(n1.clone());
     edges.push(e1);
-
-    let in_att_box = c_args.cur_box.clone();
-
-    if in_att_box.metadata.as_ref().is_some() {
-        metadata_idx = in_att_box.metadata.unwrap();
+    if ssboxf.metadata.as_ref().is_some() {
+        metadata_idx = ssboxf.metadata.unwrap();
         let mut repeat_meta = false;
         for node in meta_nodes.iter() {
             if node.metadata_idx == metadata_idx {
@@ -2998,20 +3340,19 @@ pub fn create_att_predicate(
     // create nodes and edges for this entry, include opo's and opi's
     start += 1;
 
-    let att_box = gromet.modules[0].attributes[(c_args.att_idx - 1) as usize].clone(); // current expression attribute
-    let parent_att_box =
-        gromet.modules[0].attributes[(c_args.parent_node.contents - 1) as usize].clone(); // parent attribute
+    let eboxf = gromet.modules[0].attributes[(idx - 1) as usize].clone(); // current expression attribute
+    let pboxf = gromet.modules[0].attributes[(parent_node.contents - 1) as usize].clone(); // parent attribute
 
     // construct opo nodes, if not none
     // not calling the opo port constuctors since they are based on grabbing in the name from top level of the gromet,
     // not a parental attribute
-    if att_box.opo.is_some() {
+    if eboxf.opo.is_some() {
         // grab name which is one level up and based on indexing
         // this can be done by the parent nodes contents field should give the index of the attributes
         // this constructs a vec for the names in the pofs, if any.
         let mut opo_name: Vec<String> = vec![];
-        for port in parent_att_box.pof.as_ref().unwrap().iter() {
-            if port.r#box == c_args.box_counter as u8 {
+        for port in pboxf.pof.as_ref().unwrap().iter() {
+            if port.r#box == box_counter {
                 if port.name.is_some() {
                     opo_name.push(port.name.as_ref().unwrap().clone());
                 } else {
@@ -3020,7 +3361,7 @@ pub fn create_att_predicate(
             }
         }
         let mut oport: u32 = 0;
-        for _op in att_box.opo.as_ref().unwrap().iter() {
+        for _op in eboxf.opo.as_ref().unwrap().iter() {
             let n2 = Node {
                 n_type: String::from("Opo"),
                 value: None,
@@ -3028,10 +3369,9 @@ pub fn create_att_predicate(
                 node_id: format!("n{}", start),
                 out_idx: Some([oport + 1].to_vec()),
                 in_indx: None,
-                contents: c_args.att_idx,      // this is the attribute index
-                nbox: c_args.bf_counter,       // inherited from parent
-                att_bf_idx: c_args.att_bf_idx, // This is a reference to the index in parent function
-                box_counter: c_args.box_counter, // current box idx of parent
+                contents: idx,                  // current att idx
+                nbox: bf_counter,               // top parent top level idx
+                att_bf_idx: box_counter as u32, // current box idx of parent
             };
             nodes.push(n2.clone());
             // construct edge: expression -> Opo
@@ -3042,15 +3382,13 @@ pub fn create_att_predicate(
                 prop: None,
             };
             edges.push(e3);
-            if att_box.opo.clone().as_ref().unwrap()[oport as usize]
+            if eboxf.opo.clone().as_ref().unwrap()[oport as usize]
                 .metadata
                 .clone()
                 .as_ref()
                 .is_some()
             {
-                metadata_idx = att_box.opo.clone().unwrap()[oport as usize]
-                    .metadata
-                    .unwrap();
+                metadata_idx = eboxf.opo.clone().unwrap()[oport as usize].metadata.unwrap();
                 let mut repeat_meta = false;
                 for node in meta_nodes.iter() {
                     if node.metadata_idx == metadata_idx {
@@ -3074,10 +3412,10 @@ pub fn create_att_predicate(
         }
     }
     // construct opi nodes, in not none
-    if att_box.opi.is_some() {
+    if eboxf.opi.is_some() {
         // grab name which is NOT one level up as in opo
         let mut opi_name: Vec<String> = vec![];
-        for port in att_box.opi.as_ref().unwrap().iter() {
+        for port in eboxf.opi.as_ref().unwrap().iter() {
             if port.name.is_some() {
                 opi_name.push(port.name.as_ref().unwrap().clone());
             } else {
@@ -3085,7 +3423,7 @@ pub fn create_att_predicate(
             }
         }
         let mut iport: u32 = 0;
-        for _op in att_box.opi.as_ref().unwrap().iter() {
+        for _op in eboxf.opi.as_ref().unwrap().iter() {
             let n2 = Node {
                 n_type: String::from("Opi"),
                 value: None,
@@ -3093,10 +3431,9 @@ pub fn create_att_predicate(
                 node_id: format!("n{}", start),
                 out_idx: None,
                 in_indx: Some([iport + 1].to_vec()),
-                contents: c_args.att_idx,      // this is the attribute index
-                nbox: c_args.bf_counter,       // inherited from parent
-                att_bf_idx: c_args.att_bf_idx, // This is a reference to the index in parent function
-                box_counter: c_args.box_counter,
+                contents: idx,
+                nbox: bf_counter,
+                att_bf_idx: box_counter as u32,
             };
             nodes.push(n2.clone());
             // construct edge: expression -> Opo
@@ -3107,15 +3444,13 @@ pub fn create_att_predicate(
                 prop: None,
             };
             edges.push(e3);
-            if att_box.opi.clone().as_ref().unwrap()[iport as usize]
+            if eboxf.opi.clone().as_ref().unwrap()[iport as usize]
                 .metadata
                 .clone()
                 .as_ref()
                 .is_some()
             {
-                metadata_idx = att_box.opi.clone().unwrap()[iport as usize]
-                    .metadata
-                    .unwrap();
+                metadata_idx = eboxf.opi.clone().unwrap()[iport as usize].metadata.unwrap();
                 let mut repeat_meta = false;
                 for node in meta_nodes.iter() {
                     if node.metadata_idx == metadata_idx {
@@ -3138,31 +3473,37 @@ pub fn create_att_predicate(
         }
     }
     // now to construct the nodes inside the expression, Literal and Primitives
-    let mut new_c_args = c_args.clone();
-    let mut box_counter: usize = 1;
-    for sub_att_box in att_box.bf.as_ref().unwrap().iter() {
-        new_c_args.box_counter = box_counter;
-        new_c_args.cur_box = sub_att_box.clone();
-        new_c_args.att_idx = sub_att_box.contents.unwrap() as usize;
-        match sub_att_box.function_type {
+    let mut box_counter: u8 = 1;
+    for sboxf in eboxf.bf.as_ref().unwrap().iter() {
+        match sboxf.function_type {
             FunctionType::Literal => {
-                create_att_literal(
-                    &gromet, // gromet for metadata
-                    nodes,   // nodes
-                    edges,
-                    meta_nodes,
+                (nodes, edges, meta_nodes) = create_att_literal(
+                    &gromet.clone(),
+                    eboxf.clone(),
+                    sboxf.clone(),
+                    nodes.clone(),
+                    edges.clone(),
+                    n1.clone(),
+                    idx,
+                    box_counter,
+                    bf_counter,
                     start,
-                    new_c_args.clone(),
+                    meta_nodes.clone(),
                 );
             }
             FunctionType::Primitive => {
-                create_att_primitive(
-                    &gromet, // gromet for metadata
-                    nodes,   // nodes
-                    edges,
-                    meta_nodes,
+                (nodes, edges, meta_nodes) = create_att_primitive(
+                    &gromet.clone(),
+                    eboxf.clone(),
+                    sboxf.clone(),
+                    nodes.clone(),
+                    edges.clone(),
+                    n1.clone(),
+                    idx,
+                    box_counter,
+                    bf_counter,
                     start,
-                    new_c_args.clone(),
+                    meta_nodes.clone(),
                 );
             }
             _ => {}
@@ -3171,32 +3512,31 @@ pub fn create_att_predicate(
         start += 1;
     }
     // Now we perform the internal wiring of this branch
-    internal_wiring(
-        att_box,
-        &mut nodes.clone(),
-        edges,
-        c_args.att_idx,
-        c_args.bf_counter,
-    );
+    edges = internal_wiring(eboxf, nodes.clone(), edges, idx, bf_counter);
+    (nodes, edges, start, meta_nodes)
 }
 #[allow(unused_assignments)]
 pub fn create_att_literal(
-    gromet: &ModuleCollection, // needed still for metadata unfortunately
-    nodes: &mut Vec<Node>,
-    edges: &mut Vec<Edge>,
-    meta_nodes: &mut Vec<MetadataNode>,
-    mut start: u32, // for node and edge indexing
-    c_args: ConstructorArgs,
-) -> () {
-    let att_box = c_args.att_box;
-    let lit_box = c_args.cur_box;
+    gromet: &ModuleCollection,
+    eboxf: FunctionNet,
+    sboxf: GrometBox,
+    mut nodes: Vec<Node>,
+    mut edges: Vec<Edge>,
+    parent_node: Node,
+    idx: u32,
+    box_counter: u8,
+    bf_counter: u8,
+    start: u32,
+    mut meta_nodes: Vec<MetadataNode>,
+) -> (Vec<Node>, Vec<Edge>, Vec<MetadataNode>) {
     // first find the pof's for box
     let mut pof: Vec<u32> = vec![];
-    if att_box.pof.is_some() {
+    if eboxf.pof.is_some() {
         let mut po_idx: u32 = 1;
-        for port in att_box.pof.unwrap().iter() {
-            if port.r#box == c_args.box_counter as u8 {
+        for port in eboxf.pof.unwrap().iter() {
+            if port.r#box == box_counter {
                 pof.push(po_idx);
+            } else {
             }
             po_idx += 1;
         }
@@ -3205,27 +3545,26 @@ pub fn create_att_literal(
     let mut metadata_idx = 0;
     let n3 = Node {
         n_type: String::from("Literal"),
-        value: Some(format!("{:?}", lit_box.value.as_ref().unwrap())),
+        value: Some(format!("{:?}", sboxf.value.as_ref().unwrap())),
         name: None,
         node_id: format!("n{}", start),
         out_idx: Some(pof),
         in_indx: None, // literals should only have out ports
-        contents: c_args.att_idx,
-        nbox: c_args.bf_counter,
-        att_bf_idx: c_args.att_bf_idx,
-        box_counter: c_args.box_counter,
+        contents: idx,
+        nbox: bf_counter,
+        att_bf_idx: box_counter as u32,
     };
     nodes.push(n3.clone());
     // make edge connecting to expression
     let e4 = Edge {
-        src: c_args.parent_node.node_id,
+        src: parent_node.node_id,
         tgt: n3.node_id.clone(),
         e_type: String::from("Contains"),
         prop: None,
     };
     edges.push(e4);
-    if lit_box.metadata.is_some() {
-        metadata_idx = lit_box.metadata.unwrap();
+    if sboxf.metadata.is_some() {
+        metadata_idx = sboxf.metadata.unwrap();
         let mut repeat_meta = false;
         for node in meta_nodes.iter() {
             if node.metadata_idx == metadata_idx {
@@ -3243,23 +3582,28 @@ pub fn create_att_literal(
             edges.push(me1);
         }
     }
-    start += 1;
+    (nodes, edges, meta_nodes)
 }
 #[allow(unused_assignments)]
 pub fn create_att_primitive(
-    gromet: &ModuleCollection, // needed still for metadata unfortunately
-    nodes: &mut Vec<Node>,
-    edges: &mut Vec<Edge>,
-    meta_nodes: &mut Vec<MetadataNode>,
-    mut start: u32, // for node and edge indexing
-    c_args: ConstructorArgs,
-) -> () {
+    gromet: &ModuleCollection,
+    eboxf: FunctionNet,
+    sboxf: GrometBox,
+    mut nodes: Vec<Node>,
+    mut edges: Vec<Edge>,
+    parent_node: Node,
+    idx: u32,
+    box_counter: u8,
+    bf_counter: u8,
+    start: u32,
+    mut meta_nodes: Vec<MetadataNode>,
+) -> (Vec<Node>, Vec<Edge>, Vec<MetadataNode>) {
     // first find the pof's for box
     let mut pof: Vec<u32> = vec![];
-    if c_args.att_box.pof.is_some() {
+    if eboxf.pof.is_some() {
         let mut po_idx: u32 = 1;
-        for port in c_args.att_box.pof.clone().unwrap().iter() {
-            if port.r#box == c_args.box_counter as u8 {
+        for port in eboxf.pof.clone().unwrap().iter() {
+            if port.r#box == box_counter {
                 pof.push(po_idx);
             }
             po_idx += 1;
@@ -3267,10 +3611,10 @@ pub fn create_att_primitive(
     }
     // then find pif's for box
     let mut pif: Vec<u32> = vec![];
-    if c_args.att_box.pif.is_some() {
+    if eboxf.pif.is_some() {
         let mut pi_idx: u32 = 1;
-        for port in c_args.att_box.pif.unwrap().iter() {
-            if port.r#box == c_args.box_counter as u8 {
+        for port in eboxf.pif.unwrap().iter() {
+            if port.r#box == box_counter {
                 pif.push(pi_idx);
             }
             pi_idx += 1;
@@ -3281,26 +3625,25 @@ pub fn create_att_primitive(
     let n3 = Node {
         n_type: String::from("Primitive"),
         value: None,
-        name: c_args.cur_box.name.clone(),
+        name: sboxf.name.clone(),
         node_id: format!("n{}", start),
         out_idx: Some(pof),
         in_indx: Some(pif),
-        contents: c_args.att_idx,
-        nbox: c_args.bf_counter,
-        att_bf_idx: c_args.att_bf_idx,
-        box_counter: c_args.box_counter,
+        contents: idx,
+        nbox: bf_counter,
+        att_bf_idx: box_counter as u32,
     };
     nodes.push(n3.clone());
     // make edge connecting to expression
     let e4 = Edge {
-        src: c_args.parent_node.node_id,
+        src: parent_node.node_id,
         tgt: n3.node_id.clone(),
         e_type: String::from("Contains"),
         prop: None,
     };
     edges.push(e4);
-    if c_args.cur_box.metadata.is_some() {
-        metadata_idx = c_args.cur_box.metadata.unwrap();
+    if sboxf.metadata.is_some() {
+        metadata_idx = sboxf.metadata.unwrap();
         let mut repeat_meta = false;
         for node in meta_nodes.iter() {
             if node.metadata_idx == metadata_idx {
@@ -3318,28 +3661,29 @@ pub fn create_att_primitive(
             edges.push(me1);
         }
     }
-    start += 1;
+    (nodes, edges, meta_nodes)
 }
 // This is for the construction of Opo's
 #[allow(unused_assignments)]
 pub fn create_opo(
-    gromet: &ModuleCollection, // needed still for metadata unfortunately
-    nodes: &mut Vec<Node>,
-    edges: &mut Vec<Edge>,
-    meta_nodes: &mut Vec<MetadataNode>,
-    mut start: u32, // for node and edge indexing
-    c_args: ConstructorArgs,
-) -> () {
-    let att_box = gromet.modules[0].attributes[(c_args.parent_node.contents - 1) as usize].clone();
+    mut nodes: Vec<Node>,
+    mut edges: Vec<Edge>,
+    mut meta_nodes: Vec<MetadataNode>,
+    gromet: &ModuleCollection,
+    mut start: u32,
+    parent_node: Node,
+    box_counter: u8,
+) -> (Vec<Node>, Vec<Edge>, Vec<MetadataNode>, u32) {
+    let eboxf = gromet.modules[0].attributes[(parent_node.contents - 1) as usize].clone();
     // construct opo nodes, if not none
-    if att_box.opo.is_some() {
+    if eboxf.opo.is_some() {
         // grab name which is one level up and based on indexing
         let mut opo_name = "un-named";
         let mut oport: u32 = 0;
-        for op in att_box.opo.as_ref().unwrap().iter() {
+        for op in eboxf.opo.as_ref().unwrap().iter() {
             if op.name.as_ref().is_none() && gromet.modules[0].r#fn.pof.as_ref().is_some() {
                 for port in gromet.modules[0].r#fn.pof.as_ref().unwrap().iter() {
-                    if port.r#box == c_args.bf_counter as u8
+                    if port.r#box == box_counter
                         && oport == (port.id.unwrap() as u32 - 1)
                         && port.name.is_some()
                     {
@@ -3356,15 +3700,14 @@ pub fn create_opo(
                 node_id: format!("n{}", start),
                 out_idx: Some([oport + 1].to_vec()),
                 in_indx: None,
-                contents: c_args.parent_node.contents,
-                nbox: c_args.bf_counter,
-                att_bf_idx: c_args.parent_node.att_bf_idx,
-                box_counter: c_args.parent_node.box_counter,
+                contents: parent_node.contents,
+                nbox: box_counter,
+                att_bf_idx: parent_node.att_bf_idx,
             };
             nodes.push(n2.clone());
             // construct edge: expression -> Opo
             let e3 = Edge {
-                src: c_args.parent_node.node_id.clone(),
+                src: parent_node.node_id.clone(),
                 tgt: n2.node_id.clone(),
                 e_type: String::from("Port_Of"),
                 prop: None,
@@ -3372,15 +3715,13 @@ pub fn create_opo(
             edges.push(e3);
 
             let mut metadata_idx = 0;
-            if att_box.opo.clone().as_ref().unwrap()[oport as usize]
+            if eboxf.opo.clone().as_ref().unwrap()[oport as usize]
                 .metadata
                 .clone()
                 .as_ref()
                 .is_some()
             {
-                metadata_idx = att_box.opo.clone().unwrap()[oport as usize]
-                    .metadata
-                    .unwrap();
+                metadata_idx = eboxf.opo.clone().unwrap()[oport as usize].metadata.unwrap();
                 let mut repeat_meta = false;
                 for node in meta_nodes.iter() {
                     if node.metadata_idx == metadata_idx {
@@ -3403,28 +3744,30 @@ pub fn create_opo(
             oport += 1;
         }
     }
+    (nodes, edges, meta_nodes, start)
 }
 
 // This is for the construction of Opi's
 #[allow(unused_assignments)]
 pub fn create_opi(
-    gromet: &ModuleCollection, // needed still for metadata unfortunately
-    nodes: &mut Vec<Node>,
-    edges: &mut Vec<Edge>,
-    meta_nodes: &mut Vec<MetadataNode>,
-    mut start: u32, // for node and edge indexing
-    c_args: ConstructorArgs,
-) -> () {
-    let att_box = gromet.modules[0].attributes[(c_args.parent_node.contents - 1) as usize].clone();
+    mut nodes: Vec<Node>,
+    mut edges: Vec<Edge>,
+    mut meta_nodes: Vec<MetadataNode>,
+    gromet: &ModuleCollection,
+    mut start: u32,
+    parent_node: Node,
+    box_counter: u8,
+) -> (Vec<Node>, Vec<Edge>, Vec<MetadataNode>, u32) {
+    let eboxf = gromet.modules[0].attributes[(parent_node.contents - 1) as usize].clone();
     // construct opi nodes, if not none
-    if att_box.opi.is_some() {
+    if eboxf.opi.is_some() {
         // grab name which is one level up and based on indexing
         let mut opi_name = "un-named";
         let mut oport: u32 = 0;
-        for op in att_box.opi.as_ref().unwrap().iter() {
+        for op in eboxf.opi.as_ref().unwrap().iter() {
             if op.name.as_ref().is_none() && gromet.modules[0].r#fn.pif.as_ref().is_some() {
                 for port in gromet.modules[0].r#fn.pif.as_ref().unwrap().iter() {
-                    if port.r#box == c_args.bf_counter as u8
+                    if port.r#box == box_counter
                         && oport == (port.id.unwrap() as u32 - 1)
                         && port.name.is_some()
                     {
@@ -3441,15 +3784,14 @@ pub fn create_opi(
                 node_id: format!("n{}", start),
                 out_idx: None,
                 in_indx: Some([oport + 1].to_vec()),
-                contents: c_args.parent_node.contents,
-                nbox: c_args.bf_counter,
-                att_bf_idx: c_args.parent_node.att_bf_idx,
-                box_counter: c_args.parent_node.box_counter,
+                contents: parent_node.contents,
+                nbox: box_counter,
+                att_bf_idx: parent_node.att_bf_idx,
             };
             nodes.push(n2.clone());
             // construct edge: expression <- Opi
             let e3 = Edge {
-                src: c_args.parent_node.node_id.clone(),
+                src: parent_node.node_id.clone(),
                 tgt: n2.node_id.clone(),
                 e_type: String::from("Port_Of"),
                 prop: None,
@@ -3457,15 +3799,13 @@ pub fn create_opi(
             edges.push(e3);
 
             let mut metadata_idx = 0;
-            if att_box.opi.clone().as_ref().unwrap()[oport as usize]
+            if eboxf.opi.clone().as_ref().unwrap()[oport as usize]
                 .metadata
                 .clone()
                 .as_ref()
                 .is_some()
             {
-                metadata_idx = att_box.opi.clone().unwrap()[oport as usize]
-                    .metadata
-                    .unwrap();
+                metadata_idx = eboxf.opi.clone().unwrap()[oport as usize].metadata.unwrap();
                 let mut repeat_meta = false;
                 for node in meta_nodes.iter() {
                     if node.metadata_idx == metadata_idx {
@@ -3488,16 +3828,17 @@ pub fn create_opi(
             oport += 1;
         }
     }
+    (nodes, edges, meta_nodes, start)
 }
 // having issues with deeply nested structure, it is breaking in the internal wiring of the function level.
 // wfopi: pif -> opi
 pub fn wfopi_wiring(
     eboxf: FunctionNet,
-    nodes: &mut Vec<Node>,
-    edges: &mut Vec<Edge>,
+    nodes: Vec<Node>,
+    mut edges: Vec<Edge>,
     idx: u32,
     bf_counter: u8,
-) -> () {
+) -> Vec<Edge> {
     // iterate through all wires of type
     for wire in eboxf.wfopi.unwrap().iter() {
         let mut prop = None;
@@ -3505,9 +3846,9 @@ pub fn wfopi_wiring(
         // find the src node
         for node in nodes.iter() {
             // make sure in correct box
-            if bf_counter == node.nbox as u8 {
+            if bf_counter == node.nbox {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if (idx) == node.contents as u32 {
+                if (idx) == node.contents {
                     // only include nodes with pifs
                     if node.in_indx.is_some() {
                         // exclude opi's
@@ -3528,9 +3869,9 @@ pub fn wfopi_wiring(
         // find the tgt node
         for node in nodes.iter() {
             // make sure in correct box
-            if bf_counter == node.nbox as u8 {
+            if bf_counter == node.nbox {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if (idx) == node.contents as u32 {
+                if (idx) == node.contents {
                     // only opi's
                     if node.n_type == "Opi" {
                         // iterate through port to check for tgt
@@ -3549,29 +3890,30 @@ pub fn wfopi_wiring(
                 src: wfopi_src_tgt[0].clone(),
                 tgt: wfopi_src_tgt[1].clone(),
                 e_type: String::from("Wire"),
-                prop: Some(prop.unwrap() as usize),
+                prop,
             };
             edges.push(e6);
         }
     }
+    edges
 }
 
 pub fn wfopo_wiring(
     eboxf: FunctionNet,
-    nodes: &mut Vec<Node>,
-    edges: &mut Vec<Edge>,
+    nodes: Vec<Node>,
+    mut edges: Vec<Edge>,
     idx: u32,
     bf_counter: u8,
-) -> () {
+) -> Vec<Edge> {
     // iterate through all wires of type
     for wire in eboxf.wfopo.unwrap().iter() {
         let mut wfopo_src_tgt: Vec<String> = vec![];
         // find the src node
         for node in nodes.iter() {
             // make sure in correct box
-            if bf_counter == node.nbox as u8 {
+            if bf_counter == node.nbox {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if (idx) == node.contents as u32 {
+                if (idx) == node.contents {
                     // only opo's
                     if node.n_type == "Opo" {
                         // iterate through port to check for tgt
@@ -3588,9 +3930,9 @@ pub fn wfopo_wiring(
         // finding the tgt node
         for node in nodes.iter() {
             // make sure in correct box
-            if bf_counter == node.nbox as u8 {
+            if bf_counter == node.nbox {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if (idx) == node.contents as u32 {
+                if (idx) == node.contents {
                     // only include nodes with pofs
                     if node.out_idx.is_some() {
                         // exclude opo's
@@ -3617,17 +3959,18 @@ pub fn wfopo_wiring(
             edges.push(e7);
         }
     }
+    edges
 }
 // this is duplicating wires a ton. (01/15/23)
 // shouldn't use bf_counter, should use att_bf_idx since that is variable for functions, need to pull the box from the
 // ports the wires point to.
 pub fn wff_wiring(
     eboxf: FunctionNet,
-    nodes: &mut Vec<Node>,
-    edges: &mut Vec<Edge>,
+    nodes: Vec<Node>,
+    mut edges: Vec<Edge>,
     idx: u32,
     bf_counter: u8,
-) -> () {
+) -> Vec<Edge> {
     // iterate through all wires of type
     for wire in eboxf.wff.unwrap().iter() {
         let mut wff_src_tgt: Vec<String> = vec![];
@@ -3652,11 +3995,11 @@ pub fn wff_wiring(
         // find the src node
         for node in nodes.iter() {
             // make sure in correct box
-            if src_nbox == node.nbox as u8 {
+            if src_nbox == node.nbox {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if src_att == node.contents as u32 {
+                if src_att == node.contents {
                     // matche the box
-                    if (src_box as u32) == node.att_bf_idx as u32 {
+                    if (src_box as u32) == node.att_bf_idx {
                         // only include nodes with pifs
                         if node.in_indx.is_some() {
                             // exclude opo's
@@ -3678,11 +4021,11 @@ pub fn wff_wiring(
         // finding the tgt node
         for node in nodes.iter() {
             // make sure in correct box
-            if tgt_nbox == node.nbox as u8 {
+            if tgt_nbox == node.nbox {
                 // make sure only looking in current attribute nodes for srcs and tgts
-                if tgt_att == node.contents as u32 {
+                if tgt_att == node.contents {
                     // match internal box
-                    if (tgt_box as u32) == node.att_bf_idx as u32 {
+                    if (tgt_box as u32) == node.att_bf_idx {
                         // only include nodes with pofs
                         if node.out_idx.is_some() {
                             // exclude opo's
@@ -3705,20 +4048,21 @@ pub fn wff_wiring(
                 src: wff_src_tgt[0].clone(),
                 tgt: wff_src_tgt[1].clone(),
                 e_type: String::from("Wire"),
-                prop: Some(prop.unwrap() as usize),
+                prop,
             };
             edges.push(e8);
         }
     }
+    edges
 }
 
 pub fn wopio_wiring(
     eboxf: FunctionNet,
-    nodes: &mut Vec<Node>,
-    edges: &mut Vec<Edge>,
-    idx: usize,
-    bf_counter: usize,
-) -> () {
+    nodes: Vec<Node>,
+    mut edges: Vec<Edge>,
+    idx: u32,
+    bf_counter: u8,
+) -> Vec<Edge> {
     // iterate through all wires of type
     for wire in eboxf.wopio.unwrap().iter() {
         let mut wopio_src_tgt: Vec<String> = vec![];
@@ -3773,15 +4117,16 @@ pub fn wopio_wiring(
             edges.push(e7);
         }
     }
+    edges
 }
 
 pub fn internal_wiring(
     eboxf: FunctionNet,
-    nodes: &mut Vec<Node>,
-    edges: &mut Vec<Edge>,
-    idx: usize,
-    bf_counter: usize,
-) -> () {
+    nodes: Vec<Node>,
+    mut edges: Vec<Edge>,
+    idx: u32,
+    bf_counter: u8,
+) -> Vec<Edge> {
     // first lets wire the wfopi, note we need to first limit ourselves
     // to only nodes in the current attribute by checking the contents field
     // and then we run find the ports that match the wire src and tgt.
@@ -3792,53 +4137,38 @@ pub fn internal_wiring(
 
     // check if wire exists, wfopi
     if eboxf.wfopi.is_some() {
-        wfopi_wiring(
-            eboxf.clone(),
-            &mut nodes.clone(),
-            edges,
-            idx as u32,
-            bf_counter as u8,
-        );
+        edges = wfopi_wiring(eboxf.clone(), nodes.clone(), edges, idx, bf_counter);
     }
 
     // check if wire exists, wfopo
     if eboxf.wfopo.is_some() {
-        wfopo_wiring(
-            eboxf.clone(),
-            &mut nodes.clone(),
-            edges,
-            idx as u32,
-            bf_counter as u8,
-        );
+        edges = wfopo_wiring(eboxf.clone(), nodes.clone(), edges, idx, bf_counter);
     }
 
     // check if wire exists, wff
     if eboxf.wff.is_some() {
-        wff_wiring(
-            eboxf.clone(),
-            &mut nodes.clone(),
-            edges,
-            idx as u32,
-            bf_counter as u8,
-        );
+        edges = wff_wiring(eboxf.clone(), nodes.clone(), edges, idx, bf_counter);
     }
 
     // check if wire exists, wopio
     if eboxf.wopio.is_some() {
-        wopio_wiring(eboxf, nodes, edges, idx, bf_counter);
+        edges = wopio_wiring(eboxf, nodes, edges, idx, bf_counter);
     }
+
+    edges
 }
 
 // now for the wiring used for imports
 // needs to handle top level and function level wiring that uses the function net at the call of the import.
 pub fn import_wiring(
     gromet: &ModuleCollection,
-    nodes: &mut Vec<Node>,
-    edges: &mut Vec<Edge>,
-    idx: usize,
-    bf_counter: usize,
+    _eboxf: FunctionNet,
+    nodes: Vec<Node>,
+    mut edges: Vec<Edge>,
+    idx: u32,
+    bf_counter: u8,
     parent_node: Node,
-) -> () {
+) -> Vec<Edge> {
     // first based on the parent_node determine if we need to grab the pof's from the top scope or a sub scope
     if parent_node.att_bf_idx == 0 {
         // this means top level wiring
@@ -3870,7 +4200,7 @@ pub fn import_wiring(
                     // make sure only looking in current attribute nodes for srcs and tgts
                     if src_att == node.contents {
                         // matche the box
-                        if (src_box as u32) == node.att_bf_idx as u32 {
+                        if (src_box as u32) == node.att_bf_idx {
                             // only include nodes with pifs
                             if node.in_indx.is_some() {
                                 // exclude opo's
@@ -3895,7 +4225,7 @@ pub fn import_wiring(
                     // make sure only looking in current attribute nodes for srcs and tgts
                     if tgt_att == node.contents {
                         // match internal box
-                        if (tgt_box as u32) == node.att_bf_idx as u32 {
+                        if (tgt_box as u32) == node.att_bf_idx {
                             // only include nodes with pofs
                             if node.out_idx.is_some() {
                                 // exclude opo's
@@ -3923,6 +4253,7 @@ pub fn import_wiring(
                 edges.push(e8);
             }
         }
+        edges
     } else {
         // this means we are in function scope, concerns on if this is cross attributal or just internal wiring...
         let eboxf = gromet.modules[0].attributes[(parent_node.contents - 1) as usize].clone();
@@ -3954,7 +4285,7 @@ pub fn import_wiring(
                     // make sure only looking in current attribute nodes for srcs and tgts
                     if src_att == node.contents {
                         // matche the box
-                        if (src_box as u32) == node.att_bf_idx as u32 {
+                        if (src_box as u32) == node.att_bf_idx {
                             // only include nodes with pifs
                             if node.in_indx.is_some() {
                                 // exclude opo's
@@ -3979,7 +4310,7 @@ pub fn import_wiring(
                     // make sure only looking in current attribute nodes for srcs and tgts
                     if tgt_att == node.contents {
                         // match internal box
-                        if (tgt_box as u32) == node.att_bf_idx as u32 {
+                        if (tgt_box as u32) == node.att_bf_idx {
                             // only include nodes with pofs
                             if node.out_idx.is_some() {
                                 // exclude opo's
@@ -4007,16 +4338,17 @@ pub fn import_wiring(
                 edges.push(e8);
             }
         }
+        edges
     }
 }
 
 pub fn cross_att_wiring(
     eboxf: FunctionNet, // This is the current attribute
-    nodes: &mut Vec<Node>,
-    edges: &mut Vec<Edge>,
-    idx: usize,        // this +1 is the current attribute index
-    bf_counter: usize, // this is the current box
-) -> () {
+    mut nodes: Vec<Node>,
+    mut edges: Vec<Edge>,
+    idx: u32,       // this +1 is the current attribute index
+    bf_counter: u8, // this is the current box
+) -> (Vec<Edge>, Vec<Node>) {
     // wire id corresponds to subport index in list so ex: wff src.id="1" means the first opi in the list of the src.box, this is the in_idx in the opi or out_indx in opo.
     // This will have to run wfopo wfopi and wff all in order to get the cross attribual wiring that can exist in all of them, refactoring won't do much in code saving space though.
     // for cross attributal wiring they will construct the following types of wires
@@ -4026,46 +4358,29 @@ pub fn cross_att_wiring(
 
     // check if wire exists, wfopi
     if eboxf.wfopi.is_some() {
-        wfopi_cross_att_wiring(
-            eboxf.clone(),
-            &mut nodes.clone(),
-            edges,
-            idx as u32,
-            bf_counter as u8,
-        );
+        edges = wfopi_cross_att_wiring(eboxf.clone(), nodes.clone(), edges, idx, bf_counter);
     }
 
     // check if wire exists, wfopo
     if eboxf.wfopo.is_some() {
-        wfopo_cross_att_wiring(
-            eboxf.clone(),
-            &mut nodes.clone(),
-            edges,
-            idx as u32,
-            bf_counter as u8,
-        );
+        edges = wfopo_cross_att_wiring(eboxf.clone(), nodes.clone(), edges, idx, bf_counter);
     }
 
     // check if wire exists, wff
     if eboxf.wff.is_some() {
-        wff_cross_att_wiring(
-            eboxf,
-            &mut nodes.clone(),
-            edges,
-            idx as u32,
-            bf_counter as u8,
-        );
+        (edges, nodes) = wff_cross_att_wiring(eboxf, nodes.clone(), edges, idx, bf_counter);
     }
+    (edges, nodes)
 }
 // this will construct connections from the function opi's to the sub module opi's, tracing inputs through the function
 // opi(sub)->opi(fun)
 pub fn wfopi_cross_att_wiring(
     eboxf: FunctionNet, // This is the current attribute
-    nodes: &mut Vec<Node>,
-    edges: &mut Vec<Edge>,
+    nodes: Vec<Node>,
+    mut edges: Vec<Edge>,
     idx: u32,       // this is the current attribute index
     bf_counter: u8, // this is the current box
-) -> () {
+) -> Vec<Edge> {
     for wire in eboxf.wfopi.as_ref().unwrap().iter() {
         // collect info to identify the opi src node
         let src_idx = wire.src; // port index
@@ -4093,11 +4408,11 @@ pub fn wfopi_cross_att_wiring(
             // find the src node
             for node in nodes.iter() {
                 // make sure in correct box
-                if src_nbox == node.nbox as u8 {
+                if src_nbox == node.nbox {
                     // make sure only looking in current attribute nodes for srcs and tgts
-                    if src_att == node.contents as u32 {
+                    if src_att == node.contents {
                         // make sure box index also lines up
-                        if (src_box as u32) == node.att_bf_idx as u32 {
+                        if (src_box as u32) == node.att_bf_idx {
                             // only opo's
                             if node.n_type == "Opi" {
                                 // iterate through port to check for tgt
@@ -4114,9 +4429,9 @@ pub fn wfopi_cross_att_wiring(
             }
             for node in nodes.iter() {
                 // make sure in correct box
-                if tgt_nbox == node.nbox as u8 {
+                if tgt_nbox == node.nbox {
                     // make sure only looking in current attribute nodes for srcs and tgts
-                    if tgt_att == node.contents as u32 {
+                    if tgt_att == node.contents {
                         // only opo's
                         if node.n_type == "Opi" {
                             // iterate through port to check for tgt
@@ -4141,16 +4456,17 @@ pub fn wfopi_cross_att_wiring(
             }
         }
     }
+    edges
 }
 // this will construct connections from the function opo's to the sub module opo's, tracing outputs through the function
 // opo(fun)->opo(sub)
 pub fn wfopo_cross_att_wiring(
     eboxf: FunctionNet, // This is the current attribute
-    nodes: &mut Vec<Node>,
-    edges: &mut Vec<Edge>,
+    nodes: Vec<Node>,
+    mut edges: Vec<Edge>,
     idx: u32,       // this +1 is the current attribute index
     bf_counter: u8, // this is the current box
-) -> () {
+) -> Vec<Edge> {
     for wire in eboxf.wfopo.as_ref().unwrap().iter() {
         // collect info to identify the opo tgt node
         let tgt_idx = wire.tgt; // port index
@@ -4178,9 +4494,9 @@ pub fn wfopo_cross_att_wiring(
             // find the src node
             for node in nodes.iter() {
                 // make sure in correct box
-                if src_nbox == node.nbox as u8 {
+                if src_nbox == node.nbox {
                     // make sure only looking in current attribute nodes for srcs and tgts
-                    if src_att == node.contents as u32 {
+                    if src_att == node.contents {
                         // only opo's
                         if node.n_type == "Opo" {
                             // iterate through port to check for tgt
@@ -4196,10 +4512,9 @@ pub fn wfopo_cross_att_wiring(
             }
             for node in nodes.iter() {
                 // make sure in correct box
-                if tgt_nbox == node.nbox as u8 {
+                if tgt_nbox == node.nbox {
                     // make sure only looking in current attribute nodes for srcs and tgts
-                    if tgt_att == node.contents as u32 && (tgt_box as u32) == node.att_bf_idx as u32
-                    {
+                    if tgt_att == node.contents && (tgt_box as u32) == node.att_bf_idx {
                         // only opo's
                         if node.n_type == "Opo" {
                             // iterate through port to check for tgt
@@ -4224,17 +4539,18 @@ pub fn wfopo_cross_att_wiring(
             }
         }
     }
+    edges
 }
 // this will construct connections from the sub function modules opi's to another sub module opo's, tracing data inside the function
 // opi(sub)->opo(sub)
 #[allow(unused_assignments)]
 pub fn wff_cross_att_wiring(
     eboxf: FunctionNet, // This is the current attribute, should be the function if in a function
-    nodes: &mut Vec<Node>,
-    edges: &mut Vec<Edge>,
+    mut nodes: Vec<Node>,
+    mut edges: Vec<Edge>,
     idx: u32,       // this +1 is the current attribute index
     bf_counter: u8, // this is the current box
-) -> () {
+) -> (Vec<Edge>, Vec<Node>) {
     for wire in eboxf.wff.as_ref().unwrap().iter() {
         // collect info to identify the opi src node
         let src_idx = wire.src; // port index
@@ -4279,11 +4595,9 @@ pub fn wff_cross_att_wiring(
                 // internal wiring module as well?
                 for node in nodes.iter() {
                     // make sure in correct box
-                    if src_nbox == node.nbox as u8 {
+                    if src_nbox == node.nbox {
                         // make sure only looking in current attribute nodes for srcs and tgts
-                        if src_att == node.contents as u32
-                            && (src_box as u32) == node.att_bf_idx as u32
-                        {
+                        if src_att == node.contents && (src_box as u32) == node.att_bf_idx {
                             // only opo's
                             if node.n_type == "Opi" {
                                 // iterate through port to check for tgt
@@ -4299,11 +4613,9 @@ pub fn wff_cross_att_wiring(
                 }
                 for node in nodes.iter() {
                     // make sure in correct box
-                    if tgt_nbox == node.nbox as u8 {
+                    if tgt_nbox == node.nbox {
                         // make sure only looking in current attribute nodes for srcs and tgts
-                        if tgt_att == node.contents as u32
-                            && (tgt_box as u32) == node.att_bf_idx as u32
-                        {
+                        if tgt_att == node.contents && (tgt_box as u32) == node.att_bf_idx {
                             // only opo's
                             if node.n_type == "Opo" {
                                 // iterate through port to check for tgt
@@ -4333,11 +4645,9 @@ pub fn wff_cross_att_wiring(
                 // find the src node
                 for node in nodes.iter() {
                     // make sure in correct box
-                    if src_nbox == node.nbox as u8 {
+                    if src_nbox == node.nbox {
                         // make sure only looking in current attribute nodes for srcs and tgts
-                        if src_att == node.contents as u32
-                            && (src_box as u32) == node.att_bf_idx as u32
-                        {
+                        if src_att == node.contents && (src_box as u32) == node.att_bf_idx {
                             // only opo's
                             if node.n_type == "Opi" {
                                 // iterate through port to check for tgt
@@ -4353,11 +4663,9 @@ pub fn wff_cross_att_wiring(
                 }
                 for node in nodes.iter() {
                     // make sure in correct box
-                    if tgt_nbox == node.nbox as u8 {
+                    if tgt_nbox == node.nbox {
                         // make sure only looking in current attribute nodes for srcs and tgts
-                        if tgt_att == node.contents as u32
-                            && (tgt_box as u32) == node.att_bf_idx as u32
-                        {
+                        if tgt_att == node.contents && (tgt_box as u32) == node.att_bf_idx {
                             // only opo's
                             if node.n_type == "Primitive" || node.n_type == "Literal" {
                                 // iterate through port to check for tgt
@@ -4419,11 +4727,9 @@ pub fn wff_cross_att_wiring(
                 // find the src node
                 for node in nodes.iter() {
                     // make sure in correct box
-                    if src_nbox == node.nbox as u8 {
+                    if src_nbox == node.nbox {
                         // make sure only looking in current attribute nodes for srcs and tgts
-                        if src_att == node.contents as u32
-                            && (src_box as u32) == node.att_bf_idx as u32
-                        {
+                        if src_att == node.contents && (src_box as u32) == node.att_bf_idx {
                             // only opo's
                             if node.n_type == "Primitive" {
                                 // iterate through port to check for tgt
@@ -4439,11 +4745,9 @@ pub fn wff_cross_att_wiring(
                 }
                 for node in nodes.iter() {
                     // make sure in correct box
-                    if tgt_nbox == node.nbox as u8 {
+                    if tgt_nbox == node.nbox {
                         // make sure only looking in current attribute nodes for srcs and tgts
-                        if tgt_att == node.contents as u32
-                            && (tgt_box as u32) == node.att_bf_idx as u32
-                        {
+                        if tgt_att == node.contents && (tgt_box as u32) == node.att_bf_idx {
                             // only opo's
                             if node.n_type == "Opo" {
                                 // iterate through port to check for tgt
@@ -4469,13 +4773,14 @@ pub fn wff_cross_att_wiring(
             }
         }
     }
+    (edges, nodes)
 }
 // external wiring is the wiring between boxes at the module level
 pub fn external_wiring(
     gromet: &ModuleCollection,
-    nodes: &mut Vec<Node>,
-    edges: &mut Vec<Edge>,
-) -> () {
+    nodes: Vec<Node>,
+    mut edges: Vec<Edge>,
+) -> Vec<Edge> {
     if gromet.modules[0].r#fn.wff.as_ref().is_some() {
         for wire in gromet.modules[0].r#fn.wff.as_ref().unwrap().iter() {
             let src_idx = wire.src; // pif wire connects to
@@ -4522,8 +4827,8 @@ pub fn external_wiring(
             // This is double counting since only check is name and box, check on attributes?
             // find the src
             for node in nodes.iter() {
-                if node.nbox == src_box as usize
-                    && src_att == node.contents as u32
+                if node.nbox == src_box
+                    && src_att == node.contents
                     && (node.n_type == "Opi" || node.n_type == "Import")
                 {
                     for p in node.in_indx.as_ref().unwrap().iter() {
@@ -4539,8 +4844,8 @@ pub fn external_wiring(
             for node in nodes.iter() {
                 // check this field
                 if node.n_type == "Opo" {
-                    if node.nbox == tgt_box as usize {
-                        if tgt_att == node.contents as u32 {
+                    if node.nbox == tgt_box {
+                        if tgt_att == node.contents {
                             for p in node.out_idx.as_ref().unwrap().iter() {
                                 // push the tgt
                                 if (tgt_id as u32) == *p {
@@ -4550,7 +4855,7 @@ pub fn external_wiring(
                             }
                         } else {
                             // perform extra check to make sure not getting sub cross attributal nodes
-                            if tgt_att == node.contents as u32 {
+                            if tgt_att == node.contents {
                                 for p in node.out_idx.as_ref().unwrap().iter() {
                                     // push the tgt
                                     if (tgt_id as u32) == *p {
@@ -4563,7 +4868,7 @@ pub fn external_wiring(
                     }
                 } else if (node.n_type == "Literal" || node.n_type == "Import")
                     && !tgt_found
-                    && node.nbox == tgt_box as usize
+                    && node.nbox == tgt_box
                 {
                     for p in node.out_idx.as_ref().unwrap().iter() {
                         // push the tgt
@@ -4584,6 +4889,7 @@ pub fn external_wiring(
             }
         }
     }
+    edges
 }
 
 pub fn parse_gromet_queries(gromet: ModuleCollection) -> Vec<String> {
